### 2021-02-23 #3057 @whedon
[PRE REVIEW]: TUM Open Infra Platform: an open source package for simultaneous viewing and analysis of digital models in the civil engineering domain
**Submitting author:** @pjanck (<a href="http://orcid.org/0000-0003-0387-3440">Štefan Jaud</a>)
**Repository:** <a href="https://github.com/tumcms/Open-Infra-Platform" target ="_blank">https://github.com/tumcms/Open-Infra-Platform</a>
**Version:** [still a moving target]
