### 2021-02-23 #3057 @pjanck
Hi @kthyng: I believe that whedon needs to compile the `joss_submission` branch. The version of the paper in the current main branch is not according to the specifications.
