### 2021-02-24 36254eb
small tweaks, correcting references
