+++ Documentation/joss/paper.bib
@@ -77,7 +77,7 @@
     abstract = { This paper demonstrates how linked data can be used to provide a semantically rich connection between the domains of building information modeling (BIM) and geographic information systems (GIS). Expertise from both domains is necessary for the planning of infrastructure projects such as roads, tunnels, bridges, and railways, as these projects usually require the consideration of widely diverging scales. As BIM and GIS rely on different types of data modeling standards, a conversion between the data models will inevitably result in data loss. To overcome this problem, we propose to utilize the concept of Linked Data which allows the original data to coexist and provide coherence by establishing references between the corresponding entities of both standards. The approach is illustrated using exemplary models of shield tunnels in both, the IFC and the CityGML format. }
 }
 
-@techreport{Hecht:2018:PCdtoAlignment,
+@techreport{Hecht:2018:PCDtoAlignment,
   author    = {Helge Hecht},
   type = {Report},
   title     = {From Point Cloud Data to IfcAlignment},
@@ -101,7 +101,7 @@
 @phdthesis {Amann:2018:Diss,
 	author = {Amann, J.},
 	title = {An object-oriented language for embedding interpretation semantics in digital building models},
-	type = {Dissertation},
+	type = {PhD Dissertation},
 	school = {Technische Universit\"{a}t M\"{u}nchen},
 	address = {M\"{u}nchen},
 	Year = {2018},
@@ -141,6 +141,7 @@
     title = {{Industry Foundation Classes (IFC) for data sharing in the construction and facility management industries -- Part 1: Data schema}},
     volume = {2018},
     address = {Geneva, CH},
+    author = {{ISO}},
     institution = {International Organization for Standardization},
     url = {https://www.iso.org/standard/70303.html}
 }
