+++ Documentation/joss/paper.md
@@ -7,7 +7,7 @@ tags:
   - EXPRESS parser
   - code generation
 authors:
-  - name: Stefan Jaud^[corresponding author]
+  - name: \v{S}tefan Jaud^[corresponding author]
     orcid: 0000-0003-0387-3440
     affiliation: 1
   - name: Helge Hecht
@@ -31,18 +31,18 @@ Building Information Modeling (BIM) is steadily gaining importance replacing the
  and getting implemented in every aspect of the very complex software and stakeholder landscape [@Jaud:2019:iccce].
 As one of the main principles, BIM describes the idea of integrating all information
  relevant to the life cycle of a structure, such as a tunnel, bridge, building or road, in a digital (BIM) model.
-The digital model is to ensure, among other things:
+The digital model is to ensure, among other things [@Amann:2018:Diss]:
 
 - that all relevant data is available to all project participants;
 - that all data is in a consistent state (data integrity should be guaranteed);
-- that the data can be used efficiently [@Amann:2018:Diss].
+- that the data can be used efficiently.
 
 *TUM Open Infra Platform* (OIP) is an open source application for
  viewing and analysis of different BIM models used in the civil engineering field.
 OIP supports reading, visualization, navigating, and handling of:
 
 - Industry Foundation Classes (IFC) models as specified in ISO 16739 [@iso16739];
-- Point Cloud Data (PCD) models as supported by the Cloud Compare library [CloudCompare].
+- Point Cloud Data (PCD) models as supported by the Cloud Compare library [@CloudCompare].
 
 Multiple models can be loaded at once and compared between each other (see Figure \ref{fig:ifcpcd}).
 Their absolute position is accounted for, so the models can be checked against one another for internal differences based on location of elements.
@@ -53,15 +53,15 @@ As such, complete contents of IFC files can be interpreted and analysed.
 Moreover, type safety is guaranteed at compile time thus reducing the risk for bugs and errors.
 
 Let it be stated here that OIP serves as a prototypical playground for developments.
-The software architecture, features, and functionalities have been changed, added or removed as need be along the way.
-Nowadays, OIP uses the IFC schema and Cloud Compare's model as independent internal data models [@Hecht:2019:FBI].
+The software architecture, features, and functionalities have been changed, added or removed as required along the way.
+Nowadays, OIP uses the IFC schema and CloudCompare's model as independent internal data models [@Hecht:2019:FBI].
 Major elements from previous developments were already realigned, while some (previously available) functionality is still considered as work-in-progress.
 
 ![A PCD model together with an IFC model loaded in OIP.\label{fig:ifcpcd}](../images/ifc_and_pcd.png)
 
 # Statement of Need
 
-During the development of IFC standard, a clear need for an independent software
+During the development of the IFC standard, a clear need for an independent software
  capable of consuming and producing IFC files according to the newly developed schemas was determined [@Amann:2015:OIP; @Amann:2015:ICCBEI].
 OIP fulfils this role by being open source, using CMake as a build system generator to support various platforms,
  and being based on well known libraries (like Qt, Eigen, and Boost).
@@ -77,17 +77,18 @@ Recently, OIP was selected among many other software solutions as best suitable 
 
 The origins of the software date back to the first projects expanding IFC for infrastructure
  at the Technical University of Munich [@Amann:2014:ECPPM; @Singer:2014:IfcAlignment; @Amann:2015:OIP; @Amann:2015:OKSTRA; @Amann:2015:ICCBEI].
[…]
