### 2021-02-24 #3061 @whedon
[REVIEW]: TUM Open Infra Platform: an open source package for simultaneous viewing and analysis of digital models in the civil engineering domain
**Submitting author:** @pjanck (<a href="http://orcid.org/0000-0003-0387-3440">Štefan Jaud</a>)
**Repository:** https://github.com/tumcms/Open-Infra-Platform
**Branch with paper.md** (empty if default branch): development
**Version:** release-v4.0.0

@aothms & @CBenghi, @Abdoulayediak please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
