### 2021-02-25 baa046f
small corrections to bib
