+++ Documentation/joss/paper.bib
@@ -2,7 +2,7 @@
 @inproceedings{Amann:2014:ECPPM,
   author = {Amann, J. and Flurl, M. and Jubierre, J. R. and Borrmann, A.},
   title = {An alignment meta-model for the comparison of alignment product models},
-  booktitle = {Proc. of the 10th European Conference on Product & Process Modelling},
+  booktitle = {Proceedings of the 10th European Conference on Product & Process Modelling},
   year = {2014},
   month = {Sep}
 }
@@ -10,9 +10,9 @@
 @inproceedings{Singer:2014:IfcAlignment,
     Author       = {Singer, D. and Amann, J.},
     Title        = {Erweiterung von IFC Alignment um Stra{\ss}enquerschnitte},
-    Booktitle    = {Proc. of the 26th Forum Bauinformatik},
+    Booktitle    = {Proceedings of the 26th Forum Bauinformatik},
     Year         = {2014},
-    Address      = {Darmstadt, Deutschland}
+    Address      = {Darmstadt, Germany}
 }
 
     
@@ -22,7 +22,7 @@
     Title        = {Open BIM for Infrastructure {--} mit OKSTRA und IFC Alignment zur internationalen Standardisierung des Datenaustauschs},
     Booktitle    = {Tagungsband zum 6. OKSTRA-Symposium},
     Year         = {2015},
-    Address      = {K{\"o}ln, Deutschland}
+    Address      = {K{\"o}ln, Germany}
 }
 
     
@@ -31,6 +31,7 @@
     Author       = {Amann, J. and Borrmann, A.},
     Title        = {Creating a 3D-BIM-compliant road design based on IFC alignment originating from an OKSTRA-accordant 2D road design using the TUM Open Infra Platform and the OKSTRA class library},
     Institution  = {Technische Universit{\"a}t M{\"u}nchen},
+    type = {Technical report},
     Year         = {2015}
 }
 
@@ -39,7 +40,7 @@
 @inproceedings{Amann:2015:ICCBEI,
     Author       = {Amann, J. and Singer, D. and Borrmann, A.},
     Title        = {Extension of the upcoming IFC alignment standard with cross sections for road design},
-    Booktitle    = {Proc. of the ICCBEI 2015},
+    Booktitle    = {Proceedings of the ICCBEI 2015},
     Year         = {2015},
     Address      = {Tokyo, Japan}
 }
@@ -79,7 +80,7 @@
 
 @techreport{Hecht:2018:PCDtoAlignment,
   author    = {Helge Hecht},
-  type = {Report},
+  type = {Technical report},
   title     = {From Point Cloud Data to IfcAlignment},
   pages     = {26 pp.},
   year      = {2018},
@@ -110,9 +111,8 @@
 
 @techreport{Beetz:2019:linkeddata,
   author    = {Jakob Beetz and Julian Amann and Andr{\'{e}} Borrmann},
-  type = {Report},
-  title     = {Linked Data - Analyse von Einsatzm{\"o}glichkeiten von verbundenen Informationen (Linked Data) und Ontologien und damit befassten Technologien (Semantic Web) im Bereich des Stra{\"s}enwesens},
-  institution = {Sonstige},
+  type = {Technical report},
+  title     = {Linked Data - Analyse von Einsatzm{\"o}glichkeiten von verbundenen Informationen (Linked Data) und Ontologien und damit befassten Technologien (Semantic Web) im Bereich des Stra{\ss}enwesens},
   pages     = {82 pp.},
   year      = {2019},
[…]
