### 2021-02-25 51f5899
Boost in CurveConverter.h
