### 2021-02-25 615d2b1
Reworked throw in SplineInterpretation
