+++ Core/src/SplineInterpretation/SplineInterpretation.cpp
@@ -36,7 +36,7 @@ OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::SplineInter
 {}
 
 // PUBLIC FUNCTIONS
-void OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::convertSketchToAlignment() const throw(...)
+void OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::convertSketchToAlignment() const noexcept(false)
 {
 	// obtain control points from sketch or from file
 	// (from file is from past sketch or from [external] test data)
@@ -112,7 +112,7 @@ void OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::conver
 }
 
 // PRIVATE FUNCTIONS
-std::vector<carve::geom::vector<3>> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::obtainControlPoints() const throw(...)
+std::vector<carve::geom::vector<3>> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::obtainControlPoints() const noexcept(false)
 {
 	// initialice MessageBox
 	QMessageBox msgBox;
@@ -147,7 +147,7 @@ std::vector<carve::geom::vector<3>> OpenInfraPlatform::Core::SplineInterpretatio
 	return controlPoints;
 }
 
-void OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::saveControlPointsIntoFile(const std::vector<carve::geom::vector<3>>& points) const throw(...)
+void OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::saveControlPointsIntoFile(const std::vector<carve::geom::vector<3>>& points) const noexcept(true)
 {
 	// get number of files, which are in the default folder of sketches
 	const QString path = "testdata/StrokeToAlignment/";
@@ -196,7 +196,7 @@ void OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::saveCo
 	}
 }
 
-std::vector<carve::geom::vector<3>> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::loadControlPointsFromFile() const throw(...)
+std::vector<carve::geom::vector<3>> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::loadControlPointsFromFile() const noexcept(false)
 {
 	// from a file open dialog, get file name (including path) which should be opened
 	const std::string fileName = QFileDialog::getOpenFileName(nullptr, // parent Widget
@@ -271,7 +271,7 @@ std::vector<carve::geom::vector<3>> OpenInfraPlatform::Core::SplineInterpretatio
 	}
 }
 
-std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::obtainKnotArrayOpenUniform(const size_t nPoints, const int order) const throw(...)
+std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::obtainKnotArrayOpenUniform(const size_t nPoints, const int order) const noexcept(true)
 {
 	std::vector<double> knotArray;
 	// .resize sets all vector values to 0
@@ -288,7 +288,7 @@ std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpr
 	return knotArray;
 }
 
-std::vector<std::pair<double, double>> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::movingAverageVariableWindow(std::vector<std::pair<double, double>>& xy) const throw(...)
[…]
