+++ Core/src/SplineInterpretation/SplineInterpretation.h
@@ -45,7 +45,7 @@ namespace OpenInfraPlatform
 				 * During execution, the user decides whether the input data will be sketched on screen or
 				 * will be loaded from a file. The function interprets the data and convert it to an alignment.
 				 */
-				void convertSketchToAlignment() const throw(...);
+				void convertSketchToAlignment() const noexcept(false);
 
 			private:
 				// VARIABLES
@@ -63,7 +63,7 @@ namespace OpenInfraPlatform
 				 *
 				 * \return A vector of the control points.
 				 */
-				std::vector<carve::geom::vector<3>> obtainControlPoints() const throw(...);
+				std::vector<carve::geom::vector<3>> obtainControlPoints() const noexcept(false);
 
 				/*! \brief Saves the control points or points from sketch input into a file.
 				 *
@@ -75,7 +75,7 @@ namespace OpenInfraPlatform
 				 *
 				 * \param [in]	controlPoints	The control points or points from sketch input, which are in a \c std::vector<carve::geom::vector<3>>.
 				 */
-				void saveControlPointsIntoFile(const std::vector<carve::geom::vector<3>>& points) const throw(...);
+				void saveControlPointsIntoFile(const std::vector<carve::geom::vector<3>>& points) const noexcept(true);
 
 				/*! \brief Loads the control points or points from a previous sketch input from a file.
 				 *
@@ -86,7 +86,7 @@ namespace OpenInfraPlatform
 				 *
 				 * \return	A vector of the control points.
 				 */
-				std::vector<carve::geom::vector<3>> loadControlPointsFromFile() const throw(...);
+				std::vector<carve::geom::vector<3>> loadControlPointsFromFile() const noexcept(false);
 
 				/*! \brief Generates the open uniform knot array of a B-Spline.
 				 *
@@ -101,7 +101,7 @@ namespace OpenInfraPlatform
 				 *
 				 * \return		The open uniform knots array.
 				 */
-				std::vector<double> obtainKnotArrayOpenUniform(const size_t nPoints, const int order) const throw(...);
+				std::vector<double> obtainKnotArrayOpenUniform(const size_t nPoints, const int order) const noexcept(true);
 
 				/*! \brief Applies the moving average with variable window on the y-data of \c xy.
 				 *
@@ -113,7 +113,7 @@ namespace OpenInfraPlatform
 				 *
 				 * \return			The pair from the input, but the y-values are smoothed by the moving average.
 				 */
-				std::vector<std::pair<double, double>> movingAverageVariableWindow(std::vector<std::pair<double, double>>& xy) const throw(...);
+				std::vector<std::pair<double, double>> movingAverageVariableWindow(std::vector<std::pair<double, double>>& xy) const noexcept(true);
 
 				/*! \brief Applies the moving average with variable window on the vector \c data.
 				 *
@@ -124,7 +124,7 @@ namespace OpenInfraPlatform
 				 *
 				 * \return				A vector of \c double, but the values are smoothed by the moving average.
 				 */
-				std::vector<double> movingAverageVariableWindow(std::vector<double>& data) const throw(...);
+				std::vector<double> movingAverageVariableWindow(std::vector<double>& data) const noexcept(true);
 
 				/* \brief Calculates the range of a variogramm.
 				 *
@@ -136,7 +136,7 @@ namespace OpenInfraPlatform
 				 *
 				 * return	Range of a semivariogramm regarding the number of values.
 				 */
-				size_t variogrammGetRange(std::vector<double>& data) const throw(...);
+				size_t variogrammGetRange(std::vector<double>& data) const noexcept(true);
 
 				/* \brief Calculates the numeric derivative of a vector of x-y-pairs.
 				 *
@@ -144,7 +144,7 @@ namespace OpenInfraPlatform
[…]
