### 2021-02-25 154ca21
Reworked const and & of function parameters
