+++ Core/src/SplineInterpretation/SplineInterpretation.cpp
@@ -308,7 +308,7 @@ std::vector<std::pair<double, double>> OpenInfraPlatform::Core::SplineInterpreta
 	return xy;
 }
 
-std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::movingAverageVariableWindow(std::vector<double>& data) const noexcept(true)
+std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::movingAverageVariableWindow(const std::vector<double>& data) const noexcept(true)
 {
 	// window size
 	const size_t range = variogrammGetRange(data);
@@ -351,7 +351,7 @@ std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpr
 	return value;
 }
 
-size_t OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::variogrammGetRange(std::vector<double>& data) const noexcept(true)
+size_t OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::variogrammGetRange(const std::vector<double>& data) const noexcept(true)
 {
 	// initialice vector of semivariogramm variable gamma(h)
 	std::vector<double> gamma(data.size() - 1, 0.0);
@@ -389,7 +389,7 @@ size_t OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::vari
 	return range;
 }
 
-std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::numericDerivative(std::vector<std::pair<double, double>>& xy) const noexcept(true)
+std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::numericDerivative(const std::vector<std::pair<double, double>>& xy) const noexcept(true)
 {
 	// initialice target vector
 	std::vector<double> dy;
@@ -403,8 +403,8 @@ std::vector<double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpr
 }
 
 std::vector<SplineInterpretationElement> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::identifyElementEndpoints(
-	std::vector<std::pair<double, double>> lengthsWithCurvatures, 
-	std::vector<double> curvatureChange) const noexcept(true)
+	const std::vector<std::pair<double, double>>& lengthsWithCurvatures, 
+	const std::vector<double>& curvatureChange) const noexcept(true)
 {
 	// indicate the change of curvature
 	//   curvatureZero: threshold
@@ -428,7 +428,7 @@ std::vector<SplineInterpretationElement> OpenInfraPlatform::Core::SplineInterpre
 }
 
 std::tuple<std::vector<int>, double> OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::indicateCurvatureChange(
-	std::vector<double> curvatureChange) const noexcept(true)
+	const std::vector<double>& curvatureChange) const noexcept(true)
 {
 	// threshold of curvatureChange
 	const double curvatureZero = obtainThreshold(curvatureChange);
@@ -444,7 +444,7 @@ std::tuple<std::vector<int>, double> OpenInfraPlatform::Core::SplineInterpretati
 	return { indicator, curvatureZero };
 }
 
-double OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::obtainThreshold(std::vector<double> data) const noexcept(true) {
+double OpenInfraPlatform::Core::SplineInterpretation::SplineInterpretation::obtainThreshold(const std::vector<double>& data) const noexcept(true) {
 	// allows +- 25 % around 0
 	return std::max(
 		0.25 * std::abs(*std::min_element(data.begin(), data.end())),
[…]
