### 2021-02-25 9f6a3fa
Fixed 5 warnings in the VS error-list, ... which showed up after the project was built.
