### 2021-02-26 0f25f00
Calculated Angles with point in 2D
