+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -958,21 +958,20 @@ namespace OpenInfraPlatform {
 					//	END_ENTITY;
 					// **************************************************************************************************************************
 					// Determine position
-					//carve::math::Matrix conicPositionMatrix = placementConverter->convertIfcAxis2Placement(circle->Position);
-
-					boost::numeric::ublas::matrix<double> conicPositionMatrix = placementConverter->convertIfcAxis2Placement(circle->Position);
-					boost::numeric::ublas::matrix<double> inverseConicPositionMatrix = -conicPositionMatrix;
+					carve::math::Matrix conicPositionMatrix = placementConverter->convertIfcAxis2Placement(circle->Position);
+					carve::math::Matrix inverseConicPositionMatrix;
+					GeomUtils::computeInverse(conicPositionMatrix, inverseConicPositionMatrix);
 
 					// Get radius
 					double circleRadius = circle->Radius * UnitConvert()->getLengthInMeterFactor();
 					
 					// Calculate an angle on the circle (with circle center in (0., 0., 0.)) for trimming begin.
 					carve::geom::vector<3> point = getPointOnCurve<typename IfcEntityTypesT::IfcCircle>(circle, trim1Vec, trimmingPreference);
-					double startAngle = getAngleOnCircle(conicPositionMatrix * carve::geom::VECTOR(0., 0., 0.), circleRadius, point);
+					double startAngle = getAngleOnCircle(carve::geom::VECTOR(0., 0., 0.), circleRadius, inverseConicPositionMatrix * point);
 					
 					// Calculate an angle on the circle (with circle center in (0., 0., 0.)) for trimming end.
 					point = getPointOnCurve<typename IfcEntityTypesT::IfcCircle>(circle, trim2Vec, trimmingPreference);
-					double endAngle = getAngleOnCircle(conicPositionMatrix * carve::geom::VECTOR(0., 0., 0.), circleRadius, point);
+					double endAngle = getAngleOnCircle(carve::geom::VECTOR(0., 0., 0.), circleRadius, inverseConicPositionMatrix * point);
 					
 					// Calculate an opening angle.
 					double openingAngle = calculateOpeningAngle(senseAgreement, startAngle, endAngle);
@@ -1880,73 +1879,38 @@ namespace OpenInfraPlatform {
 					if (this->GeomSettings()->areEqual(centerToTrimPoint.length(), circleRadius)) {
 						centerToTrimPoint.normalize();
 
-						if (!this->GeomSettings()->areEqual(centerToTrimPoint.x, 0.) &&
-							!this->GeomSettings()->areEqual(centerToTrimPoint.y, 0.) &&
-							!this->GeomSettings()->areEqual(centerToTrimPoint.z, 0.))
-						{
-							throw oip::InconsistentGeometryException("It is not possible to get only one Angle in Three-dimensional space");
-						}
-						else if (!this->GeomSettings()->areEqual(centerToTrimPoint.x, 0.))
-						{
-							double cosAngle = carve::geom::dot(centerToTrimPoint, carve::geom::vector<3>(carve::geom::VECTOR(1., 0., 0.)));
+						double cosAngle = carve::geom::dot(centerToTrimPoint, carve::geom::vector<3>(carve::geom::VECTOR(1., 0., 0.)));
 
-							if (!this->GeomSettings()->areEqual(centerToTrimPoint.y, 0.))
-							{
-								// Calculate angle in XY plane
-								return getAngleOnCircle2DPlane(carve::geom::VECTOR(centerToTrimPoint.x, centerToTrimPoint.y), cosAngle);
[…]
