+++ UnitTests/Schemas/IFC4X1/beam-varying-cardinal-points/src/beam-varying-cardinal-points.cpp
@@ -1,5 +1,5 @@
 /*
-	Copyright (c) 2020 Technical University of Munich
+	Copyright (c) 2021 Technical University of Munich
 	Chair of Computational Modeling and Simulation.
 
 	TUM Open Infra Platform is free software; you can redistribute it and/or modify
