+++ Tools/fileHeaderGenerator.py
@@ -27,7 +27,7 @@ from optparse import OptionParser
 #print 'Argument List:', str(sys.argv)
 
 textHeader = """/*
-    Copyright (c) 2018 Technical University of Munich
+    Copyright (c) 2021 Technical University of Munich
     Chair of Computational Modeling and Simulation.
 
     TUM Open Infra Platform is free software; you can redistribute it and/or modify
