+++ UserInterface/ViewPanel/View.cpp
@@ -24,7 +24,10 @@
 #include "BlueFramework/Rasterizer/util.h"
 
 
-QIcon homeIcon, ghostSelectedIcon, ghostUnselectedIcon, snowSelectedIcon, snowUnselectedIcon;
+QIcon homeIcon, 
+	ghostSelectedIcon, ghostUnselectedIcon, 
+	snowSelectedIcon, snowUnselectedIcon, 
+	playIcon, pauseIcon;
 
 OpenInfraPlatform::UserInterface::View::View() : QDockWidget(), // no parent
 	viewport_(nullptr)
@@ -76,6 +79,9 @@ OpenInfraPlatform::UserInterface::View::View() : QDockWidget(), // no parent
     snowSelectedIcon = QIcon("UserInterface/Data/snow_selected.ico");
     snowUnselectedIcon = QIcon("UserInterface/Data/snow_unselected.ico");
 
+	playIcon = QIcon("UserInterface/Data/play.ico");
+	pauseIcon = QIcon("UserInterface/Data/pause.ico");
+
 	if (homeIcon.isNull() || ghostSelectedIcon.isNull() || ghostUnselectedIcon.isNull())
 	{
 		BLUE_LOG(warning) << "Could not load icon from.";
@@ -88,6 +94,10 @@ OpenInfraPlatform::UserInterface::View::View() : QDockWidget(), // no parent
 	cameraGhostModeAction_ = new QAction("", nullptr);
 	cameraGhostModeAction_->setShortcut(QKeySequence(Qt::Key_T));
 
+	playAction_ = new QAction("", nullptr);
+	playAction_->setShortcut(QKeySequence(Qt::Key_P));
+	playAction_->setIcon(playIcon);
+
     snowAction_ = new QAction("", nullptr);
     snowAction_->setIcon(snowUnselectedIcon);
 
@@ -95,6 +105,7 @@ OpenInfraPlatform::UserInterface::View::View() : QDockWidget(), // no parent
 	rightBar->setObjectName("transparent");
 
 	rightBar->addAction(homeAction_);
+	rightBar->addAction(playAction_);
 	rightBar->addAction(cameraGhostModeAction_);
 
 #if _DEBUG
