+++ CMakeLists.txt
@@ -234,7 +234,7 @@ endif()
 FetchContent_Declare(
 	blue_framework
 	GIT_REPOSITORY https://github.com/tumcms/Blue-Framework.git
-	GIT_TAG a0.0.9
+	GIT_TAG a0.0.11
 )
   
 FetchContent_GetProperties(blue_framework)
