### 2021-02-26 119b902
initial commit documentation of supporting ifc geometric entities
