### 2021-02-26 db961f0
Updating function computeInverse
