### 2021-02-28 82daa69
Change to "areEqual"
