+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -1528,7 +1528,15 @@ namespace OpenInfraPlatform {
 
 						// skip duplicate vertices
 						if (!first) {
-							if (abs(vertex.x - vertex_previous.x) < 0.00000001) {
+
+							if (this->GeomSettings()->areEqual(vertex, vertex_previous))
+							{
+								// TODO: is it better to report degenerated loops, or to just omit them?
+								BLUE_LOG(warning) << "Duplicate point in polyloop. IfcPoint #" << cp->getId();
+								//continue;
+							}
+
+							/*if (abs(vertex.x - vertex_previous.x) < 0.00000001) {
 								if (abs(vertex.y - vertex_previous.y) < 0.00000001) {
 									if (abs(vertex.z - vertex_previous.z) < 0.00000001) {
 										// TODO: is it better to report degenerated loops, or to just omit them?
@@ -1536,7 +1544,7 @@ namespace OpenInfraPlatform {
 										//continue;
 									}
 								}
-							}
+							}*/
 						}
 						else
 							first = false;
