+++ UnitTests/Schemas/IFC4X1/beam-straight-i-shape-tessellated/src/beam-straight-i-shape-tessellated.cpp
@@ -40,8 +40,8 @@ protected:
 		importer = buw::makeReferenceCounted<oip::IfcImporterT<emt::IFC4X1EntityTypes>>();
 		model = importer->collectData(express_model);
 
-		_background = renderer->captureImage();
-		renderer->setModel(model);
+		_background = rendererIfc->captureImage();
+		rendererIfc->setModel(model);
 	}
 
 	virtual void TearDown() override {
@@ -74,7 +74,7 @@ TEST_F(BeamStraightIShapeTessellated, IFCHasEssentialEntities) {
 TEST_F(BeamStraightIShapeTessellated, ImageIsSaved)
 {
 	// Arrange
-	buw::Image4b image = renderer->captureImage();
+	buw::Image4b image = rendererIfc->captureImage();
 
 	// Act
 	buw::storeImage(testPath("beam-straight-i-shape-tessellated.png").string(), image);
@@ -94,22 +94,22 @@ TEST_F(BeamStraightIShapeTessellated, PlaneSurfaceViews)
 	const auto expected_back = buw::loadImage4b(dataPath("beam-straight-i-shape-tessellated_back.png").string());
 
 	// Act (Front)
-	renderer->setViewDirection(buw::eViewDirection::Front);
+	rendererIfc->setViewDirection(buw::eViewDirection::Front);
 	buw::Image4b image_front = CaptureImage();
 	// Act (Top)
-	renderer->setViewDirection(buw::eViewDirection::Top);
+	rendererIfc->setViewDirection(buw::eViewDirection::Top);
 	buw::Image4b image_top = CaptureImage();
 	// Act (Bottom)
-	renderer->setViewDirection(buw::eViewDirection::Bottom);
+	rendererIfc->setViewDirection(buw::eViewDirection::Bottom);
 	buw::Image4b image_bottom = CaptureImage();
 	// Act (Left)
-	renderer->setViewDirection(buw::eViewDirection::Left);
+	rendererIfc->setViewDirection(buw::eViewDirection::Left);
 	buw::Image4b image_left = CaptureImage();
 	// Act (Right)
-	renderer->setViewDirection(buw::eViewDirection::Right);
+	rendererIfc->setViewDirection(buw::eViewDirection::Right);
 	buw::Image4b image_right = CaptureImage();
 	// Act (Back)
-	renderer->setViewDirection(buw::eViewDirection::Back);
+	rendererIfc->setViewDirection(buw::eViewDirection::Back);
 	buw::Image4b image_back = CaptureImage();
 
 	// uncomment following lines to also save the screen shot
@@ -144,28 +144,28 @@ TEST_F(BeamStraightIShapeTessellated, VertexViews)
 	const auto expected_right_bottom_back = buw::loadImage4b(dataPath("beam-straight-i-shape-tessellated_right_bottom_back.png").string());
 
 	// Act (FrontLeftBottom)
-	renderer->setViewDirection(buw::eViewDirection::FrontLeftBottom);
+	rendererIfc->setViewDirection(buw::eViewDirection::FrontLeftBottom);
 	buw::Image4b image_front_left_bottom = CaptureImage();
 	// Act (FrontRightBottom)
-	renderer->setViewDirection(buw::eViewDirection::FrontRightBottom);
+	rendererIfc->setViewDirection(buw::eViewDirection::FrontRightBottom);
 	buw::Image4b image_front_right_bottom = CaptureImage();
 	// Act (TopLeftFront)
-	renderer->setViewDirection(buw::eViewDirection::TopLeftFront);
+	rendererIfc->setViewDirection(buw::eViewDirection::TopLeftFront);
 	buw::Image4b image_top_left_front = CaptureImage();
[…]
