### 2021-03-07 1025721
Creating screenshots for beam-varying-extrusion-paths.ifc
