+++ UnitTests/Schemas/IFC4X1/beam-varying-extrusion-paths/src/beam-varying-extrusion-paths.cpp
@@ -40,8 +40,8 @@ protected:
 		importer = buw::makeReferenceCounted<oip::IfcImporterT<emt::IFC4X1EntityTypes>>();
 		model = importer->collectData(express_model);
 
-		_background = renderer->captureImage();
-		renderer->setModel(model);
+		_background = rendererIfc->captureImage();
+		rendererIfc->setModel(model);
 	}
 
 	virtual void TearDown() override {
@@ -79,11 +79,11 @@ TEST_F(BeamVaryingExtrusionPaths, CountEssentialEntities) {
 	auto result = std::count_if(express_model->entities.begin(), express_model->entities.end(), [](auto &pair) -> bool { return pair.second->classname() == "IFCBEAMSTANDARDCASE"; });
 	EXPECT_EQ(result, 2);
 }
-/*
+
 TEST_F(BeamVaryingExtrusionPaths, ImageIsSaved)
 {
 	// Arrange
-	buw::Image4b image = renderer->captureImage();
+	buw::Image4b image = rendererIfc->captureImage();
 
 	// Act
 	buw::storeImage(testPath("beam-varying-extrusion-paths.png").string(), image);
@@ -103,33 +103,33 @@ TEST_F(BeamVaryingExtrusionPaths, PlaneSurfaceViews)
 	const auto expected_back = buw::loadImage4b(dataPath("beam-varying-extrusion-paths_back.png").string());
 
 	// Act (Front)
-	renderer->setViewDirection(buw::eViewDirection::Front);
+	rendererIfc->setViewDirection(buw::eViewDirection::Front);
 	buw::Image4b image_front = CaptureImage();
 	// Act (Top)
-	renderer->setViewDirection(buw::eViewDirection::Top);
+	rendererIfc->setViewDirection(buw::eViewDirection::Top);
 	buw::Image4b image_top = CaptureImage();
 	// Act (Bottom)
-	renderer->setViewDirection(buw::eViewDirection::Bottom);
+	rendererIfc->setViewDirection(buw::eViewDirection::Bottom);
 	buw::Image4b image_bottom = CaptureImage();
 	// Act (Left)
-	renderer->setViewDirection(buw::eViewDirection::Left);
+	rendererIfc->setViewDirection(buw::eViewDirection::Left);
 	buw::Image4b image_left = CaptureImage();
 	// Act (Right)
-	renderer->setViewDirection(buw::eViewDirection::Right);
+	rendererIfc->setViewDirection(buw::eViewDirection::Right);
 	buw::Image4b image_right = CaptureImage();
 	// Act (Back)
-	renderer->setViewDirection(buw::eViewDirection::Back);
+	rendererIfc->setViewDirection(buw::eViewDirection::Back);
 	buw::Image4b image_back = CaptureImage();
 
 	// uncomment following lines to also save the screen shot
-	
+	/*
 	buw::storeImage(testPath("beam-varying-extrusion-paths_front.png").string(), image_front);
 	buw::storeImage(testPath("beam-varying-extrusion-paths_top.png").string(), image_top);
 	buw::storeImage(testPath("beam-varying-extrusion-paths_bottom.png").string(), image_bottom);
 	buw::storeImage(testPath("beam-varying-extrusion-paths_left.png").string(), image_left);
 	buw::storeImage(testPath("beam-varying-extrusion-paths_right.png").string(), image_right);
 	buw::storeImage(testPath("beam-varying-extrusion-paths_back.png").string(), image_back);
-	
+	*/
 
 	// Assert
 	EXPECT_EQ(image_front, expected_front);
@@ -153,32 +153,32 @@ TEST_F(BeamVaryingExtrusionPaths, VertexViews)
[…]
