### 2021-03-07 09db918
Corrected a call of the templated functions
