### 2021-03-07 e94c515
adding comparison operator to EXPRESSReference
