### 2021-03-07 fb25d56
splitting transformation matrix of GeorefMetadata into pieces
