### 2021-03-08 643376a
Added a log message in case of loadFromFile fails
