### 2021-03-08 e4fc298
Adjusted numbers to floating numbers in case of double calculation.
