### 2021-03-09 3ca85d7
Adjusted unit of directional angle to radian. Now, all calculation and information storage works in radian; just the print output to the console window has radian and degree. To get the angle in degree two functions were added to SplineInterpretationElement.
