### 2021-03-09 0cca62a
Created calculateIfcPointOnCurve
