+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -2115,12 +2115,28 @@ namespace OpenInfraPlatform {
 				}
 
 				/**********************************************************************************************/
+				/*! \brief Calculates a trimming point on the curve using \c IfcCartesianPoint.
+				* \tparam TCurve					A type of the curve.
+				* \param[in] curve					A pointer to data from the curve.
+				* \param[in] cartesianPoint			A pointer to data from \c IfcCartesianPoint.
+				* \return							The location of the trimming point.
+				* \note								The position is not applied.All calculations are made based on center in(0., 0., 0.).
+				*/
+				template <typename TCurve>
+				carve::geom::vector<3> getPointOnCurve(const EXPRESSReference<TCurve> & curve,
+					const typename IfcEntityTypesT::IfcParameterValue & parameter) const throw(...)
+				{
+					throw oip::UnhandledException(curve);
+				}
+
+				/**********************************************************************************************/
 				/*! \brief Calculates a trimming point on the circle using \c IfcParameterValue.
 				* \param[in] circle					A pointer to data from \c IfcCircle.
 				* \param[in] parameter				A pointer to data from \c IfcParameterValue.
 				* \return							The location of the trimming point.
 				* \note								The position is not applied. All calculations are made based on center in ( 0., 0., 0.).
 				*/
+				template <>
 				carve::geom::vector<3> getPointOnCurve(const EXPRESSReference<typename IfcEntityTypesT::IfcCircle>& circle,
 					const typename IfcEntityTypesT::IfcParameterValue & parameter) const throw(...)
 				{
@@ -2138,6 +2154,7 @@ namespace OpenInfraPlatform {
 				* \return							The location of the trimming point.
 				* \note								The position is not applied. All calculations are made based on center in ( 0., 0., 0.).
 				*/
+				template <>
 				carve::geom::vector<3> getPointOnCurve(const EXPRESSReference<typename IfcEntityTypesT::IfcEllipse>& ellipse,
 					const typename IfcEntityTypesT::IfcParameterValue & parameter) const throw(...)
 				{
@@ -2158,6 +2175,7 @@ namespace OpenInfraPlatform {
 				* \param[in] parameter				A pointer to data from \c IfcParameterValue.
 				* \return							The location of the trimming point.
 				*/
+				template <>
 				carve::geom::vector<3> getPointOnCurve(const EXPRESSReference<typename IfcEntityTypesT::IfcLine>& line,
 					const typename IfcEntityTypesT::IfcParameterValue & parameter)const throw (...)
 				{
