+++ Core/src/IfcGeometryConverter/PlacementConverter.h
@@ -28,6 +28,7 @@
 
 #include "GeomUtils.h"
 #include "ConverterBase.h"
+#include "CurveConverter.h"
 
 #include <BlueFramework/Core/Diagnostics/log.h>
 
@@ -100,7 +101,10 @@ namespace OpenInfraPlatform {
 					//TODO this is really, really bad programming - this needs to be done more nicely!!
 #endif
 
-					// IfcPointOnCurve & IfcPointOnSurface are not supported
+					// IfcPointOnCurve 
+					if (point.isOfType<typename IfcEntityTypesT::IfcPointOnCurve>())
+						return convertIfcPointOnCurve(point.as<typename IfcEntityTypesT::IfcPointOnCurve>());
+					// IfcPointOnSurface are not supported
 					throw oip::UnhandledException(point);
 				}
 
@@ -152,6 +156,22 @@ namespace OpenInfraPlatform {
                     // scale the lengths according to the unit conversion & return
 					return point * UnitConvert()->getLengthInMeterFactor();
                 }
+
+				/*! \brief Converts \c IfcPointOnCurve to a 3D vector.
+				 *
+				 * \param[in]	pointOnCurve			\c IfcPointOnCurve entity to be interpreted.
+				 *
+				 * \return								Calculated 3D vector.
+				 */
+				carve::geom::vector<3> convertIfcPointOnCurve(
+					const EXPRESSReference<typename IfcEntityTypesT::IfcPointOnCurve>& pointOnCurve
+				) const throw(...) {
+					std::shared_ptr<PlacementConverterT<IfcEntityTypesT>> placementConverter
+						= std::make_shared<PlacementConverterT<IfcEntityTypesT>>(this->GeomSettings(), this->UnitConvert());
+
+					CurveConverterT<IfcEntityTypesT> gridConv(this->GeomSettings(), this->UnitConvert(), placementConverter);
+					return gridConv.getPointOnCurve<typename IfcEntityTypesT::IfcCurve>(pointOnCurve->BasisCurve, pointOnCurve->PointParameter);
+				}
 
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC2)
 				/*! \brief Converts \c IfcPointByDistanceExpression to a 3D vector.
