### 2021-03-09 e5ab6ec
introducing move constructor to EXPRESSReference
