+++ EarlyBinding/src/EXPRESS/EXPRESSReference.h
@@ -50,6 +50,13 @@ public:
 	    
 	}
 
+	EXPRESSReference(EXPRESSReference&& other)
+		:
+		base(other),
+		refId{ other.refId },
+		model{ other.model }
+	{}
+
 	EXPRESSReference(const std::shared_ptr<EXPRESSEntity>& to, const std::shared_ptr<EXPRESSModel>& model) {
 		if (std::dynamic_pointer_cast<T>(to) != nullptr) {
 			this->operator=(constructInstance(to->getId(), model));
