### 2021-03-09 795b117
Some corrections as was proposed in the PR
