### 2021-03-09 5d3e584
refactored EXPRESSReference a bit
