### 2021-03-09 be11154
tweaked caching of georeferencing IFC data
