+++ Core/src/IfcGeometryConverter/GeorefConverter.h
@@ -40,7 +40,8 @@ template <
 >
 class GeoreferencingConverterT : public ConverterBaseT<IfcEntityTypesT>
 {
-	typedef std::pair<size_t, std::shared_ptr<oip::GeorefMetadata>> GeorefPair;
+	typedef std::pair<EXPRESSReference<typename IfcEntityTypesT::IfcGeometricRepresentationContext>, 
+					  std::shared_ptr<oip::GeorefMetadata>> GeorefPair;
 
 public:
 	//! Constructor
@@ -69,11 +70,11 @@ public:
 		do
 		{
 			// the interpreted data
-			GeorefPair georefMeta = convertGeoref(
+			const GeorefPair georefMeta = convertGeoref(
 				EXPRESSReference<typename IfcEntityTypesT::IfcCoordinateOperation>::constructInstance(georef->first, model));
 
 			// add to the parsed map
-			georefMetadata.insert( georefMeta );
+			georefMetadata.push_back( georefMeta );
 		} 
 		while ((georef = std::find_if(++georef, model->entities.end(), [](auto pair)
 			{ return std::dynamic_pointer_cast<typename IfcEntityTypesT::IfcCoordinateOperation>(pair.second) != nullptr; }
@@ -100,7 +101,7 @@ private:
 		// get the CoordinateRereferenceSystem
 		std::shared_ptr<oip::GeorefMetadata> georefMeta = convertCRS(coordOper->TargetCRS);
 		// the used GeometricRepresentationContext->ID
-		size_t geomContextID = 0;
+		EXPRESSReference<typename IfcEntityTypesT::IfcGeometricRepresentationContext> geomContext;
 
 		// get the GeometricContext
 		switch (coordOper->SourceCRS.which())
@@ -111,7 +112,7 @@ private:
 		}
 		case 1: // IfcGeometricRepresentationContext
 		{
-			geomContextID = coordOper->SourceCRS.get<1>()->getId();
+			geomContext = coordOper->SourceCRS.template get<1>();
 			break;
 		}
 		default:
@@ -155,7 +156,7 @@ private:
 			}
 		}
 
-		return GeorefPair(geomContextID, georefMeta);
+		return GeorefPair(geomContext, georefMeta);
 	}
 
 	std::shared_ptr<oip::GeorefMetadata> convertCRS(
@@ -204,8 +205,12 @@ private:
 	}
 
 
-	//! georefeferencing metadata (first = ID of GeometricRepresentationContext; second = interpreted GeorefMetadata)
-	std::map<size_t, std::shared_ptr<oip::GeorefMetadata>> georefMetadata;
+	//! georefeferencing metadata (first = pointer to IfcGeometricRepresentationContext; second = interpreted GeorefMetadata)
+	std::vector<GeorefPair
+		//std::pair<
+		//EXPRESSReference<typename IfcEntityTypesT::IfcRepresentationContext>,
+		//std::shared_ptr<oip::GeorefMetadata>>
+	> georefMetadata;
 
 };
 
