### 2021-03-09 0a192a9
Some corrections as was proposed in the PR
