### 2021-03-09 c472a92
introduced getContextPlacement function to account for georef position
