### 2021-03-09 d8bb82e
Reworked getDirection() and getAngle() ... by applying the concepts of Default Argument, and Conditional (or Ternary) Operator (or colloquial 'Immediate-if').
