### 2021-03-09 f5f402c
Fixed a few syntax mistakes in Doxygen comments.
