+++ Core/src/SplineInterpretation/SplineInterpretation.h
@@ -126,19 +126,19 @@ namespace OpenInfraPlatform
 				 */
 				std::vector<double> movingAverageVariableWindow(const std::vector<double>& data) const noexcept(true);
 
-				/* \brief Calculates the range of a variogramm.
+				/*! \brief Calculates the range of a variogramm.
 				 *
 				 * Calculates the semevariogramm variable with the values in \c data.
 				 * The lags are defined by the vector elements; a corresponding vector of x-values should be about equally spaced.
 				 * In this implementation, the range is obtaind as the position of the first maximum turning point.
 				 *
-				 * param[in]	data	A vector of \c double.
+				 * \param[in]	data	A vector of \c double.
 				 *
-				 * return	Range of a semivariogramm regarding the number of values.
+				 * \return	Range of a semivariogramm regarding the number of values.
 				 */
 				size_t variogrammGetRange(const std::vector<double>& data) const noexcept(true);
 
-				/* \brief Calculates the numeric derivative of a vector of x-y-pairs.
+				/*! \brief Calculates the numeric derivative of a vector of x-y-pairs.
 				 *
 				 * \param[in]	xy	A vector of pairs; the first pair value denotes x, second pair value denotes y. In example, the pair denotes {curveLength, curvature}.
 				 *
@@ -146,7 +146,7 @@ namespace OpenInfraPlatform
 				 */
 				std::vector<double> numericDerivative(const std::vector<std::pair<double, double>>& xy) const noexcept(true);
 
-				/* \brief Identifies the endpoints of the alignment elements based on the change of curvature.
+				/*! \brief Identifies the endpoints of the alignment elements based on the change of curvature.
 				 *
 				 * A curve (e.g. a B-Spline curve), described by its curvature and change of curvature, is analysed to find the start- and end-points of alignment elements.
 				 * Based on the relative value of the curvature change an alternate sequence of (1) transition curves and (2) straights or circular arcs are defined by their
@@ -163,7 +163,7 @@ namespace OpenInfraPlatform
 				std::vector<SplineInterpretationElement> identifyElementEndpoints(
 					const std::vector<std::pair<double, double>>& lengthsWithCurvatures, const std::vector<double>& curvatureChange) const noexcept(true);
 
-				/* \brief Examines the value of curvature change by a relative threshold.
+				/*! \brief Examines the value of curvature change by a relative threshold.
 				 *
 				 * If an absolute value is less than the threshold, the indicator is 0, otherwise 1. The threshold is obtained by the call of the function \c obtainThreshold.
 				 *
@@ -173,7 +173,7 @@ namespace OpenInfraPlatform
 				 */
 				std::tuple<std::vector<int>, double> indicateCurvatureChange(const std::vector<double>& curvatureChange) const noexcept(true);
 
-				/* \brief Returns the threshold value from a vector of \c double.
+				/*! \brief Returns the threshold value from a vector of \c double.
 				 *
 				 * The threshold is defined by the function as 25 % of the absolute magnitude.
 				 *
@@ -183,7 +183,7 @@ namespace OpenInfraPlatform
 				 */
 				double obtainThreshold(const std::vector<double>& data) const noexcept(true);
 
-				/* \brief Converts a vector of indicators to a vector of elements.
+				/*! \brief Converts a vector of indicators to a vector of elements.
 				 *
 				 * Each sequence of identical indicators is grouped as one element. This function doesn't care about the minimum length of elements.
 				 * Per element the information about start-/end-point, length, and corresponding id in 'lengthWithCurvature' is stored in a \c SplineInterpretationElement.
[…]
