### 2021-03-09 9f5808b
IfcModel now knows about its georeferencing
