### 2021-03-09 f390e8c
showing georeferencing metadata in the UI
