### 2021-03-09 1d19e35
Changed type of isCCW from int to bool. (isCCW = is counter clockwise)
