### 2021-03-10 9b80792
trying to support georefmeta with point clouds as well
