### 2021-03-11 4f9851b
Modified functions in SplineUtilities to static functions. Adjusted all calls of these functions.
