+++ Core/src/IfcGeometryConverter/SplineUtilities.h
@@ -57,12 +57,12 @@ namespace OpenInfraPlatform
 				 * \return		The array of curve points, which can be rendered in a viewport after correction by \c GeomUtils::appendPointsToCurve.
 				 */
 				 // B-Spline curve definition according to: http://mathworld.wolfram.com/B-Spline.html
-				std::vector<carve::geom::vector<3>> computeBSplineCurveWithKnots(
+				static std::vector<carve::geom::vector<3>> computeBSplineCurveWithKnots(
 					const int& order,
 					const std::vector<double>& knotArray,
 					const std::vector<carve::geom::vector<3>>& controlPoints,
 					const uint32_t& numCurvePoints,
-					const double& accuracy) const throw(...)
+					const double& accuracy) throw(...)
 				{
 					// The following parameters corresponds to the parameter t of a curve c(t)
 					double knotStart;
@@ -104,13 +104,13 @@ namespace OpenInfraPlatform
 				 * \return		The array of curve points, which can be rendered in a viewport after correction by \c GeomUtils::appendPointsToCurve.
 				 */
 				 // B-Spline curve definition according to: http://mathworld.wolfram.com/B-Spline.html
-				std::vector<carve::geom::vector<3>> computeRationalBSplineCurveWithKnots(
+				static std::vector<carve::geom::vector<3>> computeRationalBSplineCurveWithKnots(
 					const int& order,
 					const std::vector<double>& knotArray,
 					const std::vector<carve::geom::vector<3>>& controlPoints,
 					const std::vector<double>& weightsData,
 					const uint32_t& numCurvePoints,
-					const double& accuracy) const throw(...)
+					const double& accuracy) throw(...)
 				{
 					// The following parameters corresponds to the parameter t of a curve c(t)
 					double knotStart;
@@ -148,11 +148,11 @@ namespace OpenInfraPlatform
 				 *
 				 * \return		The B-Spline point at the position t.
 				 */
-				carve::geom::vector<3> computePointOfBSpline(
+				static carve::geom::vector<3> computePointOfBSpline(
 					const int& order,
 					const double& t,
 					const std::vector<carve::geom::vector<3>>& controlPoints,
-					const std::vector<double>& knotArray) const throw(...)
+					const std::vector<double>& knotArray) throw(...)
 				{
 					const size_t numControlPoints = controlPoints.size();
 
@@ -183,12 +183,12 @@ namespace OpenInfraPlatform
 				 *
 				 * \return		The rational-B-Spline point at the position t.
 				 */
-				carve::geom::vector<3> computePointOfRationalBSpline(
+				static carve::geom::vector<3> computePointOfRationalBSpline(
 					const int& order,
 					const double& t,
 					const std::vector<carve::geom::vector<3>>& controlPoints,
 					const std::vector<double>& knotArray,
-					const std::vector<double>& weightsData) const throw(...)
+					const std::vector<double>& weightsData) throw(...)
 				{
 					const size_t numControlPoints = controlPoints.size();
 
@@ -232,11 +232,11 @@ namespace OpenInfraPlatform
 				 *
 				 * \return		Vector of pairs with curve length and curvature.
 				 */
-				std::vector<std::pair<double, double>> computeCurvatureOfBSplineCurveWithKnots(
+				static std::vector<std::pair<double, double>> computeCurvatureOfBSplineCurveWithKnots(
 					const int& order,
 					const std::vector<carve::geom::vector<3>>& controlPoints,
 					const std::vector<double>& knotArray,
-					const uint32_t& numCurvePoints) const throw(...)
+					const uint32_t& numCurvePoints) throw(...)
 				{
[…]
