### 2021-03-13 a4b9fff
following e94c515, tweaked comparison operator for EXPRESSReference
