### 2021-03-13 95c2b14
throwing an error if an EXPRESSReference cannot be resolved
