### 2021-03-13 6e4ba93
returning char arrays
