### 2021-03-14 f6d3dd9
IFC reader now generated with some exception handling
