+++ ExpressBindingGenerator/src/Generator/GeneratorOIP.cpp
@@ -2201,6 +2201,7 @@ void GeneratorOIP::generateReaderFiles(const Schema & schema)
 	writeLine(file, "const size_t id = std::stoull(line.substr(1, line.find_first_of('=') - 1));");
 	writeLine(file, "const std::string entityType = line.substr(line.find_first_of('=') + 1, line.find_first_of('(') - line.find_first_of('=') - 1);");
 	//writeLine(file, "std::string parameters = line.substr(line.find_first_of('('), line.find_last_of(')') - line.find_first_of('(') + 1);");
+	writeLine(file, "try {");
 	for (size_t idx = 0; idx < schema.getEntityCount(); idx++) {
 		auto entity = schema.getEntityByIndex(idx);
 		if (!schema.isAbstract(entity)) {
@@ -2211,8 +2212,19 @@ void GeneratorOIP::generateReaderFiles(const Schema & schema)
 			writeLine(file, "}");
 		}
 	}
-	writeLine(file, "}"); //end if line[0] == '#'
-	writeLine(file, "}"); //end for read file
+	writeLine(file, "}"); //end try 
+	writeLine(file, "catch (const std::exception& ex) {"); // begin catch
+	writeLine(file, "#pragma omp critical");
+	writeLine(file, "errors.push_back(std::string(ex.what()));");
+	writeLine(file, "}"); //end catch 
+	writeLine(file, "} // end if line[0] == '#'");
+	writeLine(file, "} // end for read file");
+	writeLine(file, "if (errors.size() != 0)	{");
+	writeLine(file, "std::string all = "";");
+	writeLine(file, "std::for_each(errors.begin(), errors.end(), [&all](const std::string& el) {all += el + \"\\n\"; });");
+	writeLine(file, "all.pop_back();");
+	writeLine(file, "throw std::exception(all.c_str());");
+	writeLine(file, "}");
 	linebreak(file);
 	writeLine(file, "// Initialize inverse parameters");
 	writeLine(file, "size_t numEntities = model->entities.size();");
@@ -2221,16 +2233,16 @@ void GeneratorOIP::generateReaderFiles(const Schema & schema)
 	writeLine(file, "auto it = model->entities.begin();");
 	writeLine(file, "std::advance(it, i);");
 	writeLine(file, "it->second->linkInverse(model);");
-	writeLine(file, "}"); // end for each entity
+	writeLine(file, "} // end for each entity"); 
 	linebreak(file);
 	writeLine(file, "return model;");
 	writeLine(file, "}"); //end try 
-	writeLine(file, "catch(std::exception e) {"); // begin catch
-	writeLine(file, "std::cout << e.what() << std::endl;");
+	writeLine(file, "catch(const std::exception& ex) {"); // begin catch
+	writeLine(file, "throw ex;");
 	writeLine(file, "}"); //end catch 
 	writeLine(file, "}");
 	writeLine(file, "else {");
-	writeLine(file, "throw std::exception(\"Could not open file.\");");
+	writeLine(file, "throw std::exception(\"Could not open IFC file.\");");
 	writeLine(file, "}"); // end else if(file.is_open())
 	writeLine(file, "return nullptr;");
 	writeLine(file, "};"); // end function FromFile
