### 2021-03-14 82d7610
forgot to initialize errors array
