### 2021-03-14 cb3a5a4
Added files of unit test Added a .sketch file to test the program, the expected result is staved in the .txt file. A short description how to do the test is included in the .md file, because a implementation in a source file is still pending.
