### 2021-03-14 ebfac5a
Added an adaption of the console size before printing of the results While writing the description about the test (md-file), I recognized that the manual change of the Screen Buffer Size could be annoying to other users. The added code changes autonomous the Screen Buffer Height dynamic depending on the number of alignment elements. (Screen Buffer Height is the number of stored lines in the console. A scroll bar appears when the number of lines doesn't fit into the current window size.)
