### 2021-03-14 31a5bd4
Correcting mistakes, added rest of entities,added some UT and linked issues
