+++ Documentation/markdown/SupportedIFCrepresentations.md
@@ -26,90 +26,130 @@ Meaning of columns:
 
 | Entity                                | Covered              | Converter                  | Unit test             | Issue |
 |---------------------------------------|----------------------|----------------------------|-----------------------|-------|
-| `IfcAdvancedBrep`                     | :clock9:             | `SolidModelConverter`      | -                     | -     |
-| `IfcAdvancedBrepWithVoids`            | :clock9:             | `SolidModelConverter`      | -                     | -     |
-| `IfcAlignmentCurve`                   | :heavy_check_mark:   | `CurveConverter`           | -                     | -     |
+| `IfcAdvancedBrep`                     | :x:                  | `SolidModelConverter`      | -                     | -     |
+| `IfcAdvancedBrepWithVoids`            | :x:                  | `SolidModelConverter`      | -                     | -     |
+| `IfcAdvancedFace`                     | :x:                  | `SolidModelConverter`      | -                     | -     |
+| `IfcAlignmentCurve`                   | :heavy_check_mark:   | `CurveConverter`           | :heavy_check_mark:    | -     |
 | `IfcAnnotationFillArea`               | :x:                  | -                          | -                     | -     |
-| `IfcAxis1Placement`                   | :clock9:             | `PlacementConverter`       | -                     | -     |
-| `IfcAxis2Placement2D`                 | :heavy_check_mark:   | `PlacementConverter`       | -                     | -     |
-| `IfcAxis2Placement3D`                 | :heavy_check_mark:   | `PlacementConverter`       | -                     | -     |
+| `IfcAxis1Placement`                   | :x:                  | `PlacementConverter`       | :heavy_check_mark:    | -     |
+| `IfcAxis2Placement2D`                 | :heavy_check_mark:   | `PlacementConverter`       | :heavy_check_mark:    | -     |
+| `IfcAxis2Placement3D`                 | :heavy_check_mark:   | `PlacementConverter`       | :heavy_check_mark:    | -     |
 | `IfcAxis2PlacementLinear`             | :x:                  | `PlacementConverter`       | -                     | -     |
-| `IfcBlock`                            | :heavy_check_mark:   | `SolidModelConverter`      | -                     | -     |
+| `IfcBlock`                            | :heavy_check_mark:   | `SolidModelConverter`      | :heavy_check_mark:    | -     |
 | `IfcBlossCurve`                       | :x:                  | `CurveConverter`           | -                     | -     |
-| `IfcBooleanClippingResult`            | :clock9:             | `SolidModelConverter`      | -                     | -     |
-| `IfcBooleanResult`                    | :heavy_check_mark:   | `SolidModelConverter`      | -                     | -     |
-| `IfcBoundedCurve`                     | :heavy_check_mark:   | `CurveConverter`           | -                     | -     |
-| `IfcBoundingBox`                      | :heavy_check_mark:   | `SolidModelConverter`      | -                     | -     |
-| `IfcBoxedHalfSpace`                   | :clock9:             | `SolidModelConverter`      | -                     | -     |
-| `IfcBSplineCurveWithKnots`            | :heavy_check_mark:   | `SplineConverter.`         | -                     | -     |
+| `IfcBooleanClippingResult`            | :heavy_check_mark:   | `SolidModelConverter`      | :x:                   | -     |
+| `IfcBooleanResult`                    | :heavy_check_mark:   | `SolidModelConverter`      | :heavy_check_mark:    | -     |
+| `IfcBoundedCurve`                     | :heavy_check_mark:   | `CurveConverter`           | :x:                   | -     |
+| `IfcBoundingBox`                      | :heavy_check_mark:   | `SolidModelConverter`      | :x:                   | -     |
[…]
