### 2021-03-14 6035ec2
implemented support for IfcFixedReferenceSweptAreaSolid
