+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -665,6 +665,9 @@ namespace OpenInfraPlatform
 						solidModel.template as<typename IfcEntityTypesT::IfcSweptAreaSolid>();
 					// Get swept area and position (attributes 1-2). 
 					oip::EXPRESSReference<typename IfcEntityTypesT::IfcProfileDef> swept_area = swept_area_solid->SweptArea;
+					// swept area
+					std::shared_ptr<ProfileConverterT<IfcEntityTypesT>> profile_converter = profileCache->getProfileConverter(swept_area);
+					const std::vector<std::vector<carve::geom::vector<2> > >& profile_coords = profile_converter->getCoordinates();
 
 					carve::math::Matrix swept_area_pos(pos);	// check if local coordinate system is specified for extrusion
 					if (swept_area_solid->Position)
@@ -684,24 +687,104 @@ namespace OpenInfraPlatform
 						// TO DO: implement
 					}
 
-					/*
-					// (2/4) IfcFixedReferenceSweptAreaSolid SUBTYPE of IfcSweptAreaSolid
-					std::shared_ptr<typename IfcEntityTypesT::IfcFixedReferenceSweptAreaSolid> fixed_ref_swept_area_solid =
-						std::dynamic_pointer_cast<typename IfcEntityTypesT::IfcFixedReferenceSweptAreaSolid>(swept_area_solid);
-					if (fixed_ref_swept_area_solid) {
-						// Get directrix, start parameter, end parameter and fixed reference (attributes 3-6).
-						//std::shared_ptr<typename IfcEntityTypesT::IfcCurve> directrix =
-						//	fixed_ref_swept_area_solid->Directrix.lock();	// TO DO: formal proposition: if no StartParam or EndParam, Directrix has to be a bounded or closed curve. 
-						double start_param = fixed_ref_swept_area_solid->StartParam;	// TO DO: optional
-						double end_param = fixed_ref_swept_area_solid->EndParam;		// TO DO: optional
-						//std::shared_ptr<typename IfcEntityTypesT::IfcDirection> fixed_ref =
-						//	fixed_ref_swept_area_solid->FixedReference.lock();
 
-						// TO DO: implement//
+					// (2/4) IfcFixedReferenceSweptAreaSolid SUBTYPE of IfcSweptAreaSolid
+					if (swept_area_solid.template isOfType<typename IfcEntityTypesT::IfcFixedReferenceSweptAreaSolid>())
+					{
+						oip::EXPRESSReference<typename IfcEntityTypesT::IfcFixedReferenceSweptAreaSolid> fixed_ref_swept_area_solid =
+							swept_area_solid.template as<typename IfcEntityTypesT::IfcFixedReferenceSweptAreaSolid>();
+
+						// Get directrix, start parameter, end parameter and fixed reference (attributes 3-6).
+						std::vector<carve::geom::vector<3> > segment_start_points;
+						std::vector<carve::geom::vector<3> > basis_curve_points;
+						curveConverter->convertIfcCurve(
+							fixed_ref_swept_area_solid->Directrix, basis_curve_points, segment_start_points);
+
+						double start_param = fixed_ref_swept_area_solid->StartParam.value_or(0.0);	// TO DO: optional
+						double end_param = fixed_ref_swept_area_solid->EndParam.value_or(1.0);		// TO DO: optional
+
+						carve::geom::vector<3> fixedRef = placementConverter->convertIfcDirection(
+							fixed_ref_swept_area_solid->FixedReference);
+
+						std::shared_ptr<carve::input::PolyhedronData> polyhedron_data(new carve::input::PolyhedronData());
+						itemData->closed_polyhedrons.push_back(polyhedron_data);
+
+						// create vertices
[…]
