### 2021-03-15 #3061 @aothms
Hi @pjanck

Thanks for submitting this. This is obviously very welcomed work by the community. Before I check off the checklist, let me start with some general comments (more might follow, just wanted to get the conversation going).

The paper, nor documentation attempts at contrasting this work with the other libraries for similar purposes. Two of your reviewers work on software libraries with overlapping aims. What are the differences?

>  License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

The file is named `License.txt` I recommend renaming this to LICENSE (I think to have the automated Github license discovery kick in, it isn't recognized now it seems). Which GPL version is used actually?

It is said that the platforms is aimed at providing a development platform, but I find no example usage of how to use the platform as such as a software library (but i know there are some interesting gems in there wrt e.g schema-agnostic template programming)

The paper text mentions a cross platform solution, but only Windows is currently supported. This needs to be better documented and perhaps explained: why only 64bit, why only msvc 2015/17 (for community edition users this is painful as msoft only allows the latest (2019) version to be downloaded for free if you didn't download earlier versions before).

No tags / releases?

Python 2 is really really end of life now (what is Python used for actually as a dependency?)

I find the paper very heavy on self citations

The paper mentions EXPRESS parser, but there is no explanation given what this entails, e.g that it is the modelling langauge used to express IFC, but there is also IfcXML is this supported?

The paper and documentation give very little detail on the implementation strategies (e.g mapping of IFC entity to C++, e.g SELECT to boost::variant / visit_struct) and geometry interpretation (which entities are supported and mapped to what geometric/topological entities).

I think a reflection on usage of Carve as a geometric modelling language would be very appreciated by the community, wrt e.g precision handling, robustness of boolean ops, support for 10303-42 modelling constructs such as surface_curve_swept_area_solid, but perhaps more in favour of Carve things like performance, simplicity and ease of use.

Not all dependencies listed: tumcms/Blue-Framework, zlib, visit_struct, ...
