### 2021-03-17 7894faf
added visit_struct license
