+++ Documentation/joss/paper.bib
@@ -1,4 +1,20 @@
 
+% ISO 10303-11 - EXPRESS
+@techreport{iso1030311,
+    type = {Standard},
+    key = {ISO 10303-11:2004},
+    month = {nov},
+    year = {2004},
+    title = {{Industrial automation systems and integration -- Product data representation and exchange -- Part 11: Description methods: The EXPRESS language reference manual}},
+    volume = {2004},
+    address = {Geneva, CH},
+    institution = {International Organization for Standardization},
+    url = {https://www.iso.org/standard/38047.html},
+    note = {Accessed: 2020-12-24},
+}
+
+
+
 @inproceedings{Amann:2014:ECPPM,
   author = {Amann, J. and Flurl, M. and Jubierre, J. R. and Borrmann, A.},
   title = {An alignment meta-model for the comparison of alignment product models},
@@ -63,6 +79,20 @@
  year = {2016},
  url = {https://www.cloudcompare.org/doc/wiki},
  note = {Accessed: 2021-02-14}
+}
+
+% ISO 10303-21 STEP files
+@techreport{iso1030321,
+    type = {Standard},
+    key = {ISO 10303-21:2016},
+    month = {mar},
+    year = {2016},
+    title = {{Industrial automation systems and integration -- Product data representation and exchange -- Part 21: Implementation methods: Clear text encoding of the exchange structure}},
+    volume = {2016},
+    address = {Geneva, CH},
+    institution = {International Organization for Standardization},
+    url = {https://www.iso.org/standard/63141.html},
+    note = {Accessed: 2020-12-24},
 }
 
 
