+++ Documentation/joss/paper.md
@@ -47,10 +47,11 @@ OIP supports reading, visualization, navigating, and handling of:
 Multiple models can be loaded at once and compared between each other (see Figure \ref{fig:ifcpcd}).
 Their absolute position is accounted for, so the models can be checked against one another for internal differences based on location of elements.
 
-Additionally, OIP incorporates its own EXPRESS parser.
+Additionally, OIP incorporates its own EXPRESS parser, that consumes data model specifications defined following `@iso1030311`.
 This enables automatic code generation for IFC early-binding library [@Amann:2018:Diss; @Hecht:2019:FBI; @Amann:2018:Programming].
-As such, complete contents of IFC files can be interpreted and analysed.
+As such, complete contents of IFC files produced following `@iso1030321` can be interpreted and analysed.
 Moreover, type safety is guaranteed at compile time thus reducing the risk for bugs and errors.
+This is achieved by a carefully designed schema-agnostic template library.
 
 Let it be stated here that OIP serves as a prototypical playground for developments.
 The software architecture, features, and functionalities have been changed, added or removed as required along the way.
