+++ Documentation/joss/paper.bib
@@ -8,6 +8,7 @@
     title = {{Industrial automation systems and integration -- Product data representation and exchange -- Part 11: Description methods: The EXPRESS language reference manual}},
     volume = {2004},
     address = {Geneva, CH},
+    author = {{ISO}},
     institution = {International Organization for Standardization},
     url = {https://www.iso.org/standard/38047.html},
     note = {Accessed: 2020-12-24},
@@ -90,10 +91,27 @@
     title = {{Industrial automation systems and integration -- Product data representation and exchange -- Part 21: Implementation methods: Clear text encoding of the exchange structure}},
     volume = {2016},
     address = {Geneva, CH},
+    author = {{ISO}},
     institution = {International Organization for Standardization},
     url = {https://www.iso.org/standard/63141.html},
     note = {Accessed: 2020-12-24},
 }
+
+% LandInfra
+@techreport{landinfra,
+    type = {Standard},
+    key = {OGC 15-111r1},
+    month = {dec},
+    year = {2016},
+    title = {{OGC\textregistered Land and Infrastructure Conceptual Model Standard (LandInfra)}},
+    volume = {2016},
+    address = {Wayland, MA, USA},
+    author = {{OGC}},
+    institution = {Open Geospatial Consortium Inc.},
+    url = {https://www.ogc.org/standards/infragml},
+    note = {Accessed: 2021-01-29},
+}
+
 
 
 @inbook{vilgertshofer:2017:iwcce,
@@ -213,4 +231,14 @@
    booktitle={Proceedings of 6th international conference siBIM: Structured data is the new gold},
    location={online},
    pages={27--35}
+}
+
+
+@misc{okstra,
+  title={OKSTRA\textregistered Objektkatalog f\"{u}r das Stra{\ss}en- und Verkehrswesen},
+  url = {https://www.okstra.de/},
+  author={{BASt}},
+  year={2021},
+  institution = {Bundesanstallt f\"{u}r Stra{\ss}enwesen},
+  note={Accessed: 2021-03-17}
 }
