+++ Documentation/joss/paper.md
@@ -67,6 +67,7 @@ Major elements from previous developments were already realigned, while some (pr
 
 During the development of the IFC standard, a clear need for an independent software
  capable of consuming and producing IFC files according to the newly developed schemas was determined [@Amann:2015:OIP; @Amann:2015:ICCBEI].
+Additionally, an interface to other infrastructure data model standards like OKSTRA and LandInfra was needed [@okstra; @landinfra].
 OIP fulfils this role by being open source, using CMake as a build system generator to support various platforms,
  and being based on well known libraries (like Qt, Eigen, Carve and Boost).
 
