+++ Documentation/joss/paper.md
@@ -66,8 +66,9 @@ Major elements from previous developments were already realigned, while some (pr
 # Statement of Need
 
 During the development of the IFC standard, a clear need for an independent software
- capable of consuming and producing IFC files according to the newly developed schemas was determined [@Amann:2015:OIP; @Amann:2015:ICCBEI].
-Additionally, an interface to other infrastructure data model standards like OKSTRA and LandInfra was needed [@okstra; @landinfra].
+ capable of consuming and producing IFC files according to the newly developed schemas was determined.
+Additionally, an interface to other infrastructure data model standards
+ like OKSTRA and LandInfra was needed [@Amann:2015:OIP; @Amann:2015:ICCBEI; @okstra; @landinfra].
 OIP fulfils this role by being open source, using CMake as a build system generator to support various platforms,
  and being based on well known libraries (like Qt, Eigen, Carve and Boost).
 
