+++ Documentation/joss/paper.bib
@@ -76,11 +76,61 @@
 }
 
 @misc{CloudCompare,
- title = {Cloud Compare Wiki},
+ title = {CloudCompare},
  year = {2016},
- url = {https://www.cloudcompare.org/doc/wiki},
+ url = {https://www.cloudcompare.org/doc},
  note = {Accessed: 2021-02-14}
 }
+
+@misc{pcl,
+ title = {Point Cloud Library},
+ year = {2021},
+ url = {https://pointclouds.org/},
+ note = {Accessed: 2021-03-17}
+}
+
+@misc{paraview,
+ title = {ParaView},
+ year = {2021},
+ url = {https://www.paraview.org/},
+ note = {Accessed: 2021-03-17}
+}
+
+@misc{ifcopenshell,
+ title = {IfcOpenShell},
+ year = {2021},
+ url = {https://ifcopenshell.github.io/docs/},
+ note = {Accessed: 2021-03-17}
+}
+
+@misc{ifcjs,
+ title = {IFC.js},
+ year = {2021},
+ url = {https://agviegas.github.io/IFC.js/docs},
+ note = {Accessed: 2021-03-17}
+}
+
+@misc{ifcplusplus,
+ title = {IFC++},
+ year = {2021},
+ url = {https://ifcquery.com/},
+ note = {Accessed: 2021-03-17}
+}
+
+@misc{xbim,
+ title = {xbim toolkit},
+ year = {2021},
+ url = {https://docs.xbim.net/},
+ note = {Accessed: 2021-03-17}
+}
+
+@misc{blender,
+ title = {BlenderBIM Add-on},
+ year = {2021},
+ url = {https://blenderbim.org/},
+ note = {Accessed: 2021-03-17}
+}
+
 
 % ISO 10303-21 STEP files
 @techreport{iso1030321,
