### 2021-03-17 #3061 @pjanck
Hi @aothms

thank you for the comments. Please find our comments below. Additionally, the paper and the branch have been updated accordingly.

> The paper, nor documentation attempts at contrasting this work with the other libraries for similar purposes. Two of your reviewers work on software libraries with overlapping aims. What are the differences?

I see that I have skipped over the instruction "including to other software addressing related needs". I added a paragraph to the `Statement of Need` section. Hopefully, I've cited everything correctly.

> The file is named License.txt I recommend renaming this to LICENSE (I think to have the automated Github license discovery kick in, it isn't recognized now it seems). Which GPL version is used actually?

Thanks for the hint, changed accordingly. It is GPL v3.

> It is said that the platforms is aimed at providing a development platform, but I find no example usage of how to use the platform as such as a software library (but i know there are some interesting gems in there wrt e.g schema-agnostic template programming)

I'm not sure I understand this comment correctly. Can you please expand on it? Where is that said?

I've added the schema-agnostic template programming - shamelessly stolen. :)

> The paper text mentions a cross platform solution, but only Windows is currently supported. This needs to be better documented and perhaps explained: why only 64bit, why only msvc 2015/17 (for community edition users this is painful as msoft only allows the latest (2019) version to be downloaded for free if you didn't download earlier versions before).

We have tried to update to vs2019, see https://github.com/tumcms/Open-Infra-Platform/pull/319 . The clash between the dependencies is stopping the progress, so the update is currently put on ice so that the dependencies clear out their problems. As a consequence we are staying with the versions we are sure that work.

> No tags / releases?

Not yet. We will do the first one when we are done with the review so that it can be referenced in the paper.

> Python 2 is really really end of life now (what is Python used for actually as a dependency?)

It is needed for a workaround for updating Boost: https://github.com/tumcms/Open-Infra-Platform/blob/development/cmake/UpdateBoostMPL.cmake .

> I find the paper very heavy on self citations.

Two reasons:

- There aren't many others using the tool. So the "self" is sort of mandatory.
- The guidelines provide these instructions:
    - "Your paper should not focus on new research results accomplished with the software."
    - "Write a short paper in Markdown format using `paper.md` as file name, including a title, summary, author names, affiliations, and *key references* [own emphasis]."
    - "Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it."

I'm can remove some, if there are too many. However, with these references, there is no need to lengthen the paper by including their results.

> The paper mentions EXPRESS parser, but there is no explanation given what this entails, e.g that it is the modelling langauge used to express IFC, but there is also IfcXML is this supported?

Do you mean the XSD schema? We only support EXPRESS. I added a reference to the standard - ISO 10303-11.

Also, the paper has this sentence explaining what the parser is useful for: "This enables automatic code generation for IFC early-binding library." I've included a sketch from Amann's dissertation that should help with the explanations.

> The paper and documentation give very little detail on the implementation strategies (e.g mapping of IFC entity to C++, e.g SELECT to boost::variant / visit_struct) and geometry interpretation (which entities are supported and mapped to what geometric/topological entities).

I believe that that would be too much content for a 250-1000 words paper. I agree that the documentation is scarce - this is one of the reasons we're submitting this paper.

- Mapping of SELECTs and the EXPRESS parsing: see the reference [@Hecht:2019:FBI]
- Which entities are supported: see https://github.com/tumcms/Open-Infra-Platform/pull/397

> I think a reflection on usage of Carve as a geometric modelling language would be very appreciated by the community, wrt e.g precision handling, robustness of boolean ops, support for 10303-42 modelling constructs such as surface_curve_swept_area_solid, but perhaps more in favour of Carve things like performance, simplicity and ease of use.

I've included a mention to Carve in the paper. Or are you having something bigger in mind?

> Not all dependencies listed: tumcms/Blue-Framework, zlib, visit_struct, ...

Would you expect a list in the paper? The list can be found at the end of the main Readme. I tried to not repeat information that is available in the repository.
