+++ Documentation/joss/paper.bib
@@ -26,7 +26,7 @@
 
 @inproceedings{Singer:2014:IfcAlignment,
     Author       = {Singer, D. and Amann, J.},
-    Title        = {Erweiterung von IFC Alignment um Stra{\ss}enquerschnitte},
+    Title        = {{Erweiterung von IFC Alignment um Stra{\ss}enquerschnitte}},
     Booktitle    = {Proceedings of the 26th Forum Bauinformatik},
     Year         = {2014},
     Address      = {Darmstadt, Germany}
@@ -36,7 +36,7 @@
 
 @inproceedings{Amann:2015:OKSTRA,
     Author       = {Amann, J. and Borrmann, A.},
-    Title        = {Open BIM for Infrastructure {--} mit OKSTRA und IFC Alignment zur internationalen Standardisierung des Datenaustauschs},
+    Title        = {{Open BIM for Infrastructure -- mit OKSTRA und IFC Alignment zur internationalen Standardisierung des Datenaustauschs}},
     Booktitle    = {Tagungsband zum 6. OKSTRA-Symposium},
     Year         = {2015},
     Address      = {K{\"o}ln, Germany}
@@ -153,7 +153,7 @@
     key = {OGC 15-111r1},
     month = {dec},
     year = {2016},
-    title = {{OGC\textregistered Land and Infrastructure Conceptual Model Standard (LandInfra)}},
+    title = {{OGC Land and Infrastructure Conceptual Model Standard (LandInfra)}},
     volume = {2016},
     address = {Wayland, MA, USA},
     author = {{OGC}},
@@ -179,7 +179,7 @@
 @techreport{Hecht:2018:PCDtoAlignment,
   author    = {Helge Hecht},
   type = {Technical report},
-  title     = {From Point Cloud Data to IfcAlignment},
+  title     = {{From Point Cloud Data to IfcAlignment}},
   pages     = {26 pp.},
   year      = {2018},
   url = {https://www.cms.bgu.tum.de/images/teaching/abim_seminar/Report_Hecht_From_Point_Cloud_to_IfcAlignment.pdf}
@@ -210,7 +210,7 @@
 @techreport{Beetz:2019:linkeddata,
   author    = {Jakob Beetz and Julian Amann and Andr{\'{e}} Borrmann},
   type = {Technical report},
-  title     = {Linked Data - Analyse von Einsatzm{\"o}glichkeiten von verbundenen Informationen (Linked Data) und Ontologien und damit befassten Technologien (Semantic Web) im Bereich des Stra{\ss}enwesens},
+  title     = {{Linked Data -- Analyse von Einsatzm{\"o}glichkeiten von verbundenen Informationen (Linked Data) und Ontologien und damit befassten Technologien (Semantic Web) im Bereich des Stra{\ss}enwesens}},
   pages     = {82 pp.},
   year      = {2019},
   url = {https://bast.opus.hbz-nrw.de/opus45-bast/frontdoor/deliver/index/docId/2181/file/LinkedData_Schlussbericht_01.0195.pdf}
@@ -285,7 +285,7 @@
 
 
 @misc{okstra,
-  title={OKSTRA\textregistered Objektkatalog f\"{u}r das Stra{\ss}en- und Verkehrswesen},
+  title={{OKSTRA: Objektkatalog f\"{u}r das Stra{\ss}en- und Verkehrswesen}},
   url = {https://www.okstra.de/},
   author={{BASt}},
   year={2021},
