### 2021-03-20 #3061 @aothms
Hi @pjanck thanks for the quick reply

> > It is said that the platforms is aimed at providing a development platform, but I find no example usage of how to use the platform as such as a software library (but i know there are some interesting gems in there wrt e.g schema-agnostic template programming)

> I'm not sure I understand this comment correctly. Can you please expand on it? Where is that said?

I was referring to `Let it be stated here that OIP serves as a prototypical playground for developments' in the article text. If I were to start on a new development (let's say provide an implementation for IfcBlossCurve) where do I begin?

> I added a reference to the standard - ISO 10303-11.

11 is the schema language. Do you mean part 21 for the SPF file format (vs part 28 for XML)?

> Which entities are supported: see tumcms/Open-Infra-Platform#397

Ok, so that's the unmerged [0] that's obviously vital information. I don't think it's as binary as this, glossing over some of the code I see attributes not being handled e.g [1] and contradicting log messages [2]. But it's a start, let's get this merged and referenced from the docs.

[0] https://github.com/tumcms/Open-Infra-Platform/blob/31a5bd4cd43906e86326b422d1121e7638d60b42/Documentation/markdown/SupportedIFCrepresentations.md
[1] https://github.com/tumcms/Open-Infra-Platform/blob/c0f621c18838b249fa6281dbb3cc8b4ffe95dc72/Core/src/IfcGeometryConverter/SolidModelConverter.h#L816
[2] https://github.com/tumcms/Open-Infra-Platform/blob/c0f621c18838b249fa6281dbb3cc8b4ffe95dc72/Core/src/IfcGeometryConverter/SolidModelConverter.h#L823

> I've included a mention to Carve in the paper. Or are you having something bigger in mind?

Yes I was hoping for more of a reflection also on the usage of Carve. To me, OpenCascade or CGAL are much more natural choices, but also a lot more complex and probably less performant. This kind of first hand usage experience I think is very valuable.

> Would you expect a list in the paper? The list can be found at the end of the main Readme. I tried to not repeat information that is available in the repository.

No readme is fine. I'm not seeing this `visit_struct` thing in the readme, I didn't do a full side by side check though.
