### 2021-03-21 #3061 @pjanck
Hi @aothms , great review!

> I was referring to `Let it be stated here that OIP serves as a prototypical playground for developments' in the article text. If I were to start on a new development (let's say provide an implementation for IfcBlossCurve) where do I begin?

Good point. I'm not quite sure how we should/will address this. I believe that a sketch about the program's architecture would be beneficial. Need to take a closer look at this.

> 11 is the schema language. Do you mean part 21 for the SPF file format (vs part 28 for XML)?

Correct, we parse the schema (specified according to part 11) and generate C++ classes and a reader that reads in part 21 and fills the runtime data. This is also the reason we don't have any IFC classes committed in the repository. See https://github.com/tumcms/Open-Infra-Platform/blob/joss_submission/Documentation/images/express_parser.png (included in the newest version of the paper as Figure 2). We don't support part 28.

> Ok, so that's the unmerged [0] that's obviously vital information. I don't think it's as binary as this, glossing over some of the code I see attributes not being handled e.g [1] and contradicting log messages [2]. But it's a start, let's get this merged and referenced from the docs.

About [0]: Yes, it is a WIP. The student has her exams, so we're not pressuring her. Hopefully, with April, she will get to it and we will merge it.
About [1] & [2]: Another WIP, another student, similar timeline: https://github.com/tumcms/Open-Infra-Platform/pull/357

It is far from optimal, I agree. But we're on a good track, I believe.

> Yes I was hoping for more of a reflection also on the usage of Carve. To me, OpenCascade or CGAL are much more natural choices, but also a lot more complex and probably less performant. This kind of first hand usage experience I think is very valuable.

We'll need to get back to you on this one, I can't remember from the top of my head, why `carve` was chosen.

> No readme is fine. I'm not seeing this `visit_struct` thing in the readme, I didn't do a full side by side check though.

I see. `visit_struct` was missing, I've added it only in the `joss_submission` branch: https://github.com/tumcms/Open-Infra-Platform/tree/joss_submission - this branch will be merged shortly so that the changes will be visible also within the default branch. The branch will still be continued with further improvements as discussed above.
