### 2021-03-21 tumcms/Open-Infra-Platform#413 @pjanck
[REQUEST] Program architecture graphic
A clear overview of the general architecture of the solution is missing.

**Describe the solution you'd like**
A graphic that entails the base blocks of the program, what functionality it provides, to help newcomers get an overview.

**Additional context**
See also: https://github.com/openjournals/joss-reviews/issues/3061#issuecomment-803536122
