### 2021-03-21 c657427
dealing with bounding box precision and calculation
