### 2021-03-21 c5eeab9
moved license notices from main readme to a separate document
