+++ Licenses/readme.md
@@ -1,0 +1,27 @@
+
+# License
+
+## TUM Open Infra Platform
+
+The license can be found here: `{root_dir}/LICENSE`
+
+## Other libraries
+
+TUM Open Infra Platform uses several software libraries. The corresponding licenses can be found in the Licenses folder distributed with this source code.
+
+| Library with version | Link | License name & version | Local file in Licenses folder |
+|:---------------------|:-----|:-----------------------|:------------------------------|
+| boost 1.65.1 | http://www.boost.org/ | Boost Software License | `boost.LICENSE_1_0.txt` |
+| Eigen 3.3.7  | http://eigen.tuxfamily.org/ | Mozilla Public License 2 | `Eigen.COPYING.MPL2` |
+| Google Test 1.8.0 | https://github.com/google/googletest | BSD 3-Clause License | `GoogleTest.LICENSE` |
+| IfcPlusPlus  | https://github.com/ifcquery/ifcplusplus | MIT License | https://github.com/ifcquery/ifcplusplus/blob/master/LICENSE.txt |
+| Qt 5.12.1    | https://download.qt.io/archive/qt/ | LGPL version 3 | `Qt.LICENSE` |
+| BlueFramwork a0.0.9 | https://github.com/tumcms/Blue-Framework | GPL v3 License | `BlueFramwork.LICENSE.txt` |
+| CloudCompare v2.10.3 | http://www.danielgm.net/cc/ | GPL v2 License) on development branch IDP for PCD processing. | `cloudcompare.license.txt` |
+| zlib 1.2.11  | https://zlib.net/ | Zlib license | `zlib.license.txt` |
+| TCLAP        | http://tclap.sourceforge.net/ | MIT License | `TCLAP.COPYING` |
+| qhull        | http://www.qhull.org/ | qhull license | `qhull.COPYING.txt` |
+| raptor2      | http://librdf.org/raptor/ | GPL v2 License | `raptor2.COPYING.txt` |
+| libLAS 1.6   | https://liblas.org/ | BSD License | `liblas.LICENSE.txt` |
+| Doxygen 1.8.18 | http://doxygen.nl/download.html | GNU General Public License | `doxygen.license.txt` |
+| visit_struct | https://github.com/garbageslam/visit_struct | Boost Software License | `visit_struct.license.txt` |
