+++ Readme.md
@@ -25,35 +25,4 @@ Out-of-sync:
 
 ## License
 
-Copyright (c) 2021 Technical University of Munich
-Chair of Computational Modeling and Simulation.
-
-TUM Open Infra Platform is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License Version 3
-as published by the Free Software Foundation.
-
-TUM Open Infra Platform is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-## Copyright notes
-TUM Open Infra Platform uses several software libraries. The corresponding licenses can be found in the Licenses folder distributed with this source code:
-
-* boost 1.65.1 (http://www.boost.org/) (*Boost Software License*) ({root_dir}/Licenses/boost.LICENSE_1_0.txt)
-* Eigen 3.3.7 (http://eigen.tuxfamily.org/) (*MPL2*) ({root_dir}/Licenses/Eigen.COPYING.MPL2)
-* Google Test 1.8.0 (https://github.com/google/googletest) (*BSD 3-Clause License*) ({root_dir}/Licenses/GoogleTest.LICENSE)
-* IfcPlusPlus (https://github.com/ifcquery/ifcplusplus) (MIT License) (https://github.com/ifcquery/ifcplusplus/blob/master/LICENSE.txt)
-* Qt 5.12.1 (https://download.qt.io/archive/qt/) (LGPL version 3) ({root_dir}/Licenses/Qt.LICENSE)
-* BlueFramwork a0.0.2 (https://github.com/tumcms/Blue-Framework) (GPL v3 License) ({root_dir}/Licenses/BlueFramwork.LICENSE.txt)
-* CloudCompare v2.10.3 (http://www.danielgm.net/cc/) (GPL v2 License) on development branch IDP for PCD processing. ({root_dir}/Licenses/cloudcompare.license.txt)
-* zlib 1.2.11 (https://zlib.net/) (Zlib license) ({root_dir}/Licenses/zlib.license.txt)
-* TCLAP (http://tclap.sourceforge.net/) (MIT License) ({root_dir}/Licenses/TCLAP.COPYING)
-* qhull (http://www.qhull.org/) without modification. ({root_dir}/Licenses/qhull.COPYING.txt)
-* raptor2 (http://librdf.org/raptor/) (GPL v2 License) ({root_dir}/Licenses/raptor2.COPYING.txt)
-* libLAS 1.6 (https://liblas.org/) (BSD License) ({root_dir}/Licenses/liblas.LICENSE.txt)
-* Doxygen 1.8.18 (http://doxygen.nl/download.html) (GNU General Public License) ({root_dir}/Licenses/doxygen.license.txt)
-* visit_struct (https://github.com/garbageslam/visit_struct) (Boost Software License) ({root_dir}/Licenses/visit_struct.license.txt)
+See TUM Open Infra Platform's [license](./LICENSE) and [third party licenses](./Licenses/readme.md).
