### 2021-03-21 9697b3f
included IfcPlusPlus license
