+++ Licenses/readme.md
@@ -7,21 +7,23 @@ The license can be found [here](../LICENSE).
 
 ## Other libraries
 
-TUM Open Infra Platform uses several software libraries. The corresponding licenses can be found in the Licenses folder distributed with this source code.
+TUM Open Infra Platform uses several software libraries. The table below lists their licenses in alphabetical order.
+
+The corresponding licenses can be found in the Licenses folder distributed with this source code.
 
 | Library with version | Link | License name & version | Local copy |
 |:---------------------|:-----|:-----------------------|:------------------------------|
+| BlueFramwork a0.0.9 | https://github.com/tumcms/Blue-Framework | GPL v3 License | [license](./BlueFramwork.LICENSE.txt) |
 | boost 1.65.1 | http://www.boost.org/ | Boost Software License | [license](./boost.LICENSE_1_0.txt) |
+| CloudCompare v2.10.3 | http://www.danielgm.net/cc/ | GPL v2 License) on development branch IDP for PCD processing. | [license](./cloudcompare.license.tx)t |
+| Doxygen 1.8.18 | http://doxygen.nl/download.html | GNU General Public License | [license](./doxygen.license.txt) |
 | Eigen 3.3.7  | http://eigen.tuxfamily.org/ | Mozilla Public License 2 | [license](./Eigen.COPYING.MPL2) |
 | Google Test 1.8.0 | https://github.com/google/googletest | BSD 3-Clause License | [license](./GoogleTest.LICENSE) |
 | IfcPlusPlus  | https://github.com/ifcquery/ifcplusplus | MIT License | [license](./IfcPlusPlus.LICENSE.txt) |
-| Qt 5.12.1    | https://download.qt.io/archive/qt/ | LGPL version 3 | [license](./Qt.LICENSE) |
-| BlueFramwork a0.0.9 | https://github.com/tumcms/Blue-Framework | GPL v3 License | [license](./BlueFramwork.LICENSE.txt) |
-| CloudCompare v2.10.3 | http://www.danielgm.net/cc/ | GPL v2 License) on development branch IDP for PCD processing. | [license](./cloudcompare.license.tx)t |
-| zlib 1.2.11  | https://zlib.net/ | Zlib license | [license](./zlib.license.txt) |
-| TCLAP        | http://tclap.sourceforge.net/ | MIT License | [license](./TCLAP.COPYING) |
-| qhull        | http://www.qhull.org/ | qhull license | [license](./qhull.COPYING.txt) |
-| raptor2      | http://librdf.org/raptor/ | GPL v2 License | [license](./raptor2.COPYING.txt) |
 | libLAS 1.6   | https://liblas.org/ | BSD License | [license](./liblas.LICENSE.txt) |
-| Doxygen 1.8.18 | http://doxygen.nl/download.html | GNU General Public License | [license](./doxygen.license.txt) |
+| qhull        | http://www.qhull.org/ | qhull license | [license](./qhull.COPYING.txt) |
+| Qt 5.12.1    | https://download.qt.io/archive/qt/ | LGPL version 3 | [license](./Qt.LICENSE) |
+| raptor2      | http://librdf.org/raptor/ | GPL v2 License | [license](./raptor2.COPYING.txt) |
+| TCLAP        | http://tclap.sourceforge.net/ | MIT License | [license](./TCLAP.COPYING) |
 | visit_struct | https://github.com/garbageslam/visit_struct | Boost Software License | [license](./visit_struct.license.txt) |
+| zlib 1.2.11  | https://zlib.net/ | Zlib license | [license](./zlib.license.txt) |
