### 2021-03-21 1b1be2a
Update licence notes
