### 2021-03-22 #3061 @pjanck
Hi @aothms - as promised, our answer to the `carve` question:

> Yes I was hoping for more of a reflection also on the usage of Carve. To me, OpenCascade or CGAL are much more natural choices, but also a lot more complex and probably less performant. This kind of first hand usage experience I think is very valuable.

It was a design decision made by `IfcPlusPlus` creators, whose code we've incorporated into our code and expanded on it. We've also updated the `readme` on the licenses to better reflect our usage of their code - see https://github.com/tumcms/Open-Infra-Platform/tree/development/Licenses .
