### 2021-03-22 #3061 @aothms
Thanks @pjanck I think that kind of proves the point as IfcPlusPlus later also adopted OpenCASCADE https://github.com/ifcquery/ifcplusplus/blob/master/IfcPlusPlus/src/ifcpp/geometry/OCC/GeometryConverterOCC.h not sure how precisely, I think carve is still the main geometry library, but not sure. So it seems like implementations are investigating various geometry libraries. It would be great if first hand experience in using these libraries, their strengths and weaknesses becomes more explicit.
