### 2021-03-22 #3061 @pjanck
I agree that a report on the first-hand experience would be beneficial, however: I don't feel we're in a good position to report on the strengths and weaknesses of geometry libraries, since they were not the focus of our developments and we haven't made any comparisons ourselves. We merely used what was available and tweaked it to compile and work with the existing code.
