+++ Base/src/GeorefMetadata.h
@@ -64,8 +64,8 @@ public:
 	{
 		return carve::math::Matrix(
 			1.0, 0.0, 0.0, 0.0,
-			0.0, cos(angle), sin(angle), 0.0,
-			0.0, sin(angle), -cos(angle), 0.0,
+			0.0, cos(angle), -sin(angle), 0.0,
+			0.0, sin(angle), cos(angle), 0.0,
 			0.0, 0.0, 0.0, 1.0);
 	}
 
