### 2021-03-22 d0988e9
supporting GeometricSubContext for georeferencing
