+++ Base/src/GeorefMetadata.h
@@ -63,9 +63,9 @@ public:
 	carve::math::Matrix rotation()
 	{
 		return carve::math::Matrix(
-			1.0, 0.0, 0.0, 0.0,
-			0.0, cos(angle), -sin(angle), 0.0,
-			0.0, sin(angle), cos(angle), 0.0,
+			cos(angle), -sin(angle), 0.0, 0.0,
+			sin(angle), cos(angle), 0.0, 0.0,
+			0.0, 0.0, 1.0, 0.0,
 			0.0, 0.0, 0.0, 1.0);
 	}
 
