### 2021-03-22 ab51bd0
moved the application of georeferencing to allow for mapped repr.
