### 2021-04-03 d6d75a2
moved some functions
