+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -2212,16 +2212,6 @@ namespace OpenInfraPlatform
 				*/
 			}
 
-			void simplifyMesh(std::shared_ptr<carve::mesh::MeshSet<3>>& meshset)
-			{
-				bool rebuild = true;
-				int num_vertices = meshset->vertex_storage.size();
-				if (rebuild && num_vertices > 8)
-				{
-					retriangulateMeshSet(meshset);
-				}
-			}
-
 			bool computeCSG(carve::mesh::MeshSet<3>* op1,
 				carve::mesh::MeshSet<3>* op2,
 				const carve::csg::CSG::OP operation,
@@ -2319,159 +2309,6 @@ namespace OpenInfraPlatform
 				box_data->addFace(7, 3, 2);
 			}
 
-			static void retriangulateMeshSet(std::shared_ptr<carve::mesh::MeshSet<3>>& meshset)
-			{
-				//carve::poly::Polyhedron* poly = carve::polyhedronFromMesh(meshset, -1);
-				//carve::mesh::MeshSet<3>* meshset_new = carve::meshFromPolyhedron(poly, -1);
-
-				int num_vertices1 = meshset->vertex_storage.size();
-				std::shared_ptr<carve::input::PolyhedronData> poly_data(new carve::input::PolyhedronData());
-				std::map<double, std::map<double, std::map<double, int> > > existing_vertices_coords;
-				std::map<double, std::map<double, std::map<double, int> > >::iterator vert_it;
-				std::map<double, std::map<double, int> >::iterator it_find_y;
-				std::map<double, int>::iterator it_find_z;
-				std::map<int, int> map_merged_idx;
-				double volume_check = 0;
-
-				for (size_t ii = 0; ii < meshset->meshes.size(); ++ii)
-				{
-					carve::mesh::Mesh<3>* mesh = meshset->meshes[ii];
-					volume_check += mesh->volume();
-					std::vector<carve::mesh::Face<3>* >& vec_faces = mesh->faces;
-
-					for (size_t i2 = 0; i2 < vec_faces.size(); ++i2)
-					{
-						carve::mesh::Face<3>* face = vec_faces[i2];
-						std::vector<int> face_idx;
-
-						carve::geom3d::Vector normal = face->plane.N;
-
-						std::vector<carve::geom::vector<2> > verts2d;
-						face->getProjectedVertices(verts2d);
-
-						// check winding order
-						//carve::geom3d::Vector normal_2d = GeomUtils::computePolygon2DNormal( verts2d );
-						//if( normal_2d.z < 0 )
-						//{
-						//	std::reverse( verts2d.begin(), verts2d.end() );
-						//}
-
-						std::vector<carve::triangulate::tri_idx> triangulated;
-						if (verts2d.size() > 3)
-						{
-							try
-							{
-								carve::triangulate::triangulate(verts2d, triangulated);
-								carve::triangulate::improve(verts2d, triangulated);
-							}
-							catch (...)
-							{
-								std::cout //<< __FUNC__
-									<< " carve::triangulate::incorporateHolesIntoPolygon failed " << std::endl;
-								continue;
-							}
-						}
-						else
-						{
-							triangulated.push_back(carve::triangulate::tri_idx(0, 1, 2));
-						}
-
-						// now insert points to polygon, avoiding points with same coordinates
-						int i_vert = 0;
-						carve::mesh::Edge<3>* edge = face->edge;
-						do
-						{
-							const carve::geom::vector<3>& v = edge->vert->v;//verts3d[i]->v;
-							edge = edge->next;
-
-							//#ifdef ROUND_IFC_COORDINATES
[…]
