### 2021-04-03 57a2c64
simplifying functions
