### 2021-04-03 1425cf8
upgrading inputs to functions, some errors remain
