### 2021-04-03 d7c2e20
a single IFC file can now produce multiple IfcModels (depends on Georef)
