+++ Core/src/IfcGeometryConverter/IfcImporterImpl.h
@@ -66,11 +66,11 @@ IfcImporterT<IfcEntityTypesT>::IfcImporterT<IfcEntityTypesT>()
 template <
 	class IfcEntityTypesT
 >
-std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>> IfcImporterT<IfcEntityTypesT>::convertIfcProduct(
+std::vector<std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>>> IfcImporterT<IfcEntityTypesT>::convertIfcProduct(
 	const EXPRESSReference<typename IfcEntityTypesT::IfcProduct>& product
 ) const
 {
-	std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>> productShape = std::make_shared<ShapeInputDataT<IfcEntityTypesT>>();
+	std::vector<std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>>> shapes;
 	try
 	{
 		carve::math::Matrix productPlacement(carve::math::Matrix::IDENT());
@@ -96,30 +96,53 @@ std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>> IfcImporterT<IfcEntityTypesT>:
 #ifdef _DEBUG
 				BLUE_LOG(trace) << "Processing " << rep->getErrorLog();
 #endif
+
+				if (!rep->ContextOfItems.template isOfType<typename IfcEntityTypesT::IfcGeometricRepresentationContext>())
+					// ignore
+					continue;
+
 				// apply global position (georeferencing given by the context of the representation)
 				const carve::math::Matrix contextPlacement =
 					repConverter->getContextPlacement(rep->ContextOfItems);
 
+				// the return value
+				std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>> productShape 
+					= std::make_shared<ShapeInputDataT<IfcEntityTypesT>>(product, rep, 
+						rep->ContextOfItems.template as<typename IfcEntityTypesT::IfcGeometricRepresentationContext>());
+
 				// convert
 				repConverter->convertIfcRepresentation(rep, contextPlacement * productPlacement, productShape);
+
 #ifdef _DEBUG
 				BLUE_LOG(trace) << "Processed " << rep->getErrorLog();
 #endif
+				computeMeshsetsFromPolyhedrons(productShape);
+				shapes.push_back(productShape);
 			}
 
-			computeMeshsetsFromPolyhedrons(productShape);
 #ifdef _DEBUG
 			BLUE_LOG(trace) << "Processed " << representation->getErrorLog();
 #endif
 		}
 
 		if (product.template isOfType<typename IfcEntityTypesT::IfcAlignment>()) {
+			auto alignment = product.template as<typename IfcEntityTypesT::IfcAlignment>();
+			auto axis = alignment->
+				Axis.as<typename IfcEntityTypesT::IfcGeometricRepresentationItem>();
+			// the return value
+			std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>> productShape
+				= std::make_shared<ShapeInputDataT<IfcEntityTypesT>>(product);
 			std::shared_ptr<ItemData> itemData(new ItemData());
-			productShape->vec_item_data.push_back(itemData);
+			// interpret the alignment
 			repConverter->convertIfcGeometricRepresentationItem(
-				product.template as<typename IfcEntityTypesT::IfcAlignment>()->
-					Axis.as<typename IfcEntityTypesT::IfcGeometricRepresentationItem>(),
+				axis,
 				productPlacement, itemData);
+			// add if the conversion was successful
+			if( !itemData->empty() )
+			{
+				productShape->getData().push_back(itemData);
+				shapes.push_back(productShape);
+			}
 		}
 	}
[…]
