### 2021-04-07 6b789bd
Add Button to UI that will open the Ifc Tree once it is implemented
