### 2021-04-08 26ab1b6
collecting schemas into a source group
