+++ CMakeLists.txt
@@ -391,6 +391,7 @@ set(OpenInfraPlatform_UserInterface_Forms
 	UserInterface/Forms/MainWindow.ui
 	UserInterface/Forms/LicenseAndCopyrightInformation.ui
 	UserInterface/Forms/PreferencesDialog.ui
+	UserInterface/Forms/IfcTreeDialog.ui
 
 	# Currently not included.
 	#UserInterface/Forms/CreateAccidentReport.ui
