### 2021-04-08 bd586b8
Add files of the unit test fixed-reference-swept-area-solid It's the initial commit for that unit test which includes the source file and all expected result images.
