### 2021-04-09 c348e3c
Starting IfcClothoid
