### 2021-04-09 2387563
Starting IfcClothoid
