### 2021-04-09 ef58ac5
Starting IfcClothoid
