### 2021-04-10 f9454b2
Refactoring convertIfcSweptDiskSolid
