+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -1266,77 +1266,77 @@ namespace OpenInfraPlatform
 				} //endif sweptDiskSolid_polygonal
 
 				// TO DO: understand what happens here
-				std::vector<carve::geom::vector<3> > segment_start_points;
-				std::vector<carve::geom::vector<3> > basis_curve_points;
-				curveConverter->convertIfcCurve(directrix_curve, basis_curve_points, segment_start_points);
+				std::vector<carve::geom::vector<3>> segmentStartPoints;
+				std::vector<carve::geom::vector<3>> basisCurvePoints;
+				curveConverter->convertIfcCurve(directrixCurve, basisCurvePoints, segmentStartPoints);
 
-				std::shared_ptr<carve::input::PolyhedronData> pipe_data(new carve::input::PolyhedronData());
-				itemData->closed_polyhedrons.push_back(pipe_data);
-				std::vector<carve::geom::vector<3> > inner_shape_points;
+				std::shared_ptr<carve::input::PolyhedronData> pipeData(new carve::input::PolyhedronData());
+				itemData->closed_polyhedrons.push_back(pipeData);
+				std::vector<carve::geom::vector<3>> outerShapePoints;
+				std::vector<carve::geom::vector<3>> innerShapePoints;
 
-				const int nvc = GeomSettings()->getNumberOfVerticesForTessellation(radius);
-				double delta_angle = GeomSettings()->getAngleLength(radius);
+				const int numberOfVertices = GeomSettings()->getNumberOfVerticesForTessellation(radius);
+				double deltaAngle = GeomSettings()->getAngleLength(radius);
 
-				std::vector<carve::geom::vector<3> > circle_points;
-				std::vector<carve::geom::vector<3> > circle_points_inner;
+				std::vector<carve::geom::vector<3>> circlePoints;
+				std::vector<carve::geom::vector<3>> circlePointsInner;
 				int i; double angle;
 				for (i = 0, angle = 0.;
-					i < nvc;
-					++i, angle += delta_angle)
+					i < numberOfVertices;
+					++i, angle += deltaAngle)
 				{
 					// cross section (circle) is defined in YZ plane
 					double x = sin(angle);
 					double y = cos(angle);
-					circle_points.push_back(carve::geom::VECTOR(0.0, x*radius, y*radius));
-					circle_points_inner.push_back(carve::geom::VECTOR(0.0, x*radius_inner, y*radius_inner));
+					circlePoints.push_back(carve::geom::VECTOR(0.0, x*radius, y*radius));
+					circlePointsInner.push_back(carve::geom::VECTOR(0.0, x*radiusInner, y*radiusInner));
 				}
 
-				int num_base_points = basis_curve_points.size();
-				carve::math::Matrix matrix_sweep;
+				int numBasePoints = basisCurvePoints.size();
+				carve::math::Matrix matrixSweep;
 
 				carve::geom::vector<3> local_z(carve::geom::VECTOR(0, 0, 1));
 
-				if (num_base_points < 2)
+				if (basisCurvePoints.size() < 2)
 				{
-					std::cout << "IfcSweptDiskSolid: num curve points < 2" << std::endl;
-					return;
+					throw oip::InconsistentModellingException("IfcSweptDiskSolid: num curve points < 2");
 				}
 
-				bool bend_found = false;
-				if (num_base_points > 3)
+				bool bendFound = false;
+				if (basisCurvePoints.size() > 3)
 				{
 					// compute local z vector by dot product of the first bend of the reference line
-					carve::geom::vector<3> vertex_back2 = basis_curve_points.at(0);
[…]
