+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -405,10 +405,10 @@ namespace OpenInfraPlatform
 
 				std::shared_ptr<carve::input::PolyhedronData> body_data = std::make_shared<carve::input::PolyhedronData>();
 				itemData->closed_polyhedrons.push_back(body_data);
-				//std::vector<carve::geom::vector<3> > inner_shape_points;  //TO DO: find out if i need inner_shape_points for the CrossSections
+				//std::vector<carve::geom::vector<3> > innerShapePoints;  //TO DO: find out if i need innerShapePoints for the CrossSections
 
 				int num_curve_points = BasisCurvePoints.size();
-				carve::math::Matrix matrix_sweep;
+				carve::math::Matrix matrixSweep;
 
 				// Less than two points is a point
 				if (num_curve_points < 2)
@@ -422,14 +422,12 @@ namespace OpenInfraPlatform
 				for (int i = 0; i < vec_cross_sections.size(); ++i)
 				{
 					std::shared_ptr<ProfileConverterT<IfcEntityTypesT>> profileConverter = profileCache->getProfileConverter(vec_cross_sections[i]);
-					const std::vector<std::vector<carve::geom::vector<2> > >& profileCoords = profileConverter->getCoordinates();
+					const std::vector<std::vector<carve::geom::vector<2>>>& profileCoords = profileConverter->getCoordinates();
 
 					// Save profile coords in paths
 					std::vector<std::vector<carve::geom::vector<2> > > profileCoords2D;
-					for (int p = 0; p < profileCoords.size(); ++p)
+					for (const auto& profileLoop : profileCoords)
 					{
-						const std::vector<carve::geom::vector<2> >& profileLoop = profileCoords[p];
-								
 						profileCoords2D.push_back(profileLoop);
 					}
 					paths.push_back(profileCoords2D);
@@ -547,7 +545,7 @@ namespace OpenInfraPlatform
 			    while (j < CrossSectionPoints.size()) 
 				{
 
-			      	//if basis_curve_points[i]==pointsOnCurve[j] ->save the information of pointsOnCurve
+			      	//if basisCurvePoints[i]==pointsOnCurve[j] ->save the information of pointsOnCurve
 					if (BasisCurvePoints[i] == CrossSectionPoints[j])
 					{
 						//1. save the information of pointsOnCurve in the new vector
@@ -578,14 +576,14 @@ namespace OpenInfraPlatform
 						++j;
 					}
 
-					//if basis_curve_points[i]!=pointsOnCurve[j] alternate depending on distance
+					//if basisCurvePoints[i]!=pointsOnCurve[j] alternate depending on distance
 					else if (BasisCurvePoints[i] != CrossSectionPoints[j])
 					{
 						// get the distance of the points 
 						double distCrossSectionPositions;
 						double distBasisCurvePoints;
 									
-						//calculate the distance from the point in basis_curve_points to the last element in the joint list
+						//calculate the distance from the point in basisCurvePoints to the last element in the joint list
 						int last = points_for_tessellation.size() - 1;
 
 						distCrossSectionPositions = distance(points_for_tessellation[last], CrossSectionPoints[j]);
@@ -855,7 +853,8 @@ namespace OpenInfraPlatform
 				throw oip::UnhandledException(sweptAreaSolid);
 			}
 
[…]
