### 2021-04-11 691e964
Refactoring convertIfcBooleanResult and convertIfcBooleanOperand
