### 2021-04-11 49fa394
Refactoring convertIfcCsgPrimitive3D
