+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -1696,275 +1696,381 @@ namespace OpenInfraPlatform
 				}
 			}
 
+			/*! \brief converts \c IfcCsgPrimitive3D to meshes.
+			 *
+			 * \param[in] csgPrimitive				The \c IfcCsgPrimitive3D to be converted.
+			 * \param[in] pos						The relative location of the origin of the representation's coordinate system within the geometric context.
+			 * \param[out] itemData					A pointer to be filled with the relevant data.
+			 *
+			 * \note See https://standards.buildingsmart.org/IFC/DEV/IFC4_2/FINAL/HTML/schema/ifcgeometricmodelresource/lexical/ifccsgprimitive3d.htm
+			*/
 			void convertIfcCsgPrimitive3D(
-				const std::shared_ptr<typename IfcEntityTypesT::IfcCsgPrimitive3D>& csgPrimitive,
+				const EXPRESSReference<typename IfcEntityTypesT::IfcCsgPrimitive3D>& csgPrimitive,
 				const carve::math::Matrix& pos,
 				std::shared_ptr<ItemData> itemData
 			) const noexcept(false)
 			{
-				std::shared_ptr<carve::input::PolyhedronData> polyhedron_data(new carve::input::PolyhedronData());
-				double length_factor = UnitConvert()->getLengthInMeterFactor();
+				// **************************************************************************************************************************
+				//	ENTITY IfcCsgPrimitive3D
+				//		ABSTRACT SUPERTYPE OF(ONEOF(
+				//			IfcBlock, 
+				//			IfcRectangularPyramid, 
+				//			IfcRightCircularCone, 
+				//			IfcRightCircularCylinder, 
+				//			IfcSphere))
+				//		SUBTYPE OF(IfcGeometricRepresentationItem);
+				//		Position: IfcAxis2Placement3D;
+				//		DERIVE
+				//			Dim : IfcDimensionCount: = 3;
+				//	END_ENTITY;
+				// **************************************************************************************************************************
 
-				// ENTITY IfcCsgPrimitive3D  ABSTRACT SUPERTYPE OF(ONEOF(IfcBlock, IfcRectangularPyramid, IfcRightCircularCone, IfcRightCircularCylinder, IfcSphere)
-
-				carve::math::Matrix primitive_placement_matrix = csgPrimitive->Position ?
-					pos * placementConverter->convertIfcAxis2Placement3D(csgPrimitive->Position) :
-					pos;
-
-
-				std::shared_ptr<typename IfcEntityTypesT::IfcBlock> block =
-					std::dynamic_pointer_cast<typename IfcEntityTypesT::IfcBlock>(csgPrimitive);
-				if (block)
+				if (csgPrimitive.template isOfType<typename IfcEntityTypesT::IfcBlock>())
 				{
-					double x_length = UnitConvert()->getLengthInMeterFactor();
-					double y_length = UnitConvert()->getLengthInMeterFactor();
-					double z_length = UnitConvert()->getLengthInMeterFactor();
-
-					if ((typename IfcEntityTypesT::IfcLengthMeasure)block->XLength)
-					{
-						x_length = (typename IfcEntityTypesT::IfcLengthMeasure)(block->XLength)*UnitConvert()->getLengthInMeterFactor();
-					}
-					if ((typename IfcEntityTypesT::IfcLengthMeasure)block->YLength)
-					{
-						y_length = (typename IfcEntityTypesT::IfcLengthMeasure)(block->YLength)*UnitConvert()->getLengthInMeterFactor();
-					}
-					if ((typename IfcEntityTypesT::IfcLengthMeasure)block->ZLength)
-					{
[…]
