### 2021-04-11 0cba750
Removed convertIfcBooleanClippingResult function, since in it alredy supported with IfcBooleanResult
