### 2021-04-12 207de31
Corrected mistake in UnhandledException and corrected code as war requested in PR
