### 2021-04-12 1f9d40c
Refactoring convertIfcHalfSpaceSolid
