+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -2058,26 +2058,185 @@ namespace OpenInfraPlatform
 			}
 
 			void convertIfcHalfSpaceSolid(
-				const std::shared_ptr<typename IfcEntityTypesT::IfcHalfSpaceSolid>& half_space_solid,
+				const EXPRESSReference<typename IfcEntityTypesT::IfcHalfSpaceSolid>& halfSpaceSolid,
 				const carve::math::Matrix& pos,
 				std::shared_ptr<ItemData> itemData,
 				const std::shared_ptr<ItemData>& otherOperand
 			) const noexcept(false)
-			{					
+			{
+				// TODO: complete this function + refactoring 
 				/*
 				//ENTITY IfcHalfSpaceSolid SUPERTYPE OF(ONEOF(IfcBoxedHalfSpace, IfcPolygonalBoundedHalfSpace))
-				std::shared_ptr<typename IfcEntityTypesT::IfcSurface> base_surface = half_space_solid->BaseSurface.lock();
-
-				double length_factor = UnitConvert()->getLengthInMeterFactor();
+				EXPRESSReference<typename IfcEntityTypesT::IfcSurface> base_surface = halfSpaceSolid->BaseSurface;
 
 				// base surface
-				std::shared_ptr<typename IfcEntityTypesT::IfcElementarySurface> elem_base_surface =
-					std::dynamic_pointer_cast<typename IfcEntityTypesT::IfcElementarySurface>(base_surface);
-				if (!elem_base_surface)
+				if (base_surface.template isOfType<typename IfcEntityTypesT::IfcElementarySurface>())
 				{
-					BLUE_LOG(warning) << "The base surface shall be an unbounded surface (subtype of IfcElementarySurface)";
-					return;
+					EXPRESSReference<typename IfcEntityTypesT::IfcElementarySurface> elem_base_surface =
+						base_surface.template as <typename IfcEntityTypesT::IfcElementarySurface>();
+
+					oip::EXPRESSReference<typename IfcEntityTypesT::IfcAxis2Placement3D>& base_surface_pos = elem_base_surface->Position;
+					carve::geom::plane<3> base_surface_plane;
+					carve::geom::vector<3> base_surface_position;
+					carve::math::Matrix base_position_matrix(carve::math::Matrix::IDENT());
+					if (base_surface_pos)
+					{
+						placementConverter->getPlane(base_surface_pos.lock(), base_surface_plane, base_surface_position, UnitConvert()->getLengthInMeterFactor());
+						base_position_matrix = placementConverter->convertIfcAxis2Placement3D(base_surface_pos);
+					}
+
+					// If the agreement flag is TRUE, then the subset is the one the normal points away from
+					bool agreement = halfSpaceSolid->AgreementFlag;
+					if (!agreement)
+					{
+						base_surface_plane.negate();
+					}
+
+					if (halfSpaceSolid.template isOfType<typename IfcEntityTypesT::IfcBoxedHalfSpace>())
+					{
+						convertIfcBoxedHalfSpace(
+							halfSpaceSolid.template as<typename IfcEntityTypesT::IfcBoxedHalfSpace>(),
+							pos, itemData);
+						return;
+					}
+
+					// check dimenstions of other operand
+					double extrusion_depth = HALF_SPACE_BOX_SIZE;
+					carve::geom::vector<3> other_operand_pos = base_surface_position;
+					if (otherOperand)
+					{
+						carve::geom::aabb<3> aabb;
+						otherOperand->createMeshSetsFromClosedPolyhedrons();
+						for (int ii = 0; ii < otherOperand->meshsets.size(); ++ii)
+						{
[…]
