### 2021-04-12 fef7225
Corrected mistakes in applyPositionToVertex
