### 2021-04-12 908e717
Started to refactor sectionedSolidHorizontal
