+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -351,29 +351,29 @@ namespace OpenInfraPlatform
 			}
 			
 			void convertIfcSectionedSolidHorizontal(
-				const EXPRESSReference<typename IfcEntityTypesT::IfcSectionedSolidHorizontal>& sectionedSolid_horizontal,
+				const EXPRESSReference<typename IfcEntityTypesT::IfcSectionedSolidHorizontal>& sectionedSolidHorizontal,
 				const carve::math::Matrix& pos,
 				std::shared_ptr<ItemData> itemData
 			) const noexcept(false)
 			{
-				if (sectionedSolid_horizontal.expired())
-					throw oip::ReferenceExpiredException(sectionedSolid_horizontal);
+				if (sectionedSolidHorizontal.expired())
+					throw oip::ReferenceExpiredException(sectionedSolidHorizontal);
 
 				//Get directrix and cross sections (attributes 1-2).
 				const EXPRESSReference<typename IfcEntityTypesT::IfcCurve>& directrix =
-					sectionedSolid_horizontal->Directrix;
-				std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcProfileDef>> vec_cross_sections =
-					sectionedSolid_horizontal->CrossSections;
+					sectionedSolidHorizontal->Directrix;
+				std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcProfileDef>> vecCrossSections =
+					sectionedSolidHorizontal->CrossSections;
 
 				// Get cross section positions and fixed axis vertical (attributes 3-4).
-				const auto& cross_section_positions = sectionedSolid_horizontal->CrossSectionPositions;
+				const auto& crossSectionPositions = sectionedSolidHorizontal->CrossSectionPositions;
 
-				bool fixed_axis_vertical = sectionedSolid_horizontal->FixedAxisVertical;
+				bool fixedAxisVertical = sectionedSolidHorizontal->FixedAxisVertical;
 
 				//check dimensions and correct attributes sizes
-				if (vec_cross_sections.size() != cross_section_positions.size())
+				if (vecCrossSections.size() != crossSectionPositions.size())
 				{
-					throw oip::InconsistentModellingException(sectionedSolid_horizontal, "CrossSections and CrossSectionsPositions are not equal in size.");
+					throw oip::InconsistentModellingException(sectionedSolidHorizontal, "CrossSections and CrossSectionsPositions are not equal in size.");
 				}
 
 				//Give directrix to Curve converter: for each station 1 Point and 1 Direction
@@ -403,29 +403,29 @@ namespace OpenInfraPlatform
 				}
 
 
-				std::shared_ptr<carve::input::PolyhedronData> body_data = std::make_shared<carve::input::PolyhedronData>();
-				itemData->closed_polyhedrons.push_back(body_data);
+				std::shared_ptr<carve::input::PolyhedronData> bodyData = std::make_shared<carve::input::PolyhedronData>();
+				itemData->closed_polyhedrons.push_back(bodyData);
 				//std::vector<carve::geom::vector<3> > innerShapePoints;  //TO DO: find out if i need innerShapePoints for the CrossSections
 
-				int num_curve_points = BasisCurvePoints.size();
+				int numCurvePoints = BasisCurvePoints.size();
 				carve::math::Matrix matrixSweep;
 
 				// Less than two points is a point
-				if (num_curve_points < 2)
+				if (numCurvePoints < 2)
 				{
-					throw oip::InconsistentModellingException(sectionedSolid_horizontal, " num curve points < 2");
[…]
