### 2021-04-14 107f220
Revert "Refactoring convertIfcRevolvedAreaSolid + applying some changes as was required in PR comments"
This reverts commit 5f97f15db1787497e720d524e0fa9ba8c4a59f2d.
