+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -430,7 +430,7 @@ namespace OpenInfraPlatform
 					{
 						profileCoords2D.push_back(profileLoop);
 					}
-					paths.push_back(profileCoords2D);
+					paths.push_back(profile_coords_2d);
 
 				}
 
@@ -953,8 +953,25 @@ namespace OpenInfraPlatform
 				}
 
 				// direction and length of extrusion
-				const double depth = extrudedArea->Depth * this->UnitConvert()->getLengthInMeterFactor();
-				carve::geom::vector<3> extrusionVector = placementConverter->convertIfcDirection(extrudedArea->ExtrudedDirection);
+				const double depth = (typename IfcEntityTypesT::IfcLengthMeasure)(extrudedArea->Depth) * UnitConvert()->getLengthInMeterFactor();
+				carve::geom::vector<3>  extrusionVector = carve::geom::VECTOR(0.0, 0.0, 0.0);
+				auto& vecDirection = extrudedArea->ExtrudedDirection->DirectionRatios;
+				
+				switch (vecDirection.size()) 
+				{
+				case 3:
+				{
+					extrusionVector.z = vecDirection[2] * depth;
+				}
+				case 2:
+				{
+					extrusionVector.x = vecDirection[0] * depth;
+					extrusionVector.y = vecDirection[1] * depth;
+					break;
+				}
+				default:
+					throw oip::InconsistentModellingException(extrudedArea, "Cardinality of the direction ratios can be only 2 or 3");
+				}
 				
 				// swept area
 				oip::EXPRESSReference<typename IfcEntityTypesT::IfcProfileDef> sweptArea = extrudedArea->SweptArea;
@@ -967,7 +984,7 @@ namespace OpenInfraPlatform
 
 				if (paths.size() == 0)
 				{
-					throw oip::InconsistentGeometryException(extrudedArea, "Paths are empty!");
+					return;
 				}
 
 				// .AREA. vs .CURVE. (is the result closed or open, i.e. full solid vs pipe)
@@ -1013,196 +1030,237 @@ namespace OpenInfraPlatform
 				std::shared_ptr<ItemData> itemData
 			) const noexcept(false)
 			{
-				// **************************************************************************************************************************
-				//	ENTITY IfcRevolvedAreaSolid
-				//		SUPERTYPE OF(IfcRevolvedAreaSolidTapered)
-				//		SUBTYPE OF(IfcSweptAreaSolid);
-				//			Axis: IfcAxis1Placement;
-				//			Angle: IfcPlaneAngleMeasure;
-				//		DERIVE
-				//			AxisLine : IfcLine: = IfcRepresentationItem() || IfcGeometricRepresentationItem() || IfcCurve() || IfcLine(Axis.Location, IfcRepresentationItem() || IfcGeometricRepresentationItem() || IfcVector(Axis.Z, 1.0));
-				//		WHERE
-				//			AxisStartInXY : Axis.Location.Coordinates[3] = 0.0;
-				//			AxisDirectionInXY: Axis.Z.DirectionRatios[3] = 0.0;
-				//	END_ENTITY;
-				// **************************************************************************************************************************
-
-				if (revolvedArea.expired())
-					throw oip::ReferenceExpiredException(revolvedArea);
-
+#ifdef _DEBUG
+				BLUE_LOG(trace) << "Processing IfcRevolvedAreaSolid #" << revolvedArea->getId();
+#endif
 				// Get axis and angle (attributes 3-4). 
-				oip::EXPRESSReference<typename IfcEntityTypesT::IfcAxis1Placement> axis = revolvedArea->Axis;
+				std::shared_ptr<typename IfcEntityTypesT::IfcAxis1Placement> axis =
+					revolvedArea->Axis.lock();
+				
+
[…]
