### 2021-04-15 773bafe
Adding first part of the tree model and tree item (not working yet)
