### 2021-04-17 392efce
Refactoring cpnvertIfcExtrudedAreaSolid one more time
