### 2021-04-17 57e281f
Refactoring convertIfcRevolvedAreaSolid once again
