+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -1013,238 +1013,202 @@ namespace OpenInfraPlatform
 				std::shared_ptr<ItemData> itemData
 			) const noexcept(false)
 			{
-#ifdef _DEBUG
-				BLUE_LOG(trace) << "Processing IfcRevolvedAreaSolid #" << revolvedArea->getId();
-#endif
-				// Get axis and angle (attributes 3-4). 
-				std::shared_ptr<typename IfcEntityTypesT::IfcAxis1Placement> axis =
-					revolvedArea->Axis.lock();
-				
+				// **************************************************************************************************************************
+				//	ENTITY IfcRevolvedAreaSolid
+				//		SUPERTYPE OF(IfcRevolvedAreaSolidTapered)
+				//		SUBTYPE OF(IfcSweptAreaSolid);
+				//			Axis: IfcAxis1Placement;
+				//			Angle: IfcPlaneAngleMeasure;
+				//		DERIVE
+				//			AxisLine : IfcLine: = IfcRepresentationItem() || IfcGeometricRepresentationItem() || IfcCurve() || IfcLine(Axis.Location, IfcRepresentationItem() || IfcGeometricRepresentationItem() || IfcVector(Axis.Z, 1.0));
+				//		WHERE
+				//			AxisStartInXY : Axis.Location.Coordinates[3] = 0.0;
+				//			AxisDirectionInXY: Axis.Z.DirectionRatios[3] = 0.0;
+				//	END_ENTITY;
+				// **************************************************************************************************************************
 
-				if (!revolvedArea->SweptArea)
-				{
-					return;
-				}
-				double length_factor = UnitConvert()->getLengthInMeterFactor();
+				if (revolvedArea.expired())
+					throw oip::ReferenceExpiredException(revolvedArea);
+
+				// Get axis and angle (attributes 3-4). 
+				oip::EXPRESSReference<typename IfcEntityTypesT::IfcAxis1Placement> axis = revolvedArea->Axis;
 
 				// angle and axis
-				double angle_factor = UnitConvert()->getAngleInRadianFactor();
-				oip::EXPRESSReference<typename IfcEntityTypesT::IfcProfileDef> sweptArea_profile = revolvedArea->SweptArea;
-				double revolution_angle = revolvedArea->Angle * angle_factor;
+			
+				oip::EXPRESSReference<typename IfcEntityTypesT::IfcProfileDef> sweptAreaProfile = revolvedArea->SweptArea;
+				double revolutionAngle = revolvedArea->Angle * this->UnitConvert()->getAngleInRadianFactor();
+			
+				carve::geom::vector<3>  axisLocation;
+				carve::geom::vector<3>  axisDirection;
 
-				carve::geom::vector<3>  axis_location;
-				carve::geom::vector<3>  axis_direction;
 				if (revolvedArea->Axis)
 				{
-					std::shared_ptr<typename IfcEntityTypesT::IfcAxis1Placement> axis_placement = revolvedArea->Axis.lock();
+					oip::EXPRESSReference<typename IfcEntityTypesT::IfcAxis1Placement> axisPlacement = revolvedArea->Axis;
 
-					if (axis_placement->Location)
-					{
-						axis_location = placementConverter->convertIfcCartesianPoint(axis_placement->Location );
-					}
 
-					if (axis_placement->Axis)
-					{
-						decltype(axis_placement->Axis)::type axis = axis_placement->Axis;
-						axis_direction = carve::geom::VECTOR(
-							axis->DirectionRatios[0],
-							axis->DirectionRatios[1],
-							axis->DirectionRatios[2]);
-					}
+					if (axisPlacement->Location)
+						axisLocation = placementConverter->convertIfcPoint(axisPlacement->Location);
[…]
