### 2021-04-17 513629d
Corrected small mistake in convertIfcRevovedAreaSolid
