### 2021-04-17 db04c97
Removing else as was requested in PR
