+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -2075,160 +2075,157 @@ namespace OpenInfraPlatform
 				EXPRESSReference<typename IfcEntityTypesT::IfcSurface> base_surface = halfSpaceSolid->BaseSurface;
 
 				// base surface
-				if (base_surface.template isOfType<typename IfcEntityTypesT::IfcElementarySurface>())
-				{
-					EXPRESSReference<typename IfcEntityTypesT::IfcElementarySurface> elem_base_surface =
-						base_surface.template as <typename IfcEntityTypesT::IfcElementarySurface>();
-
-					oip::EXPRESSReference<typename IfcEntityTypesT::IfcAxis2Placement3D>& base_surface_pos = elem_base_surface->Position;
-					carve::geom::plane<3> base_surface_plane;
-					carve::geom::vector<3> base_surface_position;
-					carve::math::Matrix base_position_matrix(carve::math::Matrix::IDENT());
-					if (base_surface_pos)
-					{
-						placementConverter->getPlane(base_surface_pos.lock(), base_surface_plane, base_surface_position, UnitConvert()->getLengthInMeterFactor());
-						base_position_matrix = placementConverter->convertIfcAxis2Placement3D(base_surface_pos);
-					}
-
-					// If the agreement flag is TRUE, then the subset is the one the normal points away from
-					bool agreement = halfSpaceSolid->AgreementFlag;
-					if (!agreement)
-					{
-						base_surface_plane.negate();
-					}
-
-					if (halfSpaceSolid.template isOfType<typename IfcEntityTypesT::IfcBoxedHalfSpace>())
-					{
-						convertIfcBoxedHalfSpace(
-							halfSpaceSolid.template as<typename IfcEntityTypesT::IfcBoxedHalfSpace>(),
-							pos, itemData);
-						return;
-					}
-
-					// check dimenstions of other operand
-					double extrusion_depth = HALF_SPACE_BOX_SIZE;
-					carve::geom::vector<3> other_operand_pos = base_surface_position;
-					if (otherOperand)
-					{
-						carve::geom::aabb<3> aabb;
-						otherOperand->createMeshSetsFromClosedPolyhedrons();
-						for (int ii = 0; ii < otherOperand->meshsets.size(); ++ii)
-						{
-							std::shared_ptr<carve::mesh::MeshSet<3>>& meshset = otherOperand->meshsets[ii];
-
-							if (!meshset)
-							{
-								throw oip::InconsistentGeometryException("Meshset not given!");
-							}
-
-							if (ii == 0)
-							{
-								aabb = meshset->getAABB();
-							}
-							else
-							{
-								aabb.unionAABB(meshset->getAABB());
-							}
-						}
-						aabb.max();
-						carve::geom::vector<3>& aabb_extent = aabb.extent;
-						double max_extent = std::max(aabb_extent.x, std::max(aabb_extent.y, aabb_extent.z));
-						extrusion_depth = 2.0*max_extent;
-						other_operand_pos = aabb.pos;
-					}
-
-
-					if (halfSpaceSolid.template isOfType<typename IfcEntityTypesT::IfcPolygonalBoundedHalfSpace>())
-					{
-						EXPRESSReference<typename IfcEntityTypesT::IfcPolygonalBoundedHalfSpace> polygonal_half_space =
-							halfSpaceSolid.template as<typename IfcEntityTypesT::IfcPolygonalBoundedHalfSpace>();
-						// ENTITY IfcPolygonalBoundedHalfSpace
-						//	SUBTYPE OF IfcHalfSpaceSolid;
-						//	Position	 :	IfcAxis2Placement3D;
[…]
