### 2021-04-17 1f251ea
Corrected mistake in convertIfcHalfSpaceSolid
