+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -2225,7 +2225,7 @@ namespace OpenInfraPlatform
 					}
 
 					return;
-				}
+				}*/
 			}
 
 			// Complete this function + refactoring 
