### 2021-04-17 9820c5f
Added ReferenceExpiredExceptions to functions
