+++ UnitTests/Schemas/IFC4X1/extruded-solid/src/extruded-solid.cpp
@@ -38,10 +38,10 @@ class ExtrudedSolidTest : public IfcVisualTest {
 		express_model = OpenInfraPlatform::IFC4X1::IFC4X1Reader::FromFile(filename.string());
 
 		importer = buw::makeReferenceCounted<oip::IfcImporterT<emt::IFC4X1EntityTypes>>();
-		model = importer->collectData(express_model);
+		auto model = importer->collectData(express_model);
 		
 		_background = rendererIfc->captureImage();
-		rendererIfc->setModel(model);
+		rendererIfc->setModel(model[0]);
 	}
 	
 	virtual void TearDown() override {
@@ -56,8 +56,6 @@ class ExtrudedSolidTest : public IfcVisualTest {
 
 	std::shared_ptr<oip::EXPRESSModel> express_model = nullptr;
 	buw::ReferenceCounted<oip::IfcImporterT<emt::IFC4X1EntityTypes>> importer = nullptr;
-	buw::ReferenceCounted<oip::IfcModel> model = buw::makeReferenceCounted<oip::IfcModel>();
-
 };
 
 TEST_F(ExtrudedSolidTest, AllEntitiesAreRead) {
