### 2021-04-20 49fe485
Small changes maily to treeModel and treeItem
