### 2021-04-20 4383357
Adding a ExpressModel to the ifcModel in order to be able to access it in the tree viewer dialog
