+++ UserInterface/Dialogues/IfcTreeDialog.cpp
@@ -18,12 +18,16 @@
 #include "IfcTreeDialog.h"
 #include "IfcTreeModel.h"
 #include "ui_IfcTreeDialog.h"
+#include "DataManagement/General/Data.h"
 
-OpenInfraPlatform::UserInterface::IfcTreeDialog::IfcTreeDialog(QWidget *parent /*= nullptr*/) :
+OpenInfraPlatform::UserInterface::IfcTreeDialog::IfcTreeDialog(OpenInfraPlatform::UserInterface::View* view, QWidget *parent /*= nullptr*/) :
 	QDialog(parent, Qt::WindowTitleHint | Qt::WindowCloseButtonHint),
-	ui_(new Ui::IfcTreeDialog)
+	ui_(new Ui::IfcTreeDialog),
+	view_(view)
 {
 	ui_->setupUi(this);
+	//ui_->ifcTreeWidget->setAnimated(true);
+	//ui_->ifcTreeView->setModel(new IfcTreeModel());
 }
 
 OpenInfraPlatform::UserInterface::IfcTreeDialog::~IfcTreeDialog()
@@ -31,11 +35,17 @@ OpenInfraPlatform::UserInterface::IfcTreeDialog::~IfcTreeDialog()
 
 }
 
-void OpenInfraPlatform::UserInterface::IfcTreeDialog::setModel()
+void OpenInfraPlatform::UserInterface::IfcTreeDialog::show()
 {
+	//for first testing just get last model, later this will be changed to selecting the model need
+	auto model = Core::DataManagement::DocumentManager::getInstance().getData().getLastModel();
+	auto ifcModel = std::static_pointer_cast<OpenInfraPlatform::Core::IfcGeometryConverter::IfcModel>(model);
+	auto entities = ifcModel->getExpressModel();
+	ui_->ifcTreeView->setAnimated(true);
 	//ui_->ifcTreeView->setModel(new IfcTreeModel());
-	//((QDialog*)this)->show();
+	((QDialog*)this)->show();
 }
+
 
 void OpenInfraPlatform::UserInterface::IfcTreeDialog::on_pushButtonClose_clicked()
 {
