### 2021-04-21 02418af
Add IfcTreeItem.cpp + some minor changes
