### 2021-04-22 2ad6aa0
Commenting out lines which are responsible for saving screenshots
