### 2021-04-22 4c21b65
check for an empty pointer
