### 2021-04-22 d638254
added a distinction for files without georeferencing
