### 2021-04-22 672b031
correcting the empty determination
