### 2021-04-22 e17070d
check if context exists
