+++ Core/src/IfcGeometryConverter/RepresentationConverter.h
@@ -94,7 +94,17 @@ namespace OpenInfraPlatform {
 				{
 				}
 
-				/*! Get the blobal placement of the context
+
+				/*! Get the default global placement of the context
+				*
+				* \return The placement matrix - all coordinates should be left-multiplied.
+				*/
+				const carve::math::Matrix getDefaultPlacement() const noexcept(false)
+				{
+					return georeferencingConverter_->getDefaultPlacement();
+				}
+
+				/*! Get the global placement of the context
 				* 
 				* \param[in] context The \c IfcRepresentationContext context to resolve.
 				* \return The placement matrix - all coordinates should be left-multiplied.
