+++ Core/src/IfcGeometryConverter/IfcImporterImpl.h
@@ -133,6 +133,9 @@ std::vector<std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>>> IfcImporterT<IfcE
 			std::shared_ptr<ShapeInputDataT<IfcEntityTypesT>> productShape
 				= std::make_shared<ShapeInputDataT<IfcEntityTypesT>>(product);
 			std::shared_ptr<ItemData> itemData(new ItemData());
+			// get the default placement
+			const carve::math::Matrix contextPlacement =
+				repConverter->getDefaultPlacement();
 			// interpret the alignment
 			repConverter->convertIfcGeometricRepresentationItem(
 				axis,
