### 2021-04-22 2df95a8
Get position and parameters of clothoid
