### 2021-04-22 18434c5
Get position and parameters of clothoid
