### 2021-04-23 ebc2872
corrected mistake in CMake
