### 2021-04-24 fc1179b
some functions need adjustment since the entities got changed
