+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -149,13 +149,13 @@ namespace OpenInfraPlatform {
 					//	END_ENTITY;
 					// **************************************************************************************************************************
 					// IfcBlossCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
-#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
-					if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcBlossCurve>())
-					{
-						return convertIfcBlossCurve(ifcCurve.as<typename IfcEntityTypesT::IfcBlossCurve>(),
-							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
-					}
-#endif
+//#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
+//					if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcBlossCurve>())
+//					{
+//						return convertIfcBlossCurve(ifcCurve.as<typename IfcEntityTypesT::IfcBlossCurve>(),
+//							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
+//					}
+//#endif
 
 					// IfcBoundedCurve SUBTYPE of IfcCurve
 					if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcBoundedCurve>())
@@ -201,14 +201,14 @@ namespace OpenInfraPlatform {
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					}
 
-					// IfcSeriesParameterCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
-#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
-					else if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcSeriesParameterCurve>())
-					{
-						return convertIfcSeriesParameterCurve(ifcCurve.as<typename IfcEntityTypesT::IfcSeriesParameterCurve>(),
-							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
-					}
-#endif
+//					// IfcSeriesParameterCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
+//#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
+//					else if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcSeriesParameterCurve>())
+//					{
+//						return convertIfcSeriesParameterCurve(ifcCurve.as<typename IfcEntityTypesT::IfcSeriesParameterCurve>(),
+//							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
+//					}
+//#endif
 
 					// IfcSurfaceCurve SUPTYPE of IfcCurve
 					else if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcSurfaceCurve>())
@@ -279,31 +279,31 @@ namespace OpenInfraPlatform {
 				} //end convertIfcCurve2D
 
 				// IfcBlossCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
-#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
-				/**********************************************************************************************/
-				/*! \brief Converts an \c IfcBlossCurve to a tesselated curve.
-				* \param[in] blossCurve				A pointer to data from \c IfcBlossCurve.
-				* \param[out] targetVec				The tessellated line.
-				* \param[out] segmentStartPoints	The starting points of separate segments.
[…]
