### 2021-04-24 38e1fa7
linear placement got reworked (not finished, though)
