### 2021-04-24 731e93a
rename GeometryModelRenderer to OpenInfraPlatform.GeometryModelRenderer
