### 2021-04-24 79c472b
introduce ScreenShotMaker (empty shell)
