### 2021-04-25 2940396
linking in IFC projects as selected by the user in CMake
