### 2021-04-25 8852f96
renamed the app, made an input mandatory
