### 2021-04-25 d58a9f0
introduced a templated function for IFC versions
