### 2021-04-25 bec96bf
made logging more manageable
