### 2021-04-25 e0df51f
small tweaks to the program code
