### 2021-04-25 f632aa7
simplifying logging
