### 2021-04-25 2e31e77
making image files named as the input ifc file
