### 2021-04-25 32472e6
started a new function to save all views at once
