### 2021-04-25 6037b43
if exception is thrown, wait for user to read it
