### 2021-04-25 7d96917
checking supported IFC schemas
