### 2021-04-25 0905c88
simplifying the code
