### 2021-04-25 cc744ab
checking input a bit better
