### 2021-05-05 f3b75fc
Moved out-commented code block into actual position. No further changes were made inside the block. This is a help step to make further changes better visible. Project compiles because the function SplineConverterT::convertIfcBSplineSurface isn't called anywhere.
