+++ Core/src/IfcGeometryConverter/SplineConverter.h
@@ -151,8 +151,120 @@ namespace OpenInfraPlatform
 						const std::vector<std::vector<carve::geom::vector<3>>>& controlPoints,
 						std::shared_ptr<carve::input::PolylineSetData>& polylineData) const throw(...)
 					{
-						// not implemented, code in commented out part at the end of the file
-						throw oip::UnhandledException(splineSurface);
+						template<>
+						inline void SplineConverterT<emt::Ifc4EntityTypes, OpenInfraPlatform::Ifc4::UnitConverter>::convertIfcBSplineSurface(
+							const std::shared_ptr<emt::Ifc4EntityTypes::IfcBoundedSurface>& splineSurfaceWithKnots,
+							const std::vector<std::vector<carve::geom::vector<3>>>& controlPoints,
+							std::shared_ptr<carve::input::PolylineSetData>& polylineData)
+						{
+							std::shared_ptr<emt::Ifc4EntityTypes::IfcBSplineSurfaceWithKnots> bsplineSurfaceWithKnots =
+								std::dynamic_pointer_cast<emt::Ifc4EntityTypes::IfcBSplineSurfaceWithKnots>(splineSurfaceWithKnots);
+
+							if (splineSurfaceWithKnots)
+							{
+								// obtain degree of both b-spline curves
+								const int degreeU = bsplineSurfaceWithKnots->m_UDegree;
+								const int orderU = degreeU + 1;
+								const int degreeV = bsplineSurfaceWithKnots->m_VDegree;
+								const int orderV = degreeV + 1;
+								// obtain number of control points
+								const int numControlPointsU = controlPoints.size();
+								const int numControlPointsV = controlPoints[0].size();
+								// obtain number of knots
+								const int numKnotsU = orderU + numControlPointsU;
+								const int numKnotsV = orderV + numControlPointsV;
+
+								// obtain knots for each direction
+								const std::vector<int>& knotMultsU = bsplineSurfaceWithKnots->m_UMultiplicities;
+								const std::vector<std::shared_ptr<emt::Ifc4EntityTypes::IfcParameterValue>>& splineKnotsU = bsplineSurfaceWithKnots->m_UKnots;
+								const std::vector<int>& knotMultsV = bsplineSurfaceWithKnots->m_VMultiplicities;
+								const std::vector<std::shared_ptr<emt::Ifc4EntityTypes::IfcParameterValue>>& splineKnotsV = bsplineSurfaceWithKnots->m_VKnots;
+
+								std::vector<double> knotsU;
+								std::vector<double> knotsV;
+								knotsU.reserve(numKnotsU);
+								knotsV.reserve(numKnotsV);
+
+								for (int i = 0; i < splineKnotsU.size(); ++i)
+								{
+									double knot = splineKnotsU[i]->m_value;
+									const int knotMult = knotMultsU[i];
+									// look at the multiplicity of the current knot
+									for (int j = 0; j < knotMult; ++j)
+									{
+										knotsU.push_back(knot);
+									}
+								}
+
+								for (int i = 0; i < splineKnotsV.size(); ++i)
+								{
+									double knot = splineKnotsV[i]->m_value;
+									const int knotMult = knotMultsV[i];
+									// look at the multiplicity of the current knot
+									for (int j = 0; j < knotMult; ++j)
+									{
+										knotsV.push_back(knot);
+									}
+								}
+
+								std::vector<std::vector<double>> weights;
+
[…]
