### 2021-05-05 4c59b39
Added temporary call of SplineConverterT::convertIfcBSplineSurface ... and adjusted the function parameters for the appropriate call. Does not compile (desired to get compiler errors as feedback of coming refactoring).
