### 2021-05-06 524af53
Added IfsSpiral converter
