+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -869,8 +869,84 @@ namespace OpenInfraPlatform
 					GeomUtils::appendPointsToCurve(basisCurvePoints, targetVec);
 				}
 
+<<<<<<< HEAD
 				// IfcClothoid SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
+=======
+				// IfcSpline SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
+#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
+>>>>>>> 92a31da5 (Added IfsSpiral converter)
+				/**********************************************************************************************/
+				/*! \brief Converts an \c IfcSpiral to a tesselated curve.
+				* \param[in] spiral					A pointer to data from \c Spiral.
+				* \param[out] targetVec				The tessellated line.
+				* \param[out] segmentStartPoints	The starting points of separate segments.
+				* \param[in] trim1Vec				The trimming of the curve as saved in IFC model - trim at start of curve.
+				* \param[in] trim2Vec				The trimming of the curve as saved in IFC model - trim at end of curve.
+				* \param[in] senseAgreement			Does the resulting geometry have the same sense agreement as the \c IfcSpiral.
+				*/
+				void convertIfcSpiral(
+					const EXPRESSReference<typename IfcEntityTypesT::IfcSpiral>& spiral,
+					std::vector<carve::geom::vector<3>>& targetVec,
+					std::vector<carve::geom::vector<3>>& segmentStartPoints,
+					const std::vector<std::shared_ptr<typename IfcEntityTypesT::IfcTrimmingSelect>>& trim1Vec,
+					const std::vector<std::shared_ptr<typename IfcEntityTypesT::IfcTrimmingSelect>>& trim2Vec,
+					const bool senseAgreement,
+					const typename IfcEntityTypesT::IfcTrimmingPreference & trimmingPreference
+				) const throw(...)
+				{
+					// **************************************************************************************************************************
+					//	 https://
+					//	ENTITY IfcSpiral
+					//		ABSTRACT SUPERTYPE OF(ONEOF
+					//			(IfcBlossCurve,
+					//          IfcClothoid
+					//          IfcCosine
+					//          IfcHelmertCurve
+					//			IfcSine))
+					//		SUBTYPE OF(IfcCurve);
+					//		Position: OPTIONAL IfcAxis2Placement;
+					//	END_ENTITY;
+					// **************************************************************************************************************************
+					// IfcSpiral SUBTYPE OF IfcBlossCurve
+					if (spiral.isOfType<typename IfcEntityTypesT::IfcBlossCurve>())
+					{
+						return convertIfcSpiral(spiral.as<typename IfcEntityTypesT::IfcSpiral>(),
+							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
+					} // end if IfcBlossCurve
+
+					// IfcEllipse SUBTYPE OF IfcClothoid
+					else if (spiral.isOfType<typename IfcEntityTypesT::IfcClothoid>())
+					{
+						return convertIfcClothoid(conic.as<typename IfcEntityTypesT::IfcClothoid>(),
+							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
+					} // end if IfcClothoid
+
+					// IfcEllipse SUBTYPE OF IfcCosine
+					else if (spiral.isOfType<typename IfcEntityTypesT::IfcCosine>())
+					{
[…]
