### 2021-05-06 b2cbb1d
Added IfsSpiral converter
