### 2021-05-06 92a31da
Added IfsSpiral converter
