### 2021-05-06 52d3501
Deleted deprecated template-function-declaration; adjusted alignment
