+++ Core/src/IfcGeometryConverter/SplineConverter.h
@@ -152,117 +152,109 @@ namespace OpenInfraPlatform
 						//const std::vector<std::vector<carve::geom::vector<3>>>& controlPoints,
 						std::shared_ptr<carve::input::PolylineSetData>& polylineData) const throw(...)
 					{
-						
-						template<>
-						inline void SplineConverterT<emt::Ifc4EntityTypes, OpenInfraPlatform::Ifc4::UnitConverter>::convertIfcBSplineSurface(
-							const std::shared_ptr<emt::Ifc4EntityTypes::IfcBoundedSurface>& splineSurfaceWithKnots,
-							const std::vector<std::vector<carve::geom::vector<3>>>& controlPoints,
-							std::shared_ptr<carve::input::PolylineSetData>& polylineData)
+						std::shared_ptr<emt::Ifc4EntityTypes::IfcBSplineSurfaceWithKnots> bsplineSurfaceWithKnots =
+							std::dynamic_pointer_cast<emt::Ifc4EntityTypes::IfcBSplineSurfaceWithKnots>(splineSurfaceWithKnots);
+
+						if (splineSurfaceWithKnots)
 						{
-							std::shared_ptr<emt::Ifc4EntityTypes::IfcBSplineSurfaceWithKnots> bsplineSurfaceWithKnots =
-								std::dynamic_pointer_cast<emt::Ifc4EntityTypes::IfcBSplineSurfaceWithKnots>(splineSurfaceWithKnots);
+							// obtain degree of both b-spline curves
+							const int degreeU = bsplineSurfaceWithKnots->m_UDegree;
+							const int orderU = degreeU + 1;
+							const int degreeV = bsplineSurfaceWithKnots->m_VDegree;
+							const int orderV = degreeV + 1;
+							// obtain number of control points
+							const int numControlPointsU = controlPoints.size();
+							const int numControlPointsV = controlPoints[0].size();
+							// obtain number of knots
+							const int numKnotsU = orderU + numControlPointsU;
+							const int numKnotsV = orderV + numControlPointsV;
 
-							if (splineSurfaceWithKnots)
+							// obtain knots for each direction
+							const std::vector<int>& knotMultsU = bsplineSurfaceWithKnots->m_UMultiplicities;
+							const std::vector<std::shared_ptr<emt::Ifc4EntityTypes::IfcParameterValue>>& splineKnotsU = bsplineSurfaceWithKnots->m_UKnots;
+							const std::vector<int>& knotMultsV = bsplineSurfaceWithKnots->m_VMultiplicities;
+							const std::vector<std::shared_ptr<emt::Ifc4EntityTypes::IfcParameterValue>>& splineKnotsV = bsplineSurfaceWithKnots->m_VKnots;
+
+							std::vector<double> knotsU;
+							std::vector<double> knotsV;
+							knotsU.reserve(numKnotsU);
+							knotsV.reserve(numKnotsV);
+
+							for (int i = 0; i < splineKnotsU.size(); ++i)
 							{
-								// obtain degree of both b-spline curves
-								const int degreeU = bsplineSurfaceWithKnots->m_UDegree;
-								const int orderU = degreeU + 1;
-								const int degreeV = bsplineSurfaceWithKnots->m_VDegree;
-								const int orderV = degreeV + 1;
-								// obtain number of control points
-								const int numControlPointsU = controlPoints.size();
-								const int numControlPointsV = controlPoints[0].size();
-								// obtain number of knots
-								const int numKnotsU = orderU + numControlPointsU;
-								const int numKnotsV = orderV + numControlPointsV;
-
-								// obtain knots for each direction
-								const std::vector<int>& knotMultsU = bsplineSurfaceWithKnots->m_UMultiplicities;
[…]
