### 2021-05-06 41e11a9
Refactored conversion from splineSurface to splineSurfaceWithKnots
