### 2021-05-07 f499991
Refactored obtainment of 'degree' and 'control points'. Loading the control points was copied from FaceConterter where a few parts of out-commented code already are.
