### 2021-05-09 0aa0cfa
implemented transformation for Face with Voids, Moved part of the code in extrude to the new function
