### 2021-05-09 #3061 @abdoulayediak
Hi all! @hugoledoux here I am finally.
First of all, I'd like to apologise for being inactive all this time, as these 2 last months have been very very crazy busy.

@pjanck it is always a pleasure to see some open source contributions, mostly in the field of BIM where there is still a lot to be done. I have been through the repo and the check-list above, and here are my first comments (most of them were already raised by @aothms):

- Going through the paper (Statement of need section), the following sentence:
> However, they focus mainly on the stable versions of the IFC standard56(like IFC2x3 and IFC4), while OIP focuses primarily on the newer developments.

left me with the impression that OIP does not deal with stable versions of IFC. I had even more that impression while checking the tool itself, as the 'Open' and 'Import' options seem very specific in dealing with specific kind of IFC. I had to try randomly an IFC2x3 file to realise that (as expected), stable release are handled as well. Therefore, I think it would be beneficial to clearly state something here with respect to that. Because if the OIP can handle stable releases just like the other tools do, then it is good plus even. If not, then it should be clearly stated so that the users can limit their expectations.

- The self-citation mentioned by @aothms is not that much a problem as you explained. But I think the paper should clarify a bit more explicitly the information that can be found in referred papers without too much details (e.g. mention that details of the OIP components are provided in Hecht & Jaud, 2019).

- Another point is the Windows-only support. Although I can see that the "cross-platform" claim was removed from the paper, it would be definitely valuable to state it early somewhere in the ReadMe of the repo.

I still have few other points to check before completing the check-list, but it all sounds on a pretty good track to me.

[For the general discussion]
@aothms I found [this thread](https://github.com/ifcquery/ifcplusplus/issues/125) suggesting (among other stuffs that may be of particular interest to you) that OCC is only partially used by IFC++. The latter mostly relies on Carve for its geometric operations.
