### 2021-05-10 11ff3dd
moved some documentation to a new "retired" folder
