+++ Documentation/retired/bloss.py
@@ -1,0 +1,185 @@
+import scipy.integrate as integrate
+import matplotlib.pyplot as plt
+import math
+import numpy as np
+
+
+def posX(K, L, R, l):
+	return integrate.quad(lambda l_: math.cos(K(L, R, l_)), 0, l)[0]
+	
+def posY(K, L, R, l):
+	return integrate.quad(lambda l_: math.sin(K(L, R, l_)), 0, l)[0]
+		
+
+# clothoid
+# k = l / A2
+# K = l^2 / (A2 * 2) 
+
+def K_Clothoid(L, R, l):
+	A2 = R * L
+	return pow(l, 2) / (A2 * 2)
+	
+def k_Clothoid(L, R, l):
+	A2 = R * L
+	return l / A2
+
+# calculated with matlab
+# needed approxCurve.md
+# k = l /(R*L)
+# [x, y, X, Y] = approxCurve(k, 4)
+def X_Clothoid(L, R, l):
+	return l - l**5/(40*L**2*R**2) + l**9/(3456*L**4*R**4) - l**13/(599040*L**6*R**6)
+	
+def Y_Clothoid(L, R, l):
+	return l**3/(6*L*R) - l**7/(336*L**3*R**3) + l**11/(42240*L**5*R**5) - l**15/(9676800*L**7*R**7)
+	
+# bloss 
+# K = l^3 / (R * L^2) - l^4 * (2 * R * L^3) 
+# k = l^2 * 3 / (R * L^2) - l^3 * 2 / (R * L^3)
+def K_Bloss(L, R, l):
+	return pow(l, 3) / (R * pow(L, 2)) - pow(l,4) / (2 * R * pow(L, 3))
+	
+def k_Bloss(L, R, l):
+	return math.pow(l, 2) * 3 / (R * math.pow(L, 2)) - math.pow(l, 3) * 2 / (R * math.pow(L, 3))
+
+# calculated with matlab
+# needed approxCurve.md
+# k = 3 * l^2 /(R*L^2) - 2 * l^3 /(R*L^3)
+# [x, y, X, Y] = approxCurve(k, 4)
+def X_Bloss(L, R, l):
+	return l - l**7/(14*L**4*R**2) + l**8/(16*L**5*R**2) - l**9/(72*L**6*R**2) + l**13/(312*L**8*R**4) - l**14/(168*L**9*R**4) + l**15/(240*L**10*R**4) - l**16/(768*L**11*R**4) + l**17/(6528*L**12*R**4) - l**19/(13680*L**12*R**6) + l**20/(4800*L**13*R**6) - l**21/(4032*L**14*R**6) + l**22/(6336*L**15*R**6) - l**23/(17664*L**16*R**6) + l**24/(92160*L**17*R**6) - l**25/(1152000*L**18*R**6)
+	
+def Y_Bloss(L, R, l):
+	return l**4/(4*L**2*R) - l**5/(10*L**3*R) - l**10/(60*L**6*R**3) + l**11/(44*L**7*R**3) - l**12/(96*L**8*R**3) + l**13/(624*L**9*R**3) + l**16/(1920*L**10*R**5) - l**17/(816*L**11*R**5) + l**18/(864*L**12*R**5) - l**19/(1824*L**13*R**5) + l**20/(7680*L**14*R**5) - l**21/(80640*L**15*R**5) - l**22/(110880*L**14*R**7) + l**23/(33120*L**15*R**7) - l**24/(23040*L**16*R**7) + l**25/(28800*L**17*R**7) - l**26/(59904*L**18*R**7) + l**27/(207360*L**19*R**7) - l**28/(1290240*L**20*R**7) + l**29/(18708480*L**21*R**7)
+	
+def getTransitionCurve(K, L, R, n):
+	xs = []
[…]
