### 2021-05-10 afa214e
adding a pdf of the architecture to the repo, expanded readme
