### 2021-05-10 #3061 @pjanck
Hi @abdoulayediak - great to hear from you!

I'm happy that you found some time and first of all: thank you for your comments and your time.

> OIP does not deal with stable versions of IFC

This is a true statement for the `development` branch, which is some ~2 years ahead and currently our main branch (should be set like that in the repo's settings). Previous versiond (I believe even the last installer) could indeed read in stable IFC versions, but could not handle the "old" stuff within the "new" standard development (like an `IfcWall` in an IFC4x1 file). This is the reason for the major overhaul in the recent three years and our lack of focus for IFC2x3 and IFC4 in the current state. Please observe additionally:

- https://github.com/tumcms/Open-Infra-Platform/issues/334
- https://github.com/tumcms/Open-Infra-Platform/issues/335

>  it would be definitely valuable to state [Windows-only support] early somewhere in the ReadMe of the repo.

Added a note to the specialized "setup" readme: https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/SetupHelp.md

>  I think the paper should clarify a bit more explicitly the information that can be found in referred papers without too much details (e.g. mention that details of the OIP components are provided in Hecht & Jaud, 2019).

I believe that this is enough: "Nowadays, OIP uses the IFC schema and CloudCompare's model as independent internal data models -- explained in detail by @Hecht:2019:FBI." Do you think I should expand (I find the paper already longer as most papers submitted with JOSS).

> self-citation

I know that we're quite heavy on self-citations (there are more like a dozen and not only 2 as noted by @hugoledoux ). However, I believe that this is one of the principles of this journal - don't repeat yourself but rather point the readers in the right directions. And I am somehow proud that we've managed to use our software in this many publications :)

I've updated the paper slightly, so I'll be recompiling the pdf again after I press enter.
