### 2021-05-12 3341b51
Added a note to the documentation about the state of the installer
