### 2021-05-12 #3061 @CBenghi
Hi

A few issues to discuss to improve the paper:

1. Looking at the authors `Contribution and authorship` concerns I wonder if you have
properly considered the contribution of Cara Coetzee

```

Cara Coetzee                    45         75224          95039            4.76
ccoetzee                        65          8184           3132            0.32

Helge Hecht                    310        367863        1759623           59.47
Jonas Schlenger                 52          2319            881            0.09
Julian Amann                    62       1165069           3021           32.65
Stefan Jaud                    429         25919           9279            0.98

Author                     Rows      Stability          Age       % in comments

Cara Coetzee               7568           10.1         19.9               19.24
ccoetzee                   5412           66.1          5.6               59.15

Helge Hecht               73881           20.1         28.3               29.61
Stefan Jaud               23391           90.2          2.7               22.73
Jonas Schlenger            1849           79.7          0.9               27.91
Julian Amann             275581           23.7          0.3               18.39
```

2. On line 17 you mention `viewing and analysis`, but it's not clear what type
of analysis can be performed. I think a more specific set of functional claims would improve the paper.

3. Perhaps connected to the previous point, on line 57 you mention the specific focus on preliminary Express schemas.
This uniqueness could be better highlighted in the project documentation, providing instructions for
custom schema implementation.

4. Mentioning limitations of OIP in comparison to other frameworks could better serve the research community (e.g. what you mentioned about the single IFC version at a time).

5. Installation documentation could be slightly improved to clarify that *plenty* of warnings in cmake should be expected.

6. I've noticed a few apps in the solution that support your testing and QA. I would invite you to produce a quick readme markdown that might help developers understand their purpose and contribute cases.
