### 2021-05-12 tumcms/Open-Infra-Platform#454 @pjanck
[TODO] Improve for JOSS
**Additional context**
See https://github.com/openjournals/joss-reviews/issues/3061#issuecomment-839733632
