### 2021-05-14 75811a2
Fixed obtainment of control points. (personal) desired is the use of the function FaceConverterT::convertIfcCartesianPoint2DVector. To call this function, the data has to be transformed from 'controlPointList' to 'controlPointListVector.
