### 2021-05-14 41df2e1
Refactored FaceConverterT::convertIfcCartesianPoint2DVector. Previously, this function wasn't able to compile. point2D.expired() was moved and changed to be called on an EXPRESSReference, instead the std::vector. InconsistentGeometryException was changed accordingly.
