### 2021-05-15 71ec6b7
Created new function GeomUtils::correctWinding
