+++ Core/src/IfcGeometryConverter/GeomUtils.cpp
@@ -123,8 +123,7 @@ void GeomUtils::extrude(
 					const std::vector<std::vector<carve::geom::vector<2> > >& face_loops_input, 
 					const carve::geom::vector<3> extrusion_vector, 
 					const bool closed,
-					std::shared_ptr<carve::input::PolyhedronData>& poly_data, 
-					std::stringstream& err )
+					std::shared_ptr<carve::input::PolyhedronData>& poly_data)
 {
 	if( face_loops_input.size() == 0 )
 	{
@@ -150,60 +149,8 @@ void GeomUtils::extrude(
 	//  0-------face_loops[0]--------1
 
 	carve::geom::vector<3> normal_first_loop;
-	std::vector<std::vector<carve::geom2d::P2>>	face_loops;
-	for( std::vector<std::vector<carve::geom::vector<2>>>::const_iterator it_face_loops = face_loops_input.begin(); it_face_loops != face_loops_input.end(); ++it_face_loops )
-	{
-		const std::vector<carve::geom::vector<2>>& loop = (*it_face_loops);
-
-		if( loop.size() < 3 )
-		{
-			throw oip::InconsistentGeometryException("GeomUtils::extrude(): loop.size() < 3");
-		}
-
-		// check winding order
-		bool reverse_loop = false;
-		std::vector<carve::geom2d::P2> loop_2d( loop );
-		carve::geom::vector<3>  normal_2d = computePolygon2DNormal( loop_2d );
-		if( it_face_loops == face_loops_input.begin() )
-		{
-			normal_first_loop = normal_2d;
-			if( normal_2d.z < 0 )
-			{
-				reverse_loop = true;
-				normal_first_loop = -normal_first_loop;
-			}
-		}
-		else
-		{
-			if( normal_2d.z > 0 )
-			{
-				reverse_loop = true;
-			}
-		}
-		if( reverse_loop )
-		{
-			std::reverse( loop_2d.begin(), loop_2d.end() );
-		}
-				
-		if( loop_2d.size() < 3 )
-		{
-			throw oip::InconsistentGeometryException("GeomUtils::extrude(): loop_2d.size() < 3");
-		}
-		
-		// close loop, insert first point at end if not already there
-//		while( loop_2d.size() > 2 )
-		{
-			carve::geom::vector<2> first = loop_2d.front();
-			carve::geom::vector<2>& last = loop_2d.back();
-
-			if( abs(first.x-last.x) > 0.00001 || abs(first.y-last.y) > 0.00001 )
-			{
-				loop_2d.push_back( first );
-			}
-		}
-
-		face_loops.push_back(loop_2d);
-	}
+	std::vector<std::vector<carve::geom2d::P2>>	face_loops = GeomUtils::correctWinding(face_loops_input, normal_first_loop);
+	
 
 	bool flip_faces = false;
 	double extrusion_dot_normal = dot( extrusion_vector, normal_first_loop );
@@ -216,17 +163,8 @@ void GeomUtils::extrude(
 	std::vector<carve::geom2d::P2> merged_path;
 	std::vector<carve::triangulate::tri_idx> triangulated;
 	std::vector<std::pair<size_t, size_t>> path_all_loops;
-	try
-	{
-		path_all_loops = carve::triangulate::incorporateHolesIntoPolygon(face_loops);
-		GeomUtils::createVoids(merged_path, triangulated, path_all_loops, face_loops, err);
-	}
-	catch(...)
-	{
-		err << "carve::triangulate::incorporateHolesIntoPolygon failed " << std::endl;
-		return;
-	}
-
+	GeomUtils::incorporateVoids(face_loops, merged_path, triangulated, path_all_loops);
+	
 	// now insert points to polygon, avoiding points with same coordinates
 	std::map<double, std::map<double, int>> existing_vertices_coords;
[…]
