### 2021-05-15 1805f06
Transformed for loop
