### 2021-05-15 522c552
Modified obtainment of control points. Tried to simplify the transformation form EXPRESSContainer<EXPRESSContainer<IfcCartesianPoint>> to std::vector<std::vector<IfcCartesianPoint>>. The transformation is still necessary because FaceConverter::convertIfcCartesianPoint2DVector does not accept the EXPRESSContainer type (at least according to my current knowledge)
