### 2021-05-15 39b09d7
Created loops of actual points for IfcIndexedPolygonalFaceWithVoids
