### 2021-05-15 02d6b05
calculating 2D Plane of the Points
