### 2021-05-15 9b4c92d
Changed parameter type of FaceConverter::convertIfcCartesieanPoint2DVector ... from std::vector to EXPRESSContainer. Accordingly, the transformation in SplineConverter::convertIfcBSplineSurface isn't necessary anymore.
