+++ Core/src/IfcGeometryConverter/SplineConverter.h
@@ -167,31 +167,13 @@ namespace OpenInfraPlatform
 							//const int numControlPointsU = controlPoints.size();
 							//const int numControlPointsV = controlPoints[0].size();
 
-							// load array of control points from ifc-entity
-							// simplified, it's in the type EXPRESSContainer<EXPRESSContainer<IfcCartesianPoint>>
-							const auto controlPointList = bsplineSurfaceWithKnots->ControlPointsList;
-							// temporary control point container of type std::vector<std::vector<..>> which is needed for converter function
-							std::vector<std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcCartesianPoint>>> controlPointListVector;
-							controlPointListVector.resize(controlPointList.size());
-							// convert from EXPRESSContainer<EXPRESSContainer<..>> to std::vector<std::vector<..>>
-							// loop over u-indeces (=K1) of control points
-							for (size_t i = 0; i < controlPointList.size(); i++)
-							{
-								controlPointListVector[i].resize(controlPointList[i].size());
-								// loop over v-indeces (=K2) of control points
-								for (size_t j = 0; j < controlPointList[i].size(); j++)
-								{
-									controlPointListVector[i][j] = controlPointList[i][j];
-								}
-							}
-
 							// temporary instances to curve and face converter, potential relocation will be considered in further refactoring
 							std::shared_ptr<CurveConverterT<IfcEntityTypesT>> curveConverter =
 								std::make_shared<CurveConverterT<IfcEntityTypesT>>(CurveConverterT<IfcEntityTypesT>(GeomSettings(), UnitConvert(), placementConverter));
 							FaceConverterT<IfcEntityTypesT> faceConverter(GeomSettings(), UnitConvert(), placementConverter, curveConverter);
 
 							// obtain control points
-							const std::vector<std::vector<carve::geom::vector<3>>> controlPoints = faceConverter.convertIfcCartesianPoint2DVector(controlPointListVector);
+							const std::vector<std::vector<carve::geom::vector<3>>> controlPoints = faceConverter.convertIfcCartesianPoint2DVector(bsplineSurfaceWithKnots->ControlPointsList);
 
 							/*
 							// obtain number of knots
