### 2021-05-15 c5322c6
Refactoring of FaceConverter::convertIfcCartesianPoint2DVector. Function design according to the function CuveConverter::convertIfcCartesianPointVector which does the same one level below. Treatment of 'lengthFactor' and number of dimension are handled by the called function PlacementConverter::convertIfcCartesianPoint.
