### 2021-05-16 4db3b19
Started to implement conversion from 3D to 2D
