### 2021-05-16 e5e0196
Removed redundant code
