### 2021-05-16 80f4770
calculated plane matrix
