### 2021-05-17 48612b2
Added typename definition (alias) for EXPRESSContainer in FaceConverter
