### 2021-05-20 ff27b9d
Calculated vector, length, points
