### 2021-05-20 599dd48
Calculated vector, length, points
