### 2021-05-20 bfdad56
Calculated vector, length, points
