+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -911,7 +911,7 @@ namespace OpenInfraPlatform
 					// IfcSpiral SUBTYPE OF IfcBlossCurve
 					if (spiral.isOfType<typename IfcEntityTypesT::IfcBlossCurve>())
 					{
-						return convertIfcSpiral(spiral.as<typename IfcEntityTypesT::IfcSpiral>(),
+						return convertIfcSpiral(spiral.as<typename IfcEntityTypesT::IfcBlossCurve>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcBlossCurve
 
@@ -932,14 +932,14 @@ namespace OpenInfraPlatform
 					// IfcEllipse SUBTYPE OF IfcHelmertCurve
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcHelmertCurve>())
 					{
-						return convertIfcClothoid(conic.as<typename IfcEntityTypesT::IfcHelmertCurve>(),
+						return convertIfcHelmertCurve(conic.as<typename IfcEntityTypesT::IfcHelmertCurve>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcHelmertCurve
 
 						// IfcEllipse SUBTYPE OF IfcSine
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcSine>())
 					{
-						return convertIfcClothoid(conic.as<typename IfcEntityTypesT::IfcSine>(),
+						return convertIfcSine(conic.as<typename IfcEntityTypesT::IfcSine>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcSine
 
@@ -998,16 +998,58 @@ namespace OpenInfraPlatform
 					carve::geom::vector<3> endPoint = getPointOnCurve<typename IfcEntityTypesT::IfcClothoid>(clothoid, trim2Vec, trimmingPreference);
 
 					//get L, A, R 
-					
+					// A - Clothoid constant.
 					double A = clothoid->ClothoidConstant * this->UnitConvert()->getLengthInMeterFactor();
-				 
-				
-					//Fresnel equastion 
-					for (int i = 1; i < 200; i++)
+
+					// Calculate the length "L" between two ponts.
+					//double L = sqrt(pow(endPoint.x- firstPoint.x)+pow(endPoint.y - firstPoint.y)+ pow(endPoint.z - firstPoint.z));
+					double L = carve::geom::distance(firstPoint, endPoint);
+
+					// Get number of tesselated segments (note default is 100, write new function!)
+					int numSegments = 100;
+					// Create a tesselated vector
+					std::vector < double > ti(numSegments,0);
+					double pointInterval_ti = (endPoint.x - firstPoint.x) / (numSegments - 1);
+					for (int i = 1; i < ti.size(); ++i)
 					{
+						ti.push_back(ti[i - 1] + pointInterval_ti);
+					}
+					
+					//Calculate clothoid points. Fresnel equastion.
 
-					};
+					std::vector<carve::geom::vector<3> > clothoidPoints;
+					clothoidPoints.x[0, 0];
+					clothoidPoints.y[0, 0];
+					clothoidPoints.z[0, 0];
+					
+					for (int i = 1; i < numSegments; ++i) 
+					{
+						clothoidPoints.push_back(carve::geom::vector<3>(
+							carve::geom::VECTOR(
[…]
