+++ UserInterface/Dialogues/IfcTreeModel.cpp
@@ -1,0 +1,169 @@
+/*
+	Copyright (c) 2021 Technical University of Munich
+	Chair of Computational Modeling and Simulation.
+
+	TUM Open Infra Platform is free software; you can redistribute it and/or modify
+	it under the terms of the GNU General Public License Version 3
+	as published by the Free Software Foundation.
+
+	TUM Open Infra Platform is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#include "IfcTreeModel.h"
+
+#include <type_traits>
+#include <cstdlib>
+#include <memory>
+
+
+OpenInfraPlatform::UserInterface::IfcTreeModel::IfcTreeModel(std::shared_ptr<OpenInfraPlatform::EarlyBinding::EXPRESSModel> expressModel, QObject *parent)
+	: QAbstractItemModel(parent)
+{
+	//rootItem_ = std::make_shared<IfcTreeItem>(expressModel->entities.find(1)->second, nullptr); //make ifcProject the root item 
+	//std::string text = "Title";
+	rootItem_ = std::make_shared<IfcTreeItem>();
+	setupModelData(expressModel, rootItem_);
+}
+
+OpenInfraPlatform::UserInterface::IfcTreeModel::~IfcTreeModel()
+{
+	
+}
+
+QModelIndex OpenInfraPlatform::UserInterface::IfcTreeModel::index(int row, int column, const QModelIndex &parent) const
+{
+	if (!hasIndex(row, column, parent))
+		return QModelIndex();
+
+	std::shared_ptr<IfcTreeItem> parentItemShared;
+
+	if (!parent.isValid())
+		parentItemShared = rootItem_;
+		//return this->createIndex(row, column, nullptr);
+	else
+	{
+		IfcTreeItem *parentItem = static_cast<IfcTreeItem*>(parent.internalPointer());
+		std::shared_ptr<IfcTreeItem> parentTemp(parentItem);
+		parentItemShared = parentTemp;
+	}
+
+	std::shared_ptr<IfcTreeItem> childItemShared = parentItemShared->child(row);
+	IfcTreeItem *childItem = childItemShared.get();
+	if (childItem)
+		return createIndex(row, column, childItem);
+	return QModelIndex();
+}
+
+QVariant OpenInfraPlatform::UserInterface::IfcTreeModel::headerData(int section, Qt::Orientation orientation,
+	int role) const
+{
+	return rootItem_->data(section);
+}
+
+QModelIndex OpenInfraPlatform::UserInterface::IfcTreeModel::parent(const QModelIndex &index) const
+{
+	if (!index.isValid())
+		return QModelIndex();
+
+	IfcTreeItem *childItem = static_cast<IfcTreeItem*>(index.internalPointer());
+	std::shared_ptr<IfcTreeItem> childItemShared(childItem);
+	std::shared_ptr<IfcTreeItem> parentItemShared = childItem->parentItem();
+	IfcTreeItem *parentItem = parentItemShared.get();
+
+	if (parentItemShared == rootItem_)
+		return QModelIndex();
+
+	return createIndex(parentItem->row(), 0, parentItem);
+}
+
+int OpenInfraPlatform::UserInterface::IfcTreeModel::rowCount(const QModelIndex &parent) const
+{
[…]
