### 2021-05-21 bc8c767
Small changes to tree items (not working yet!)
