### 2021-05-21 a1f2e4a
Change everything from shared pointers to pointers (not working yet!)
