### 2021-05-22 7a17b33
Corrected function, so it can call functions from GeomUtils.cpp
