+++ Core/src/IfcGeometryConverter/FaceConverter.h
@@ -1297,7 +1297,6 @@ namespace OpenInfraPlatform {
 												point[2]) * UnitConvert()->getLengthInMeterFactor();
 
 											// apply transformation
-											vertex = pos * vertex;
 											outer_loop.push_back(vertex);
 											break;
 										}
@@ -1311,7 +1310,7 @@ namespace OpenInfraPlatform {
 									std::vector<carve::geom::vector<3>> loop;
 									for (auto innerLoopPoint : innerLoop)
 									{
-										double i = 0.;
+										size_t i = 0;
 										for (const auto& point : faceSet->Coordinates->CoordList)
 										{
 											if (i == innerLoopPoint- 1) {
@@ -1322,7 +1321,6 @@ namespace OpenInfraPlatform {
 														point[2]) * UnitConvert()->getLengthInMeterFactor();
 
 												// apply transformation
-												vertex = pos * vertex;
 												loop.push_back(vertex);
 												break;
 											}
@@ -1337,35 +1335,33 @@ namespace OpenInfraPlatform {
 								carve::geom::vector<3> normalOfPlane = GeomUtils::computePolygonNormal(loops[0]);
 
 								carve::math::Matrix planeMatrix;
-								GeomUtils::convertPlane2Matrix(normalOfPlane, carve::geom::VECTOR(0., 0., 0.),
-									outerLoop[2] - outerLoop[1], planeMatrix);
+								GeomUtils::convertPlane2Matrix(normalOfPlane, outerLoop[0],
+									outerLoop[1] - outerLoop[0], planeMatrix);
 
 								carve::math::Matrix inversePlaneMatrix = GeomUtils::computeInverse(planeMatrix);
 
-
-								for (int i = 0; i < loops.size(); i++) 
+								for (auto loop : loops) 
 								{
-									std::vector<carve::geom::vector<3>> loop = loops[i];
 									std::vector<carve::geom::vector<2>> loop2D;
-									for (int j = 0; j < loop.size(); j++)
+									std::vector<carve::geom::vector<3>> loop3Dto2D;
+									for (auto point : loop)
 									{
-										carve::geom::vector<3> point3Dto2D = inversePlaneMatrix * loop[j];
-										
-										carve::geom::vector<2> point2D = carve::geom::VECTOR(point3Dto2D.x, point3Dto2D.y);
-										loop2D.push_back(point2D);
+										carve::geom::vector<3> point3Dto2D = inversePlaneMatrix * point;
+										loop3Dto2D.push_back(point3Dto2D);
 									}
+									loop2D = removeEmptyCoordinate(loop3Dto2D);
 									loops2D.push_back(loop2D);
 								}
 								
 								// 2. Call geomUtils
-								/*carve::geom::vector<3> normal_first_loop;
-								std::vector<std::vector<carve::geom2d::P2>>	face_loops = GeomUtils::correctWinding(loops, normal_first_loop);
+								carve::geom::vector<3> normal_first_loop;
+								std::vector<std::vector<carve::geom2d::P2>>	face_loops = GeomUtils::correctWinding(loops2D, normal_first_loop);
 
 								std::vector<carve::geom2d::P2> merged_path;
 								std::vector<carve::triangulate::tri_idx> triangulated;
 								std::vector<std::pair<size_t, size_t>> path_all_loops;
 
-								GeomUtils::incorporateVoids(face_loops, merged_path, triangulated, path_all_loops);*/
+								GeomUtils::incorporateVoids(face_loops, merged_path, triangulated, path_all_loops);
 								// 3. Given the results, find corresponding points in the original array. 
 								
 							}
[…]
