### 2021-05-22 0b515e0
Setup the base structure of ifc-type handling
