+++ Core/src/IfcGeometryConverter/SplineConverter.h
@@ -152,123 +152,123 @@ namespace OpenInfraPlatform
 						//const std::vector<std::vector<carve::geom::vector<3>>>& controlPoints
 						) const throw(...)
 					{
-							// obtain degree of both b-spline curves
-							const int degreeU = bsplineSurfaceWithKnots->UDegree;
-							const int orderU = degreeU + 1;
-							const int degreeV = bsplineSurfaceWithKnots->VDegree;
-							const int orderV = degreeV + 1;
-							// obtain number of control points
-							//const int numControlPointsU = controlPoints.size();
-							//const int numControlPointsV = controlPoints[0].size();
+						// obtain degree of both b-spline curves
+						const int degreeU = bsplineSurfaceWithKnots->UDegree;
+						const int orderU = degreeU + 1;
+						const int degreeV = bsplineSurfaceWithKnots->VDegree;
+						const int orderV = degreeV + 1;
+						// obtain number of control points
+						//const int numControlPointsU = controlPoints.size();
+						//const int numControlPointsV = controlPoints[0].size();
 
-							// temporary instances to curve and face converter, potential relocation will be considered in further refactoring
-							std::shared_ptr<CurveConverterT<IfcEntityTypesT>> curveConverter =
-								std::make_shared<CurveConverterT<IfcEntityTypesT>>(CurveConverterT<IfcEntityTypesT>(GeomSettings(), UnitConvert(), placementConverter));
-							FaceConverterT<IfcEntityTypesT> faceConverter(GeomSettings(), UnitConvert(), placementConverter, curveConverter);
+						// temporary instances to curve and face converter, potential relocation will be considered in further refactoring
+						std::shared_ptr<CurveConverterT<IfcEntityTypesT>> curveConverter =
+							std::make_shared<CurveConverterT<IfcEntityTypesT>>(CurveConverterT<IfcEntityTypesT>(GeomSettings(), UnitConvert(), placementConverter));
+						FaceConverterT<IfcEntityTypesT> faceConverter(GeomSettings(), UnitConvert(), placementConverter, curveConverter);
 
-							// obtain control points
-							const std::vector<std::vector<carve::geom::vector<3>>> controlPoints = faceConverter.convertIfcCartesianPoint2DVector(bsplineSurfaceWithKnots->ControlPointsList);
+						// obtain control points
+						const std::vector<std::vector<carve::geom::vector<3>>> controlPoints = faceConverter.convertIfcCartesianPoint2DVector(bsplineSurfaceWithKnots->ControlPointsList);
 
-							/*
-							// obtain number of knots
-							const int numKnotsU = orderU + numControlPointsU;
-							const int numKnotsV = orderV + numControlPointsV;
+						/*
+						// obtain number of knots
+						const int numKnotsU = orderU + numControlPointsU;
+						const int numKnotsV = orderV + numControlPointsV;
 
-							// obtain knots for each direction
-							const std::vector<int>& knotMultsU = bsplineSurfaceWithKnots->m_UMultiplicities;
-							const std::vector<std::shared_ptr<emt::Ifc4EntityTypes::IfcParameterValue>>& splineKnotsU = bsplineSurfaceWithKnots->m_UKnots;
-							const std::vector<int>& knotMultsV = bsplineSurfaceWithKnots->m_VMultiplicities;
-							const std::vector<std::shared_ptr<emt::Ifc4EntityTypes::IfcParameterValue>>& splineKnotsV = bsplineSurfaceWithKnots->m_VKnots;
+						// obtain knots for each direction
[…]
