### 2021-05-22 48d2604
Moved the obtainment of controlPoints to a suitable place
