### 2021-05-22 31c3e93
Split function SplineConverter::loadKnotArray into two functions. The new function loadKnotArrayCurve is called with an IfcBSplineCurveWithKnots. A new function loadKnotArraySurface will be added in a following commit. Both function will call obtainKnotArray to reuse its code in both cases (curve and surface).
