### 2021-05-23 51241cd
Added function SplineConverter::loadKnotArraySurface. This function calls the previously created function obtainKnotArray.
