### 2021-05-23 775a4ab
Removed parts of legacy code (obtainment of knots)
