### 2021-05-23 0df7fb3
implemented search of the indices. Implemented addFace for IfcIndexedPolygonalFaceWithVoids
