### 2021-05-23 4bfa5a6
Structure to decide between SplineSurface and RationalSplineSurface
