### 2021-05-23 044a541
Adjust function name of convertIfcBSplineSurface in SplineConverter
