### 2021-05-23 cb17f9a
Edited function convertPlane2Matrix
