### 2021-05-23 4ce7377
Adjusted existing function loadWeightsData to be reused for surfaces. The reuse by SplineSurface will be implemented in the next commit.
