### 2021-05-23 cbfd66a
Added function loadWeightsDataSurface
