### 2021-05-24 299dc6d
Split obtainProperties into sub-functions, call them in convertSurface
