+++ Core/src/IfcGeometryConverter/SplineConverter.h
@@ -86,10 +86,9 @@ namespace OpenInfraPlatform
 							const std::vector<double> knotArray = loadKnotArrayCurve(splineCurve.as<typename IfcEntityTypesT::IfcBSplineCurveWithKnots>(), numKnotsArray);
 							const std::vector<double> weightsData = loadWeightsData((splineCurve.as<typename IfcEntityTypesT::IfcRationalBSplineCurveWithKnots>())->WeightsData);
 
-							uint32_t numCurvePoints;
+							const uint32_t numCurvePoints = obtainNumCurvePoints(knotArray.size());
 							// at the end, subtract current knot value with accuracy to avoid zero-vectors (since last knot value is excluded by definition)
-							double accuracy;
-							std::tie(numCurvePoints, accuracy) = obtainProperties(knotArray.size());
+							const double accuracy = obtainAccuracy();
 
 							std::vector<carve::geom::vector<3>> curvePoints = 
 								IfcGeometryConverter::SplineUtilities::computeRationalBSplineCurveWithKnots(order, knotArray, controlPoints, weightsData, numCurvePoints, accuracy);
@@ -103,10 +102,9 @@ namespace OpenInfraPlatform
 						{
 							const std::vector<double> knotArray = loadKnotArrayCurve(splineCurve.as<typename IfcEntityTypesT::IfcBSplineCurveWithKnots>(), numKnotsArray);
 
-							uint32_t numCurvePoints;
+							const uint32_t numCurvePoints = obtainNumCurvePoints(knotArray.size());
 							// at the end, subtract current knot value with accuracy to avoid zero-vectors (since last knot value is excluded by definition)
-							double accuracy;
-							std::tie(numCurvePoints, accuracy) = obtainProperties(knotArray.size());
+							const double accuracy = obtainAccuracy();
 
 							std::vector<carve::geom::vector<3>> curvePoints = 
 								IfcGeometryConverter::SplineUtilities::computeBSplineCurveWithKnots(order, knotArray, controlPoints, numCurvePoints, accuracy);
@@ -166,6 +164,11 @@ namespace OpenInfraPlatform
 							orderU + controlPoints.size(), // number of knots in u direction
 							orderV + controlPoints[0].size()); // number of knots in v direction
 
+						const uint32_t numCurvePointsU = obtainNumCurvePoints(knotsU.size());
+						const uint32_t numCurvePointsV = obtainNumCurvePoints(knotsV.size());
+						// at the end, subtract current knot value with accuracy to avoid zero-vectors (since last knot value is excluded by definition)
+						const double accuracy = obtainAccuracy();
+
 						// distinction between IfcBSplineSurfaceWithKnots and IfcRationalBSplineSurfaceWithKnots
 						// (2/2) reverse order - IfcRationalBSplineSurfaceWithKnots SUBTYPE of IfcBSplineSurfaceWithKnots
 						if (bsplineSurfaceWithKnots.isOfType<typename IfcEntityTypesT::IfcRationalBSplineSurfaceWithKnots>())
@@ -183,12 +186,9 @@ namespace OpenInfraPlatform
 						}
 						/*
 						// reserve memory for all surface points on b-spline surface
-						//! TEMPORARY default number of curve points
-						const uint8_t numCurvePointsU = splineKnotsU.size() * 10;
-						const uint8_t numCurvePointsV = splineKnotsV.size() * 10;
 						std::vector<carve::geom::vector<3>> curvePoints;
 						curvePoints.reserve(numCurvePointsU * numCurvePointsV);
-
+						
 						computeBSplineSurface(orderU, orderV, numCurvePointsU, numCurvePointsV,
[…]
