### 2021-05-24 f868123
Changed structure of curvePoints from vector to matrix. matrix means vector of vectors, it might be easier to read and understand - especially because it's a 2D field of points.
