### 2021-05-24 9af9691
Removed deprecated entry in existing doxygen comment
