### 2021-05-24 4f1a128
Implementation of rational-BSpline-Surface calculation The function SplineUtilities::computeRationalBSplineSurfaceWithKnots computes the entire rational surface, it uses the function computeRationalPointOfBSplineSurface. The appropriate function call was added in SplineConverter.
