### 2021-05-24 957bd7a
Implementation of BSpline-Surface calculation. The function SplineUtilities::computeBSplineSurfaceWithKnots computes the entire surface, it uses the function computePointOfBSplineSurface. The appropriate function call was added in SplineConverter. The implementation of a rational BSpline-Surface will be implemented in a separate commit.
