+++ Core/src/IfcGeometryConverter/SplineUtilities.h
@@ -137,6 +137,80 @@ namespace OpenInfraPlatform
 					return curvePoints;
 				}
 
+				/*! \brief Computes the surface points of the B-Spline surface.
+				 *
+				 * All information has to be loaded from an \c IfcBSplineSurfaceWithKnots entity before calling this function.
+				 *
+				 * \param[in]	orderU				In direction u, order of the B-Spline or rather the basis functions ( =degree+1 )
+				 * \param[in]	orderV				In direction v, order of the B-Spline or rather the basis functions ( =degree+1 )
+				 * \param[in]	knotArrayU			The array / vector of knots in direction u, the function \c loadKnotArray gives this vector.
+				 * \param[in]	knotArrayV			The array / vector of knots in direction v, the function \c loadKnotArray gives this vector.
+				 * \param[in]	controlPoints		The field of the B-Spline surface control points.
+				 * \param[in]	numCurvePointsU		The number of curve points in direction u where the surface s(u,v) has to be evaluated.
+				 * \param[in]	numCurvePointsV		The number of curve points in direction v where the surface s(u,v) has to be evaluated.
+				 * \param[in]	accuracy			Accuracy which is technically needed in the calculation.
+				 *
+				 * \return		The vector of vectors of surface points, which can be rendered in a viewport after some additional transformation.
+				 */
+				static std::vector<std::vector<carve::geom::vector<3>>> computeBSplineSurfaceWithKnots(
+					const int orderU,
+					const int orderV,
+					const std::vector<double>& knotsU,
+					const std::vector<double>& knotsV,
+					const std::vector<std::vector<carve::geom::vector<3>>>& controlPoints,
+					const uint32_t numCurvePointsU, 
+					const uint32_t numCurvePointsV,
+					const double accuracy)  throw(...)
+				{
+					// The following parameters corresponds to the parameter u of a surface s(u,v)
+					double knotUStart;
+					double knotUEnd;
+					double stepU;
+					std::tie(knotUStart, knotUEnd, stepU) = obtainKnotRange(orderU, knotsU, numCurvePointsU);
+					// The following parameters corresponds to the parameter v of a surface s(u,v)
+					double knotVStart;
+					double knotVEnd;
+					double stepV;
+					std::tie(knotVStart, knotVEnd, stepV) = obtainKnotRange(orderV, knotsV, numCurvePointsV);
+
+					// target vector of vectors (field of points)
+					std::vector<std::vector<carve::geom::vector<3>>> curvePoints;
+					curvePoints.resize(numCurvePointsU);
+					for (std::vector<carve::geom::vector<3>>& it : curvePoints)
+						it.resize(numCurvePointsV);
+
+					// u: representatives of the parameter vectors, start with first valid knots
+					double u = knotUStart;
+					
+
+					// the for-loop iterates over all values u of the parameter vector
+					for (size_t j = 0; j < numCurvePointsU; ++j) 
+					{
+						// if at last curve point, use u slightly left of last knot value (because of definition in base functions)
+						if (j == numCurvePointsU - 1) { u = knotUEnd - accuracy; }
+
+						// v: representatives of the parameter vectors, start with first valid knots
+						double v = knotVStart;
+
+						// the for-loop iterates over all values v of the parameter vector
+						for (size_t i = 0; i < numCurvePointsV; ++i)
+						{
+							// if at last curve point, use v slightly left of last knot value (because of definition in base functions)
+							if (i == numCurvePointsV - 1) { v = knotVEnd - accuracy; }
+
+							curvePoints[j][i] = computePointOfBSplineSurface(orderU, orderV, u, v, controlPoints, knotsU, knotsV);
+							
[…]
