### 2021-05-24 c627ad8
Apply pos to the curvePoints pos: The relative location of the origin of the representation's coordinate system within the geometric context.
