### 2021-05-24 cf79e56
Adjust doxygen comment
