### 2021-05-24 5dbdffe
Declaration of return value, out-commented last piece of code. The code compiles and runs: The mashes of the IfcBSplineSurfaceWithKnots and of the IfcRationalBSplineWithKnots are displayed. Probably, this last piece of code still needs some refactoring - or at least some comments.
