### 2021-05-24 2e63404
Add comments to polyLine-construction
