### 2021-05-24 61195a1
Removed legacy code & comments
