+++ Core/src/IfcGeometryConverter/SplineConverter.h
@@ -457,99 +457,7 @@ namespace OpenInfraPlatform
 						// ! TEMPORARY default number of curve points
 						return numKnotsArray * 10;
 					}
-
-					// B-Spline surface definition according to: 
-					// http://www.buildingsmart-tech.org/ifc/IFC4/final/html/schema/ifcgeometryresource/lexical/ifcbsplinesurface.htm
-					void computeBSplineSurface(
-						const uint8_t orderU,
-						const uint8_t orderV,
-						const uint32_t numCurvePointsU,
-						const uint32_t numCurvePointsV,
-						const uint32_t numControlPointsU,
-						const uint32_t numControlPointsV,
-						const std::vector<std::vector<carve::geom::vector<3>>>& controlPoints,
-						const std::vector<std::vector<double>>& weights,
-						const std::vector<double>& knotVectorU,
-						const std::vector<double>& knotVectorV,
-						std::vector<carve::geom::vector<3>>& curvePoints) const throw(...)
-					{
-						// curve is defined for [t_p;t_m-p], m := number of knots - 1
-						const uint32_t firstIndexU = orderU - 1;
-						const uint32_t lastIndexU = knotVectorU.size() - orderU;
-						const uint32_t firstIndexV = orderV - 1;
-						const uint32_t lastIndexV = knotVectorV.size() - orderV;
-
-						const double knotStartU = knotVectorU[firstIndexU];
-						const double knotEndU = knotVectorU[lastIndexU];
-						const double knotRangeU = knotEndU - knotStartU;
-
-						const double knotStartV = knotVectorV[firstIndexV];
-						const double knotEndV = knotVectorV[lastIndexV];
-						const double knotRangeV = knotEndV - knotStartV;
-
-						// compute step size for each direction
-						const double stepU = knotRangeU / static_cast<double>(numCurvePointsU - 1);
-						const double stepV = knotRangeV / static_cast<double>(numCurvePointsV - 1);
-
-						std::vector<double> basisFuncsU(numControlPointsU, 0.0);
-						std::vector<double> basisFuncsV(numControlPointsV, 0.0);
-
-						const double accuracy = 0.0000001;
-
-						// start with first valid knot in each direction
-						double tV = knotStartV;
-
-						for (int j = 0; j < numCurvePointsV; ++j) {
-							if (j == numCurvePointsV - 1) { tV = knotEndV - accuracy; }
-
-							double tU = knotStartU;
-
-							for (int i = 0; i < numCurvePointsU; ++i) {
-								if (i == numCurvePointsU - 1) { tU = knotEndU - accuracy; }
-
-								// 1) Evaluate basis functions at curve point tU and tV
-								IfcGeometryConverter::SplineUtilities::computeBSplineBasisFunctions(orderU, tU, numControlPointsU, knotVectorU, basisFuncsU);
-								IfcGeometryConverter::SplineUtilities::computeBSplineBasisFunctions(orderV, tV, numControlPointsV, knotVectorV, basisFuncsV);
-
-								// 2) Compute exact point on surface
-								carve::geom::vector<3> point = carve::geom::VECTOR(0, 0, 0);
-
-								// 2i) If B-spline surface is rational, weights and their sum have to considered, as well
-								double weightSum = 0.0;
-
-								for (int x = 0; x < numControlPointsU; ++x) {
-									const double basisFuncU = basisFuncsU[x];
-
-									for (int y = 0; y < numControlPointsV; ++y) {
-										const double basisFuncV = basisFuncsV[y];
-										const carve::geom::vector<3>& controlPoint = controlPoints[x][y];
-
-										if (!weights.empty()) {
-											// 3a) apply formula for rational B-spline surfaces
[…]
