### 2021-05-24 a373272
Moved code of PolylineSetData construction ... from SplineConverter into FaceConverter.
