### 2021-05-24 7bafb5a
Moved comment about information-only-attributes to right place. Added 'For information only' to SurfaceForm, corresponding to 'Definition according to ISO/CD 10303-42:1992' (https://standards.buildingsmart.org/IFC/RELEASE/IFC4_1/FINAL/HTML/schema/ifcgeometryresource/lexical/ifcbsplinesurface.htm)
