### 2021-05-25 a45f2b5
Adjustment with respect to double initialization
