### 2021-05-25 cdd784e
Add doxycomment to the function convertPlane2Matrix
