### 2021-05-25 482f403
Forgot about changes in GeomUtils.h
