### 2021-05-25 9cc48d8
Correction to the IndexedPolygonalFaceWithVoids. Now it works
