### 2021-05-25 2f6cea6
Some simplifications to the IndexedPolygonalFaceWithVoids.
