### 2021-05-27 3c782b8
Copy-Paste code of convertIfcFixedReferenceSweptAreaSolid ... form previous commit 6035ec2 "implemented sopport for IfcFixedReferenceSweptAreaSolid (Stefan Jaud, 15th March 2021), respectively commit c0f621c "Merge pull request # 406 from pjanck/Solving405" (Stefan Jaud, 16th March 2021)
