+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -818,6 +818,9 @@ namespace OpenInfraPlatform
 
 				// Get swept area and position (attributes 1-2). 
 				oip::EXPRESSReference<typename IfcEntityTypesT::IfcProfileDef> sweptArea = sweptAreaSolid->SweptArea;
+				// swept area
+				std::shared_ptr<ProfileConverterT<IfcEntityTypesT>> profile_converter = profileCache->getProfileConverter(sweptArea);
+				const std::vector<std::vector<carve::geom::vector<2> > >& profile_coords = profile_converter->getCoordinates();
 
 				carve::math::Matrix sweptAreaPos(pos);	// check if local coordinate system is specified for extrusion
 				if (sweptAreaSolid->Position)
@@ -833,7 +836,7 @@ namespace OpenInfraPlatform
 				{
 					convertIfcFixedReferenceSweptAreaSolid(
 						sweptAreaSolid.template as<IfcEntityTypesT::IfcFixedReferenceSweptAreaSolid>(),
-						sweptAreaPos, itemData);
+						profile_coords, sweptAreaPos, itemData);
 					return;
 				}
 				
@@ -870,21 +873,103 @@ namespace OpenInfraPlatform
 
 			void convertIfcFixedReferenceSweptAreaSolid(
 				const EXPRESSReference<typename IfcEntityTypesT::IfcFixedReferenceSweptAreaSolid>& fixedRefSweptAreaSolid,
+				const std::vector<std::vector<carve::geom::vector<2> > >& profile_coords,
 				const carve::math::Matrix& pos,
 				std::shared_ptr<ItemData> itemData
 			) const noexcept(false)
 			{
-				throw oip::UnhandledException(fixedRefSweptAreaSolid);
-				/*
-				// Get directrix, start parameter, end parameter and fixed reference (attributes 3-6).
-					//std::shared_ptr<typename IfcEntityTypesT::IfcCurve> directrix =
-					//	fixed_ref_sweptAreaSolid->Directrix.lock();	// TO DO: formal proposition: if no StartParam or EndParam, Directrix has to be a bounded or closed curve.
-				double startParam = fixed_ref_sweptAreaSolid->StartParam;	// TO DO: optional
-				double endParam = fixed_ref_sweptAreaSolid->EndParam;		// TO DO: optional
-				//std::shared_ptr<typename IfcEntityTypesT::IfcDirection> fixed_ref =
-				//	fixed_ref_sweptAreaSolid->FixedReference.lock();
+				
 
-				// TO DO: implement//*/
+				// Get directrix, start parameter, end parameter and fixed reference (attributes 3-6).
+				std::vector<carve::geom::vector<3> > segment_start_points;
+				std::vector<carve::geom::vector<3> > basis_curve_points;
+				curveConverter->convertIfcCurve(
+					fixedRefSweptAreaSolid->Directrix, basis_curve_points, segment_start_points);
+
+				double start_param = fixedRefSweptAreaSolid->StartParam.value_or(0.0);	// TO DO: optional
+				double end_param = fixedRefSweptAreaSolid->EndParam.value_or(1.0);		// TO DO: optional
+
+				carve::geom::vector<3> fixedRef = placementConverter->convertIfcDirection(
+					fixedRefSweptAreaSolid->FixedReference);
+
+				std::shared_ptr<carve::input::PolyhedronData> polyhedron_data(new carve::input::PolyhedronData());
+				itemData->closed_polyhedrons.push_back(polyhedron_data);
+
+				// create vertices
+				size_t num_segments = basis_curve_points.size();
+				for (size_t i = 0; i < num_segments; ++i)
+				{
+					size_t
+						prev = (i == 0 ? 0 : i - 1),
[…]
