### 2021-05-28 6038f35
Fixes bug "vector subscript out of range". Unfortunately removing outerLoop and outerIndexLoop does not work properly
