### 2021-05-28 7fc7e0c
moved function removeEmptyCoordinate to the GeomUtils
