### 2021-05-28 4f41d28
movedIfcIndexedPoltgonalFaceWithVoids to the separate function
