+++ Core/src/IfcGeometryConverter/FaceConverter.h
@@ -1134,7 +1134,7 @@ namespace OpenInfraPlatform {
 					const EXPRESSReference<typename IfcEntityTypesT::IfcTessellatedItem>& tessItem,
 					const carve::math::Matrix& pos,
 					std::shared_ptr<ItemData>& itemData
-					) const throw(...) 
+				) const throw(...)
 				{
 					if (tessItem.as<typename IfcEntityTypesT::IfcTriangulatedFaceSet>())
 					{
@@ -1159,7 +1159,7 @@ namespace OpenInfraPlatform {
 
 						auto& coordinatesIndices = faceSet->CoordIndex;
 						auto& pnIndices = faceSet->PnIndex; // optional
-						
+
 						std::vector<int> flags;
 						if (tessItem.isOfType<typename IfcEntityTypesT::IfcTriangulatedIrregularNetwork>())
 						{
@@ -1176,7 +1176,7 @@ namespace OpenInfraPlatform {
 							{
 								throw oip::InconsistentModellingException(tessItem, "invalid size of coordIndex attribute.");
 							}
-							
+
 							// determine vertices' indices
 							size_t i0, i1, i2;
 
@@ -1202,7 +1202,7 @@ namespace OpenInfraPlatform {
 								// add break line
 								else
 								{
-									if (flags[i-1] & 1) // first flag set
+									if (flags[i - 1] & 1) // first flag set
 									{
 										std::shared_ptr<carve::input::PolylineSetData> polylineData = std::make_shared<carve::input::PolylineSetData>();
 										polylineData->beginPolyline();
@@ -1212,7 +1212,7 @@ namespace OpenInfraPlatform {
 										polylineData->addPolylineIndex(1);
 										itemData->polylines.push_back(polylineData);
 									}
-									if (flags[i-1] & 2) // second flag set
+									if (flags[i - 1] & 2) // second flag set
 									{
 										std::shared_ptr<carve::input::PolylineSetData> polylineData = std::make_shared<carve::input::PolylineSetData>();
 										polylineData->beginPolyline();
@@ -1222,7 +1222,7 @@ namespace OpenInfraPlatform {
 										polylineData->addPolylineIndex(1);
 										itemData->polylines.push_back(polylineData);
 									}
-									if (flags[i-1] & 4) // third flag set
+									if (flags[i - 1] & 4) // third flag set
 									{
 										std::shared_ptr<carve::input::PolylineSetData> polylineData = std::make_shared<carve::input::PolylineSetData>();
 										polylineData->beginPolyline();
@@ -1246,11 +1246,11 @@ namespace OpenInfraPlatform {
 					{
 						std::shared_ptr<carve::input::PolyhedronData> polygon(new carve::input::PolyhedronData());
 						auto faceSet = tessItem.template as<typename IfcEntityTypesT::IfcPolygonalFaceSet>();
-						
+
 						// obtain vertices from coordinates list and add them to the new polygon
 						for (const auto& point : faceSet->Coordinates->CoordList)
 						{
-							
+
 							carve::geom::vector<3> vertex = //placementConverter->convertIfcPoint(point);
 								carve::geom::VECTOR(point[0],
 									point[1],
@@ -1266,7 +1266,7 @@ namespace OpenInfraPlatform {
 						auto& faces = faceSet->Faces;
 						auto& pnIndices = faceSet->PnIndex; // optional
 
-						for( auto& face : faces )
+						for (auto& face : faces)
 						{
 							if (!face.template isOfType<typename IfcEntityTypesT::IfcIndexedPolygonalFaceWithVoids>())
 							{
@@ -1276,110 +1276,118 @@ namespace OpenInfraPlatform {
[…]
