### 2021-05-28 63bee28
Refactoring function convertIfcTessellatedItem
