### 2021-05-28 1888ac2
Added Doxy-comments to new functions
