### 2021-05-28 778e0df
Removing redundant code
