+++ Core/src/IfcGeometryConverter/FaceConverter.h
@@ -1283,8 +1283,7 @@ namespace OpenInfraPlatform {
 					std::shared_ptr<ItemData>& itemData) const noexcept(false)
 				{
 					std::shared_ptr<carve::input::PolyhedronData> polygon(new carve::input::PolyhedronData());
-					//auto faceSet = tessItem.template as<typename IfcEntityTypesT::IfcPolygonalFaceSet>();
-
+					
 					// obtain vertices from coordinates list and add them to the new polygon
 					for (const auto& point : faceSet->Coordinates->CoordList)
 					{
