### 2021-05-30 aeedbd6
Created unit test for polygonal-face-tessellation
