### 2021-05-30 98c024b
Added areEqual function to GeomUtils.h
