### 2021-05-30 875f326
Fixed mistakes in Spiral function
