+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -855,7 +855,7 @@ namespace OpenInfraPlatform {
 					GeomUtils::appendPointsToCurve(basisCurvePoints, targetVec);
 				}
 
-				// IfcSpline SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
+				// IfcSpiral SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
 				/**********************************************************************************************/
 				/*! \brief Converts an \c IfcSpiral to a tesselated curve.
@@ -889,35 +889,35 @@ namespace OpenInfraPlatform {
 					//		Position: OPTIONAL IfcAxis2Placement;
 					//	END_ENTITY;
 					// **************************************************************************************************************************
-					// IfcSpiral SUBTYPE OF IfcBlossCurve
+					// IfcBlossCurve SUBTYPE OF IfcSpiral
 					if (spiral.isOfType<typename IfcEntityTypesT::IfcBlossCurve>())
 					{
 						return convertIfcSpiral(spiral.as<typename IfcEntityTypesT::IfcBlossCurve>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcBlossCurve
 
-					// IfcEllipse SUBTYPE OF IfcClothoid
+					// IfcClothoid SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcClothoid>())
 					{
 						return convertIfcClothoid(conic.as<typename IfcEntityTypesT::IfcClothoid>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcClothoid
 
-					// IfcEllipse SUBTYPE OF IfcCosine
+					// IfcCosine SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcCosine>())
 					{
 						return convertIfcCosine(conic.as<typename IfcEntityTypesT::IfcCosine>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcCosine
 
-					// IfcEllipse SUBTYPE OF IfcHelmertCurve
+					// IfcHelmertCurve SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcHelmertCurve>())
 					{
 						return convertIfcHelmertCurve(conic.as<typename IfcEntityTypesT::IfcHelmertCurve>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcHelmertCurve
 
-						// IfcEllipse SUBTYPE OF IfcSine
+					// IfcSine SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcSine>())
 					{
 						return convertIfcSine(conic.as<typename IfcEntityTypesT::IfcSine>(),
@@ -927,7 +927,7 @@ namespace OpenInfraPlatform {
 					// the rest we do not support
 					throw oip::UnhandledException(spiral);
 				}
-				// IfcClothoid SUBTYPE of IfcSpline (exists starting IFC4X3_RC3)
+				// IfcClothoid SUBTYPE of IfcSpiral (exists starting IFC4X3_RC3)
 				/**********************************************************************************************/
[…]
