### 2021-05-30 c71bf09
Fixed mistakes in Spiral function
