+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -870,10 +870,14 @@ namespace OpenInfraPlatform
 				}
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 				// IfcClothoid SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
 =======
 				// IfcSpline SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
+=======
+				// IfcSpiral SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
+>>>>>>> 875f326b (Fixed mistakes in Spiral function)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
 >>>>>>> 92a31da5 (Added IfsSpiral converter)
 				/**********************************************************************************************/
@@ -908,35 +912,35 @@ namespace OpenInfraPlatform
 					//		Position: OPTIONAL IfcAxis2Placement;
 					//	END_ENTITY;
 					// **************************************************************************************************************************
-					// IfcSpiral SUBTYPE OF IfcBlossCurve
+					// IfcBlossCurve SUBTYPE OF IfcSpiral
 					if (spiral.isOfType<typename IfcEntityTypesT::IfcBlossCurve>())
 					{
 						return convertIfcSpiral(spiral.as<typename IfcEntityTypesT::IfcBlossCurve>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcBlossCurve
 
-					// IfcEllipse SUBTYPE OF IfcClothoid
+					// IfcClothoid SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcClothoid>())
 					{
 						return convertIfcClothoid(conic.as<typename IfcEntityTypesT::IfcClothoid>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcClothoid
 
-					// IfcEllipse SUBTYPE OF IfcCosine
+					// IfcCosine SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcCosine>())
 					{
 						return convertIfcCosine(conic.as<typename IfcEntityTypesT::IfcCosine>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcCosine
 
-					// IfcEllipse SUBTYPE OF IfcHelmertCurve
+					// IfcHelmertCurve SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcHelmertCurve>())
 					{
 						return convertIfcHelmertCurve(conic.as<typename IfcEntityTypesT::IfcHelmertCurve>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcHelmertCurve
 
-						// IfcEllipse SUBTYPE OF IfcSine
+					// IfcSine SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcSine>())
 					{
 						return convertIfcSine(conic.as<typename IfcEntityTypesT::IfcSine>(),
@@ -946,7 +950,7 @@ namespace OpenInfraPlatform
 					// the rest we do not support
 					throw oip::UnhandledException(spiral);
 				}
-				// IfcClothoid SUBTYPE of IfcSpline (exists starting IFC4X3_RC3)
[…]
