### 2021-05-30 0a30e04
Fixed mistakes in Spiral function
