### 2021-05-31 ece748a
Refactoring GeometrySettings
