+++ Core/src/IfcGeometryConverter/GeometrySettings.h
@@ -72,11 +72,9 @@ namespace OpenInfraPlatform
 
 				\return The number of vertices
 				*/
-				int getNumberOfVerticesForTessellation(const double dRadius, const double dArcStart, const double dArcEnd)
-				{
-					return getNumberOfVerticesForTessellation(dRadius, abs(dArcStart - dArcEnd));
-				}
+				int getNumberOfVerticesForTessellation(const double dRadius, const double dArcStart, const double dArcEnd);
 				
+
 				/*! Calculates the number of vertices needed for tessellation of an arc
 
 				\param[in] dRadius		The radius of the arc
@@ -84,22 +82,8 @@ namespace OpenInfraPlatform
 
 				\return The number of vertices
 				*/
-				int getNumberOfVerticesForTessellation(const double dRadius, const double dArcExtent = 2. * M_PI)
-				{
-					// if radius is smaller then the precision of the model, it is a straight (curvature -> infinity)
-					if (dRadius < getPrecision())
-						return 2;
-
-					// what's the biggest angle so that the precision still holds
-					// that is, the maximum distance between the arc and the line between two points on the arc < precision
-					double alpha = acos((dRadius - getTessellationPrecision()) / dRadius);
-
-					// we need at least this many segments along the arc
-					int numOfVertices = (int) ceil(dArcExtent / alpha);
-
-					// return at least 2 (start & end)
-					return carve::util::max_functor()(numOfVertices , 2);
-				}
+				int getNumberOfVerticesForTessellation(const double dRadius, const double dArcExtent = 2. * M_PI);
+				
 
 				/*! Calculates the arc segment length when tessellating an arc
 
@@ -108,10 +92,8 @@ namespace OpenInfraPlatform
 
 				\return The angular length of the segment [in radians]
 				*/
-				double getAngleLength(const double dRadius, const double dArcExtent = 2.0 * M_PI)
-				{
-					return dArcExtent / (double)(getNumberOfSegmentsForTessellation(dRadius, dArcExtent));
-				}
+				double getAngleLength(const double dRadius, const double dArcExtent = 2.0 * M_PI);
+				
 
 				/*! Calculates the number of segments when tessellating an arc
 
@@ -120,27 +102,21 @@ namespace OpenInfraPlatform
 
 				\return The number of segments [in radians]
 				*/
-				int getNumberOfSegmentsForTessellation(const double dRadius, const double dArcExtent = 2.0 * M_PI)
-				{
-					return getNumberOfVerticesForTessellation(dRadius, dArcExtent) - 1;
-				}
+				int getNumberOfSegmentsForTessellation(const double dRadius, const double dArcExtent = 2.0 * M_PI);
+				
 
 				/*! returns the precision of the model
 				
 				
 				*/
-				double getPrecision() const
-				{
-					return carve::EPSILON; //TODO remove constant and replace with content from IFC (i.e. introduce member, getter / setter)
-				}
+				double getPrecision() const;
+				
 				
 				/*! returns the tessellation precision of the model
 
 				*/
-				double getTessellationPrecision() const
-				{
-					return 0.01;
-				}
+				double getTessellationPrecision() const;
+				
 
 				/*! \brief Normalizes given angle to lie within the specified interval.
 
@@ -151,20 +127,9 @@ namespace OpenInfraPlatform
 				\note The interval is increased to a minimum of \c dMin \c + \c 2*PI if needed.
 				\note The borders are swaped to fulfil \c dMin \c < \c dMax.
 				*/
-				void normalizeAngle(double& dAngle, double dMin = 0., double dMax = M_TWOPI)
-				{
-					if (dMax < dMin)
-						std::swap(dMin, dMax);
-					if (dMax - dMin < M_TWOPI)
-						dMax = dMin + M_TWOPI;
[…]
