### 2021-06-02 2af454e
Added tests IFCHasAnEssentialEntity and CountEssentialEntities to unit test.
