### 2021-06-02 68f8c49
Adds doxy-comments to GeomUtils.h
