+++ Core/src/IfcGeometryConverter/GeomUtils.cpp
@@ -340,6 +340,8 @@ void GeomUtils::extrude(
 	}
 }
 
+/**********************************************************************************************/
+
 void GeomUtils::incorporateVoids(const std::vector<std::vector<carve::geom2d::P2>>& face_loops_input,
 	std::vector<carve::geom2d::P2>& merged_path,
 	std::vector<carve::triangulate::tri_idx>& triangulated, 
@@ -392,6 +394,8 @@ void GeomUtils::incorporateVoids(const std::vector<std::vector<carve::geom2d::P2
 		throw oip::InconsistentGeometryException("Error in function GeomUtils::incorporateVoids");
 	}
 }
+
+/**********************************************************************************************/
 
 std::vector<std::vector<carve::geom2d::P2>> GeomUtils::correctWinding(
 	const std::vector<std::vector<carve::geom::vector<2>>> & face_loops_input, carve::geom::vector<3>& normal_first_loop)
@@ -455,12 +459,7 @@ std::vector<std::vector<carve::geom2d::P2>> GeomUtils::correctWinding(
 }
 
 /**********************************************************************************************/
-/*! \brief Computes an inverse of the matrix.
-*
-* \param[in] matrix_a				An initial invertible matrix.
-*
-* \return							Inverted matrix
-*/
+
 carve::math::Matrix GeomUtils::computeInverse(const carve::math::Matrix& matrix_a)
 {
 	int i, j;	// col, row
@@ -666,13 +665,14 @@ bool GeomUtils::isPointOnLineSegment( double& target_lambda,
 }
 
 /**********************************************************************************************/
-/*! \brief Calculates distances between start points of the segment and their intersection. 
-* \param[in] initialPointSeg1		Initial point of the first intersecting segment.			
+/*! \brief Calculates distances between start points of the segment and their intersection.
+* \param[in] initialPointSeg1		Initial point of the first intersecting segment.
 * \param[in] terminalPointSeg1		Terminal point of the first intersecting segment.
-* \param[in] initialPointSeg2		Initial point of the second intersecting segment.			
+* \param[in] initialPointSeg2		Initial point of the second intersecting segment.
 * \param[in] terminalPointSeg2		Terminal point of the second intersecting segment.
-* \param[out] distToIntesection1	Distance from initialPointSeg1 to the segment intresection, which is calculated as section of the first segment. 
-* \param[out] distToIntesection2	Distance from initialPointSeg2 to the segment intresection, which is calculated as section of the second segment. 
+* \param[out] distToIntesection1	Distance from initialPointSeg1 to the segment intresection, which is calculated as section of the first segment.
+* \param[out] distToIntesection2	Distance from initialPointSeg2 to the segment intresection, which is calculated as section of the second segment.
+
 * \return							Returns true, if lines are intersecting with each other. False otherwise.
 */
 bool LineToLineIntersectionHelper(const carve::geom::vector<2>& initialPointSeg1,
@@ -756,15 +756,9 @@ bool GeomUtils::LineSegmentToLineSegmentIntersection(const carve::geom::vector<2
 	}
 	return false;
 }
+
 /**********************************************************************************************/
-/*! \brief Calculates coordinates of the intersection point.
-* \param[in] initialPointSeg1		Initial point of the first intersecting line.
-* \param[in] terminalPointSeg1		Terminal point of the first intersecting line.
-* \param[in] initialPointSeg2		Initial point of the second intersecting line.
-* \param[in] terminalPointSeg2		Terminal point of the second intersecting line.
-* \param[out] intersectionPoint		Coordinates of the intersection point between first line segment and second line segment.
[…]
