+++ Core/src/IfcGeometryConverter/GeomUtils.h
@@ -65,93 +65,280 @@ namespace OpenInfraPlatform
 			class GeomUtils {
 			public:
 
+				/**********************************************************************************************/
+				/*! \brief Compare two doubles for equality to the custom precision
+				*
+				* \param[in] first					The first value
+				* \param[in] second					The second value
+				* \param[in] precision				Precision of model
+				*
+				* \return							True, if the absolute difference is smaller than the precision. False otherwise
+				*/
 				static bool GeomUtils::areEqual(const double first, 
 					const double second, const double precision);
 
+				/**********************************************************************************************/
+				/*! \brief Calculates a geometric center of the plane figure. 
+				*
+				* \param[in] polygon				Polygon, which contains points of a plane figure. 
+				*
+				* \return							Geometric center of the plane figure. 
+				*/
 				static carve::geom::vector<3> computePolygonCentroid(
-					const std::vector<carve::geom::vector<3> >& polygon);
-				static carve::geom::vector<3> computePolygonNormal(
-					const std::vector<carve::geom::vector<3> >& polygon);
-				static carve::geom::vector<3> computePolygon2DNormal(
-					const std::vector<carve::geom::vector<2> >& polygon);
+					const std::vector<carve::geom::vector<3>>& polygon);
 
+				/**********************************************************************************************/
+				/*! \brief Calculates a normal of the plane using points of the polygon on this plane.
+				*
+				* \param[in] polygon				Polygon with points (in 3D) on the plane. 
+				*
+				* \return							Normal vector of the plane. 
+				*/
+				static carve::geom::vector<3> computePolygonNormal(
+					const std::vector<carve::geom::vector<3>>& polygon);
+
+				/**********************************************************************************************/
+				/*! \brief Calculates a normal of the plane using points of the polygon on this plane.
+				*
+				* \param[in] polygon				Polygon with points (in 2D) on the plane.
+				*
+				* \return							Normal vector of the plane.
+				*/
+				static carve::geom::vector<3> computePolygon2DNormal(
+					const std::vector<carve::geom::vector<2>>& polygon);
+
+				/**********************************************************************************************/
+				/*! \brief Calculates coordinates of the intersection point.
+				*
+				* \param[in] initialPointSeg1		Initial point of the first intersecting segment.
+				* \param[in] terminalPointSeg1		Terminal point of the first intersecting segment.
+				* \param[in] initialPointSeg2		Initial point of the second intersecting segment, which describes the direction of the line.
+				* \param[in] terminalPointSeg2		Terminal point of the second intersecting segment, which describes the direction of the line.
+				* \param[out] intersectionPoint		Coordinates of the intersection point between first line segment and second line.
+				*
+				* \return							Returns true, if lines are intersecting with each other. False otherwise.
+				*/
 				static bool LineSegmentToLineIntersection(const carve::geom::vector<2>& initialPointSeg1,
 					const carve::geom::vector<2>& terminalPointSeg1,
 					const carve::geom::vector<2>& initialPointSeg2,
 					const carve::geom::vector<2>& terminalPointSeg2,
 					carve::geom::vector<2> & intersectionPoint);
 
+				/**********************************************************************************************/
+				/*! \brief Calculates coordinates of the intersection point.
+				*
+				* \param[in] initialPointSeg1		Initial point of the first intersecting segment, which describes the direction of the line.
+				* \param[in] terminalPointSeg1		Terminal point of the first intersecting segment, which describes the direction of the line.
[…]
