### 2021-06-02 5e4618a
Adds description to GeomUtils class
