### 2021-06-09 36d839f
Renamed function FaceConverter::convertIfcCartesianPoint2DVector ... to convertIfcCartesianPointVectorVector. Adjusted variable names inside the function to prevent confusion. Added an additional comment to the typename definition of EXPRESSContainer in FaceConverter.
