### 2021-06-09 0787fe7
Implemented polyhedron construction in convertIfcBSplineSurface
