### 2021-06-11 3ae78cd
Added itemData as parameter of all sub-calls from convertIfcSurface. Changed return values accordingly. itemData will be needed to store the polyline-grid and the polyhedrons.
