+++ Core/src/IfcGeometryConverter/FaceConverter.h
@@ -72,33 +72,15 @@ namespace OpenInfraPlatform {
 
 				/*! \brief Converts an \c IfcSurface to an array of segments to be rendered on screen.
 				 *
-				 * \param[in] surface		The \c IfcSurface to be converted.
-				 * \param[in] pos			The relative location of the origin of the representation's coordinate system within the geometric context.
-				 * \param[out] itemData		A pointer to be filled with the relevant data.
-				 *
-				 * \note Calls the other overload FaceConverterT::convertIfcSurface with corresponding input parameters. Adds their result to itemData.
+				 * \param[in]	surface		The \c IfcSurface to be converted.
+				 * \param[in]	pos			The relative location of the origin of the representation's coordinate system within the geometric context.
+				 * \param[out]	itemData	A pointer to be filled with the relevant data of the triangulated  surface (\c PolyhedronData).
 				 */
 				void convertIfcSurface(
 					const EXPRESSReference<typename IfcEntityTypesT::IfcSurface>& surface,
 					const carve::math::Matrix& pos,
 					std::shared_ptr<ItemData>& itemData
 				) const throw(...)
-				{
-					std::shared_ptr<carve::input::PolylineSetData> polyline =
-						convertIfcSurface(surface, pos);
-					if (polyline->getVertexCount() > 1) 
-						itemData->polylines.push_back(polyline);
-				}
-
-				/*! \brief Converts \c IfcSurface to a according to subtype.
-					\param		surface	\c IfcSurface entity to be interpreted.
-					\param		pos
-					\return		polylineData
-					\note		The \c IfcSurface subtypes are \c IfcBoundedSurface,  \c IfcElementarySurface and  \c IfcSweptSurface.
-				*/
-
-				std::shared_ptr<carve::input::PolylineSetData> convertIfcSurface(const EXPRESSReference<typename IfcEntityTypesT::IfcSurface>& surface,
-					const carve::math::Matrix& pos) const throw(...) 
 				{
 					if (surface.expired()) {
 						throw oip::ReferenceExpiredException(surface);
@@ -107,21 +89,21 @@ namespace OpenInfraPlatform {
 					if (surface.isOfType<typename IfcEntityTypesT::IfcBoundedSurface>()) 
 					{
 						return convertIfcBoundedSurface(surface.as<typename IfcEntityTypesT::IfcBoundedSurface>(),
-							pos);
+							pos, itemData);
 					}
 
 					// (2/3) IfcElementarySurface SUBTYPE of IfcSurface
 					if (surface.isOfType<typename IfcEntityTypesT::IfcElementarySurface>()) 
 					{
 						return convertIfcElementarySurface(surface.as<typename IfcEntityTypesT::IfcElementarySurface>(),
-							pos);
+							pos, itemData);
 					}
 
 					// (3/3) IfcSweptSurface SUBTYPE of IfcSurface
 					if (surface.isOfType<typename IfcEntityTypesT::IfcSweptSurface>())  
 					{
 						return convertIfcSweptSurface(surface.as<typename IfcEntityTypesT::IfcSweptSurface>(),
-							pos);
+							pos, itemData);
 					}
 
 					// the rest we do not support
@@ -132,15 +114,16 @@ namespace OpenInfraPlatform {
 				// IfcBoundedSurface
 				//--------------------------------------------------------------------------------------------
 
-						/*! \brief  Converts \c IfcBoundedSurface by calling the conversion function according to subtype.
-						\param		surface	\c IfcBoundedSurface entity to be interpreted.
-						\param		pos
-						\return		polylineData
[…]
