### 2021-06-11 840bd7d
Added construction of polyhedronData in convertIfcPlane ... similar to the approach in convertIfcBSplineSurface (previous commit 0787fe7). Before this change, only a polyline was constructed in convertIfcPlane; however, it's a subtype of IfcSurface, thus it will probably need an actual surface (represented by PolyhedronData).
