### 2021-06-11 c65e7ea
Removed a deprecated comment. It was resolved in a previous commit (c627ad8)
