### 2021-06-11 9e3594f
Deleted code which is not needed ... because the 'abstract' definition of IfcBSplineCurve, there can't exist an instance which matches this condition. (Since the refactoring of the Spline-Curves last summer, now I understood the concept of abstract types / classes better.)
