### 2021-06-11 bf0bdc5
Fix of reference in for-each loop
