### 2021-06-11 6253f17
Removed repetitive code by default parameter. Calculation of normal and rational B-Spline curves was separated in two almost similar sets of functions. Now, it's combined into one set of functions; the case is distributed by a default-parameter-definition of the weights vector. This change is applied to the surfaces in the same way.
