+++ Core/src/IfcGeometryConverter/SplineUtilities.h
@@ -45,13 +45,19 @@ namespace OpenInfraPlatform
 
 				/*! \brief Computes the curve points of the B-Spline.
 				 *
-				 * All information has to be loaded from an \c IfcBSplineCurveWithKnots entity before calling this function.
+				 * This functions computes a B-Spline curve or a rational B-Spline curve, dependent on the existance of a weight vector \c weightsData.
+				 * If the function will be called without a weights vector, a B-Spline curve will be calculated.
+				 * If the function will be called with a weights vector, a rational B-Spline curve will be calculated.
+				 * All information has to be loaded from an \c IfcBSplineCurveWithKnots or \c IfcRationalBSplineCurveWithKnots before calling this function.
 				 *
 				 * \param[in]	order				Order of the B-Spline or rather the basis functions ( =degree+1 )
 				 * \param[in]	knotArray			The array / vector of knots, the function \c loadKnotArray gives this vector.
 				 * \param[in]	controlPoints		The vector of the B-Spline control points.
 				 * \param[in]	numCurvePoints		The number of curve points where the curve c(t) has to be evaluated.
 				 * \param[in]	accuracy			Accuracy which is technically needed in the calculation.
+				 * \param[in]	weightsData			The vector with the weight values per control point, the function \c loadWeightsData gives this vector.
+				 *
+				 * \note	\c weightsData is defined as default parameter: A function call without a weights vector triggers a (usual) B-Spline curve, a call with a weights vector triggers a rational B-Spline curve
 				 *
 				 * \return		The array of curve points, which can be rendered in a viewport after correction by \c GeomUtils::appendPointsToCurve.
 				 */
@@ -61,7 +67,8 @@ namespace OpenInfraPlatform
 					const std::vector<double>& knotArray,
 					const std::vector<carve::geom::vector<3>>& controlPoints,
 					const uint32_t& numCurvePoints,
-					const double& accuracy) throw(...)
+					const double& accuracy,
+					const std::vector<double>& weightsData = { } ) throw(...)
 				{
 					// The following parameters corresponds to the parameter t of a curve c(t)
 					double knotStart;
@@ -79,54 +86,7 @@ namespace OpenInfraPlatform
 						// the for-loop iterates over all values t of the parameter vector
 						if (i == numCurvePoints - 1) { t = knotEnd - accuracy; }
 
-						curvePoints.push_back(computePointOfBSpline(order, t, controlPoints, knotArray));
-
-						// go to next representative t of the parameter vector (= increase its value)
-						t += step;
-					}
-
-					return curvePoints;
-				}
-
-				/*! \brief Computes the curve points of the rational B-Spline.
-				 *
-				 * All information has to be loaded from an \c IfcRationalBSplineCurveWithKnots entity before calling this function.
-				 *
-				 * \param[in]	order				Order of the B-Spline or rather the basis functions ( =degree+1 )
-				 * \param[in]	knotArray			The array / vector of knots, the function \c loadKnotArray gives this vector.
-				 * \param[in]	controlPoints		The vector of the B-Spline control points.
-				 * \param[in]	weightsData			The vector with the weight values per control point, the function \c loadWeightsData gives this vector.
-				 * \param[in]	numCurvePoints		The number of curve points where the curve c(t) has to be evaluated.
-				 * \param[in]	accuracy			Accuracy which is technically needed in the calculation.
-				 *
-				 * \return		The array of curve points, which can be rendered in a viewport after correction by \c GeomUtils::appendPointsToCurve.
-				 */
[…]
