+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -148,8 +148,8 @@ namespace OpenInfraPlatform {
 					//		Dim : IfcDimensionCount: = IfcCurveDim(SELF);
 					//	END_ENTITY;
 					// **************************************************************************************************************************
-					// IfcBlossCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
-//#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
+					// IfcBlossCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
+//#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
 //					if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcBlossCurve>())
 //					{
 //						return convertIfcBlossCurve(ifcCurve.as<typename IfcEntityTypesT::IfcBlossCurve>(),
@@ -164,8 +164,8 @@ namespace OpenInfraPlatform {
 							targetVec, segmentStartPoints);
 					}
 
-					// IfcClothoid SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
-#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
+					// IfcClothoid SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
+#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
 					else if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcClothoid>())
 					{
 						return convertIfcClothoid(ifcCurve.as<typename IfcEntityTypesT::IfcClothoid>(), 
@@ -201,8 +201,8 @@ namespace OpenInfraPlatform {
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					}
 
-//					// IfcSeriesParameterCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
-//#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
+//					// IfcSeriesParameterCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
+//#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
 //					else if (ifcCurve.isOfType<typename IfcEntityTypesT::IfcSeriesParameterCurve>())
 //					{
 //						return convertIfcSeriesParameterCurve(ifcCurve.as<typename IfcEntityTypesT::IfcSeriesParameterCurve>(),
@@ -278,8 +278,8 @@ namespace OpenInfraPlatform {
 					}
 				} //end convertIfcCurve2D
 
-				// IfcBlossCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
-//#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
+				// IfcBlossCurve SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
+//#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
 //				/**********************************************************************************************/
 //				/*! \brief Converts an \c IfcBlossCurve to a tesselated curve.
 //				* \param[in] blossCurve				A pointer to data from \c IfcBlossCurve.
@@ -332,7 +332,7 @@ namespace OpenInfraPlatform {
 					//		SUBTYPE OF(IfcCurve);
 					//	END_ENTITY;
 					// **************************************************************************************************************************
-					// IfcAlignmentCurve SUBTYPE OF IfcBoundedCurve (Deprecated starting IFC4X3_RC3)
[…]
