### 2021-06-13 40fe375
Removed doxy-comments from GeomUtils.cpp
