+++ Core/src/IfcGeometryConverter/GeomUtils.cpp
@@ -665,17 +665,8 @@ bool GeomUtils::isPointOnLineSegment( double& target_lambda,
 }
 
 /**********************************************************************************************/
-/*! \brief Calculates distances between start points of the segment and their intersection.
-* \param[in] initialPointSeg1		Initial point of the first intersecting segment.
-* \param[in] terminalPointSeg1		Terminal point of the first intersecting segment.
-* \param[in] initialPointSeg2		Initial point of the second intersecting segment.
-* \param[in] terminalPointSeg2		Terminal point of the second intersecting segment.
-* \param[out] distToIntesection1	Distance from initialPointSeg1 to the segment intresection, which is calculated as section of the first segment.
-* \param[out] distToIntesection2	Distance from initialPointSeg2 to the segment intresection, which is calculated as section of the second segment.
 
-* \return							Returns true, if lines are intersecting with each other. False otherwise.
-*/
-bool LineToLineIntersectionHelper(const carve::geom::vector<2>& initialPointSeg1,
+bool GeomUtils::LineToLineIntersectionHelper(const carve::geom::vector<2>& initialPointSeg1,
 									const carve::geom::vector<2>& terminalPointSeg1,
 									const carve::geom::vector<2>& initialPointSeg2,
 									const carve::geom::vector<2>& terminalPointSeg2, 
@@ -701,14 +692,7 @@ bool LineToLineIntersectionHelper(const carve::geom::vector<2>& initialPointSeg1
 }
 
 /**********************************************************************************************/
-/*! \brief Calculates coordinates of the intersection point.
-* \param[in] initialPointSeg1		Initial point of the first intersecting segment.
-* \param[in] terminalPointSeg1		Terminal point of the first intersecting segment.
-* \param[in] initialPointSeg2		Initial point of the second intersecting segment, which describes the direction of the line.
-* \param[in] terminalPointSeg2		Terminal point of the second intersecting segment, which describes the direction of the line.
-* \param[out] intersectionPoint		Coordinates of the intersection point between first line segment and second line.
-* \return							Returns true, if lines are intersecting with each other. False otherwise.
-*/
+
 bool GeomUtils::LineSegmentToLineIntersection(const carve::geom::vector<2>& initialPointSeg1,
 												const carve::geom::vector<2>& terminalPointSeg1, 
 												const carve::geom::vector<2>& initialPointSeg2,
@@ -728,14 +712,7 @@ bool GeomUtils::LineSegmentToLineIntersection(const carve::geom::vector<2>& init
 }
 
 /**********************************************************************************************/
-/*! \brief Calculates coordinates of the intersection point.
-* \param[in] initialPointSeg1		Initial point of the first intersecting segment, which describes the direction of the line.
-* \param[in] terminalPointSeg1		Terminal point of the first intersecting segment, which describes the direction of the line.
-* \param[in] initialPointSeg2		Initial point of the second intersecting segment, which describes the direction of the line.
-* \param[in] terminalPointSeg2		Terminal point of the second iintersecting segment, which describes the direction of the line.
-* \param[out] intersectionPoint		Coordinates of the intersection point between first line and second line.
-* \return							Returns true, if lines are intersecting with each other. False otherwise.
-*/
+
 bool GeomUtils::LineSegmentToLineSegmentIntersection(const carve::geom::vector<2>& initialPointSeg1,
 														const carve::geom::vector<2>& terminalPointSeg1,
 														const carve::geom::vector<2>& initialPointSeg2,
