### 2021-06-13 e82ca1a
Added Doxy-comment to LineToLineIntersectionHelper and corrected doxy-comments for Intersection functions
