### 2021-06-15 897d9c9
Added Unit Test for IfcBsplineSurfaceWithKnots. ifcBSplineSurfaceWithKnots (entities 100 and 52 to 67) is extracted from the official example 'basin-advanced-brep.ifc'
