### 2021-06-16 0fad7a4
Fixed to const auto& in for-each loop
