### 2021-06-18 e0dac45
Added new functions for Taylor series integral Fixed some mistakes
