+++ Core/src/IfcGeometryConverter/GeomUtils.cpp
@@ -1219,11 +1219,300 @@ bool GeomUtils::checkMeshSet( const carve::mesh::MeshSet<3>* mesh_set,
 	}
 #endif
 
-	if( err.tellp() > 0 )
+	if (err.tellp() > 0)
 	{
 		throw oip::InconsistentModellingException("GeomUtils::checkMeshSet(): MeshSet of resulting mesh has problems: " + err.str());
 	}
 	return true;
 }
+
+	/**********************************************************************************************/
+
+	/*!Calculate recursive multiplication for calculation Taylor series
+	* \param[in]  value                  value of Taylor series
+	* \param[in]  polynomialConstants    polynomial constants for a polyomial term - a,b,c,d
+	* \param[in]  polynomialConstantCnt  type of polinomial term - cubic, linear 
+	* \param[in]  myArray
+	* \param[in]  myArrayParentIndex
+	* \param[in]  i_n
+	* \param[in]  n
+	*
+	* \return
+	*/
+	void GeomUtils::recursiveMultiplicationTaylor(
+			double	value,
+			std::vector<double> polynomialConstants,
+			size_t	polynomialConstantCnt,
+			std::vector<double> myArray,
+			size_t	myArrayParentIndex,
+			size_t	i_n,
+			size_t	n
+		)
+	{
+		assert(value);
+		if (i_n < n - 1) {
+			assert(polynomialConstantCnt && polynomialConstants[0] == 0.);
+			for (size_t i = 1; i < polynomialConstantCnt; i++) {
+				if (polynomialConstants[i]) {
+					GeomUtils::recursiveMultiplicationTaylor(
+						value * polynomialConstants[i],
+						polynomialConstants, polynomialConstantCnt,
+						myArray, myArrayParentIndex + i,
+						i_n + 1, n
+					);
+				}
+			}
+		}
+		else {
+			assert(i_n == n - 1);
+			for (size_t i = 1; i < polynomialConstantCnt; i++) {
+				if (polynomialConstants[i]) {
+					assert(myArrayParentIndex + i >= n && myArrayParentIndex + i < polynomialConstantCnt* n + 1);
+					myArray[myArrayParentIndex + i] += value * polynomialConstants[i];
+				}
+			}
+		}
+	}
+
+	/**********************************************************************************************/
+
+	/*!Calculate the Taylor series 
+	* \param[in]  n                      number of iteration points
+	* \param[in]  polynomialConstants    polynomial constants for a polyomial term - a,b,c,d
+	* \param[in]  polynomialConstantCnt  type of polinomial term - cubic, linear
+	* \param[in]  s                      the length of the curve between two points 
+	*
+	* \return
+	*/
+	double	GeomUtils::computeIntegralTaylorSeriesExpansionElement(
+		size_t	n,
+		std::vector<double> polynomialConstants,
+		size_t	polynomialConstantCnt,
+		double	s
+	)
+	{
+		//
+		//	pC = polynomialConstants
+		//	c = polynomialConstantCnt
+		// 	x =  indeterminate
+		// 	   =>	pC[c-1] * x^(c-1) + pC[c-2] * x^(c-2) + ... + pC[1] * x^(1) + pC[0] * x^(0) where pC[0] == 0
+		//
+		assert(polynomialConstantCnt && polynomialConstants[0] == 0.);
+
+		if (n) {
+			std::vector<double>	myArray(polynomialConstantCnt * n + 1);
+			GeomUtils::recursiveMultiplicationTaylor(
+				1.,
+				polynomialConstants, polynomialConstantCnt,
+				myArray, 0,
+				0, n
+			);
+
+			double	value = 0., factor = 1.;
[…]
