### 2021-06-18 62478ac
Added new functions for Taylor series integral Fixed some mistakes
