### 2021-06-19 360f7bf
Implementation of SolidModelConverter::convertIfcAdvancedBrep
