### 2021-06-19 6c8a245
Added comments to SolidModelConverter::convertIfcClosedShell
