### 2021-06-19 346e611
Adjusted doxygen comment of FaceConverter::convertIfcFaceList
