+++ Core/src/IfcGeometryConverter/FaceConverter.h
@@ -729,11 +729,10 @@ namespace OpenInfraPlatform {
 				// convertIfcFaceList, convertIfcFace, convert3DPointsTo2D, triangulateFace, convertIfcCartesianPointVectorVector
 				//------------------------------------------------------------------------------------------------------------
 
-						/*! \brief  Converts \c IfcFace to a polygon and adds it to the carve PolyhedronData vector.
-						\param	faces \c IfcFace entity to be interpreted.
-						\param	pos
-						\param	itemData
-						\return polygon carve polygon
+						/*! \brief  Converts a list of \c IfcFace -s to a polygon and adds it to the carve PolyhedronData vector.
+						\param[in]	faces		List \c IfcFace entity to be interpreted.
+						\param[in]	pos			The relative location of the origin of the representation's coordinate system within the geometric context.
+						\param[out]	itemData	Polygon carve polygon.
 						\note The \c IfcFaceList can be an open or closed shell.
 						*/
 
