### 2021-06-19 032606c
Changed to for-each-loop in FaceConverter::convertIfcFaceList
