### 2021-06-20 fb684da
Removed bool conversionFailed in FaceConverter::convertIfcFace. The defined boolean value does not change inside the function convertIfcFace. Probably, its original purpose was to indicate an error inside the function. In the current code of convertIfcFace, all inspected error-cases are handled by the call of 'throw opi::<exception>'.
