### 2021-06-20 4cfa5fa
Removed variable faceId in convertIfcFace, because isn't used anymore
