### 2021-06-20 2f65034
Encapsulated the loading of IfcFaceBound-list into separate function
