### 2021-06-20 affec0b
Modified convertIfcFaceBoundList to avoid copy of entire vector
