### 2021-06-20 5b4d832
Encapsulated the loop obtainment of IfcFaceBound into separate function
