### 2021-06-20 7a755a9
Moved call of convertIfcFaceBound into convertIfcFaceBoundList
