### 2021-06-20 ed56be2
Moved external for-loop counter next to its for-loop
