### 2021-06-20 cdb1b5e
Removed not needed variable. vertex3D and v were both declared as const carve::geom::vector<3>. By hovering with the mouse, both variables were identified as const crave::geom3d::Vector &. Thus, I can't see any 'global transformation' like promised in the comment.
