### 2021-06-21 3ae4b05
Fixed semi-bug of double triangulation of triangular ifcFaceBound. See following github comment for more details on this.
