### 2021-06-21 2383522
Removed unused variable mergedIndices in convertIfcFace. Of cause, some information was saved in the variable (line 833), but it was never used / called in the entire function.
