### 2021-06-21 967024d
Removed unused external for-loop counter
