### 2021-06-21 d8c7d76
Encapsulated process of addTriangleToPolygedronData into own function
