### 2021-06-21 18f39ec
Restructured convertIfcFace to decide between simple and general case: simple case is only on triangle without holes, general case is arbitrary number of vertices and holes
