### 2021-06-21 8e12ed5
Encapsulated process addArbitraryFaceToPolyhedronData into own function
