### 2021-06-21 a995def
Restructure if-else in convertIfcFace because of encapsulated function
