### 2021-06-21 #3061 @pjanck
Hi @hugoledoux - I believe we'll take the summer to improve on the few open points as mentioned by @CBenghi and @aothms - you can put this on pause. Apologies for taking this long, but it is quite a package :)
