+++ UserInterface/Dialogues/IfcTreeModel.cpp
@@ -21,87 +21,90 @@
 #include <cstdlib>
 #include <memory>
 
+#include <QStringList>
+#include <QString>
 
-OpenInfraPlatform::UserInterface::IfcTreeModel::IfcTreeModel(OpenInfraPlatform::EarlyBinding::EXPRESSModel *expressModel, QObject *parent)
-	: QAbstractItemModel(parent)
-{
-	//rootItem_ = std::make_shared<IfcTreeItem>(expressModel->entities.find(1)->second, nullptr); //make ifcProject the root item 
-	//std::string text = "Title";
-	rootItem_ = new IfcTreeItem();
-	setupModelData(expressModel, rootItem_);
-}
 
-OpenInfraPlatform::UserInterface::IfcTreeModel::~IfcTreeModel()
-{
-	delete rootItem_;
-}
-
-QModelIndex OpenInfraPlatform::UserInterface::IfcTreeModel::index(int row, int column, const QModelIndex &parent) const
-{
-	if (!hasIndex(row, column, parent))
-		return QModelIndex();
-
-	IfcTreeItem* parentItem;
-
-	if (!parent.isValid())
-		parentItem = rootItem_;
-		//return this->createIndex(row, column, nullptr);
-	else
-		parentItem = static_cast<IfcTreeItem*>(parent.internalPointer());
-
-	IfcTreeItem *childItem = parentItem->child(row);
-	if (childItem)
-	{
-		auto n = createIndex(row, column, childItem);
-		return n;
-	}
-		//return createIndex(row, column, childItem);
-	return QModelIndex();
-}
-
-QVariant OpenInfraPlatform::UserInterface::IfcTreeModel::headerData(int section, Qt::Orientation orientation,
-	int role) const
-{
-	return rootItem_->data(section);
-}
-
-QModelIndex OpenInfraPlatform::UserInterface::IfcTreeModel::parent(const QModelIndex &index) const
-{
-	if (!index.isValid())
-		return QModelIndex();
-
-	IfcTreeItem *child = static_cast<IfcTreeItem*>(index.internalPointer());
-	IfcTreeItem *parentItem = child->parentItem();
-
-	if (parentItem == rootItem_)
-		return QModelIndex(); 
-
-	return createIndex(parentItem->row(), 0, parentItem);
-}
-
-int OpenInfraPlatform::UserInterface::IfcTreeModel::rowCount(const QModelIndex &parent) const
-{
-	IfcTreeItem *parentItem;
-
-	if (parent.column() > 0)
-		return 0;
-
-	if (!parent.isValid())
-		parentItem = rootItem_;
-	else
-		parentItem = static_cast<IfcTreeItem*>(parent.internalPointer());
-	
-	int nrOfRows = parentItem->childCount();
-	return nrOfRows;
-}
-
-int OpenInfraPlatform::UserInterface::IfcTreeModel::columnCount(const QModelIndex &parent) const
-{
-	if (parent.isValid())
-		return static_cast<IfcTreeItem*>(parent.internalPointer())->columnCount();
-	int nrOfColumns = rootItem_->columnCount();
-	return nrOfColumns;
-}
+//OpenInfraPlatform::UserInterface::IfcTreeModel::IfcTreeModel(OpenInfraPlatform::EarlyBinding::EXPRESSModel *expressModel, QObject *parent)
+//	: QAbstractItemModel(parent)
+//{
+//	//rootItem_ = std::make_shared<IfcTreeItem>(expressModel->entities.find(1)->second, nullptr); //make ifcProject the root item 
+//	//std::string text = "Title";
+//	rootItem_ = new IfcTreeItem();
+//	setupModelData(expressModel, rootItem_);
+//}
+//
[…]
