+++ UserInterface/Dialogues/IfcTreeItem.cpp
@@ -23,60 +23,60 @@
 #include "visit_struct\visit_struct.hpp"
 
 
-OpenInfraPlatform::UserInterface::IfcTreeItem::IfcTreeItem(OpenInfraPlatform::EarlyBinding::EXPRESSEntity *data, IfcTreeItem *parentItem)
-{
-	data_ = data;
-	parentItem_ = parentItem;
-	//itemData_.push_back(data_->getStepLine());
-	itemData_.push_back(data_->classname());
-}
-
-OpenInfraPlatform::UserInterface::IfcTreeItem::IfcTreeItem()
-{
-	data_ = nullptr;
-	parentItem_ = nullptr;
-	std::string text = "Title";
-	itemData_.push_back(text);
-}
-
-OpenInfraPlatform::UserInterface::IfcTreeItem::~IfcTreeItem()
-{
-	//childItems_.clear();
-	qDeleteAll(childItems_);
-}
-
-void OpenInfraPlatform::UserInterface::IfcTreeItem::appendchild(IfcTreeItem *child)
-{
-	childItems_.append(child);
-}
-
-OpenInfraPlatform::UserInterface::IfcTreeItem* OpenInfraPlatform::UserInterface::IfcTreeItem::child(int row)
-{
-	if (row < 0 || row >= childItems_.size())
-		throw oip::UnhandledException("Child index out of range (IfcTreeItem::child)");
-
-	return childItems_.at(row);
-}
-
-int OpenInfraPlatform::UserInterface::IfcTreeItem::childCount() const
-{
-	int nrOfChilds = childItems_.count();
-	return nrOfChilds;
-}
-
-int OpenInfraPlatform::UserInterface::IfcTreeItem::columnCount() const
-{
-	int nrOfColumns = itemData_.size(); //size right here?
-	return nrOfColumns;
-}
-
-QVariant OpenInfraPlatform::UserInterface::IfcTreeItem::data(int column) const
-{
-	if (column < 0 || column >= itemData_.size())
-		return QVariant();
-	//QString text = QString::fromStdString(itemData_.at(column));
-	return QString("text");
-}
+//OpenInfraPlatform::UserInterface::IfcTreeItem::IfcTreeItem(OpenInfraPlatform::EarlyBinding::EXPRESSEntity *data, IfcTreeItem *parentItem)
+//{
+//	data_ = data;
+//	parentItem_ = parentItem;
+//	//itemData_.push_back(data_->getStepLine());
+//	itemData_.push_back(data_->classname());
+//}
+//
+//OpenInfraPlatform::UserInterface::IfcTreeItem::IfcTreeItem()
+//{
+//	data_ = nullptr;
+//	parentItem_ = nullptr;
+//	std::string text = "Title";
+//	itemData_.push_back(text);
+//}
+//
+//OpenInfraPlatform::UserInterface::IfcTreeItem::~IfcTreeItem()
+//{
+//	//childItems_.clear();
+//	qDeleteAll(childItems_);
+//}
+//
+//void OpenInfraPlatform::UserInterface::IfcTreeItem::appendchild(IfcTreeItem *child)
+//{
+//	childItems_.append(child);
+//}
+//
+//OpenInfraPlatform::UserInterface::IfcTreeItem* OpenInfraPlatform::UserInterface::IfcTreeItem::child(int row)
+//{
+//	if (row < 0 || row >= childItems_.size())
+//		throw oip::UnhandledException("Child index out of range (IfcTreeItem::child)");
+//
+//	return childItems_.at(row);
+//}
+//
+//int OpenInfraPlatform::UserInterface::IfcTreeItem::childCount() const
+//{
+//	int nrOfChilds = childItems_.count();
+//	return nrOfChilds;
[…]
