### 2021-07-01 9df4951
Adapt to the newest version of ifcImporter.h
