### 2021-07-01 15d3a2f
Letting the treeviewer show the list of all IfcEntities within the loaded file
