### 2021-07-01 5a28961
Clean up IfcTreeItem
