### 2021-07-01 a7d19f5
Clean up IfcTreeDialog
