### 2021-07-01 007fa6f
Clean up IfcTreeModel
