+++ UserInterface/Dialogues/IfcTreeModel.h
@@ -36,13 +36,11 @@ namespace OpenInfraPlatform
 			Q_OBJECT;
 
 		public:
-			//explicit IfcTreeModel(const QString &data, QObject *parent = nullptr);
 			explicit IfcTreeModel(OpenInfraPlatform::EarlyBinding::EXPRESSModel *expressModel, QObject *parent = nullptr);
 			~IfcTreeModel();
 			
 			//override from QAbstractItemModel
 			QVariant data(const QModelIndex &index, int role) const override;
-			//Qt::ItemFlags flags(const QModelIndex &index) const override;
 			QVariant headerData(int section, Qt::Orientation orientation,
 				int role = Qt::DisplayRole) const override;
 			QModelIndex index(int row, int column,
@@ -52,12 +50,10 @@ namespace OpenInfraPlatform
 			int columnCount(const QModelIndex &parent = QModelIndex()) const override;
 
 		private:
-			//void setupModelData(const QStringList &lines, IfcTreeItem *parent);
 			void setupModelData(const OpenInfraPlatform::EarlyBinding::EXPRESSModel *expressModel, IfcTreeItem *parent);
 
 			IfcTreeItem *rootItem;
 		};
-
 	}
 }
 
