+++ UserInterface/Dialogues/IfcTreeModel.cpp
@@ -33,145 +33,6 @@ OpenInfraPlatform::UserInterface::IfcTreeModel::IfcTreeModel(OpenInfraPlatform::
 	rootItem = new IfcTreeItem(rootData);
 	setupModelData(expressModel, rootItem);
 }
-//
-//OpenInfraPlatform::UserInterface::IfcTreeModel::~IfcTreeModel()
-//{
-//	delete rootItem_;
-//}
-//
-//QModelIndex OpenInfraPlatform::UserInterface::IfcTreeModel::index(int row, int column, const QModelIndex &parent) const
-//{
-//	if (!hasIndex(row, column, parent))
-//		return QModelIndex();
-//
-//	IfcTreeItem* parentItem;
-//
-//	if (!parent.isValid())
-//		parentItem = rootItem_;
-//		//return this->createIndex(row, column, nullptr);
-//	else
-//		parentItem = static_cast<IfcTreeItem*>(parent.internalPointer());
-//
-//	IfcTreeItem *childItem = parentItem->child(row);
-//	if (childItem)
-//	{
-//		auto n = createIndex(row, column, childItem);
-//		return n;
-//	}
-//		//return createIndex(row, column, childItem);
-//	return QModelIndex();
-//}
-//
-//QVariant OpenInfraPlatform::UserInterface::IfcTreeModel::headerData(int section, Qt::Orientation orientation,
-//	int role) const
-//{
-//	return rootItem_->data(section);
-//}
-//
-//QModelIndex OpenInfraPlatform::UserInterface::IfcTreeModel::parent(const QModelIndex &index) const
-//{
-//	if (!index.isValid())
-//		return QModelIndex();
-//
-//	IfcTreeItem *child = static_cast<IfcTreeItem*>(index.internalPointer());
-//	IfcTreeItem *parentItem = child->parentItem();
-//
-//	if (parentItem == rootItem_)
-//		return QModelIndex(); 
-//
-//	return createIndex(parentItem->row(), 0, parentItem);
-//}
-//
-//int OpenInfraPlatform::UserInterface::IfcTreeModel::rowCount(const QModelIndex &parent) const
-//{
-//	IfcTreeItem *parentItem;
-//
-//	if (parent.column() > 0)
-//		return 0;
-//
-//	if (!parent.isValid())
-//		parentItem = rootItem_;
-//	else
-//		parentItem = static_cast<IfcTreeItem*>(parent.internalPointer());
-//	
-//	int nrOfRows = parentItem->childCount();
-//	return nrOfRows;
-//}
-//
-//int OpenInfraPlatform::UserInterface::IfcTreeModel::columnCount(const QModelIndex &parent) const
-//{
-//	if (parent.isValid())
-//		return static_cast<IfcTreeItem*>(parent.internalPointer())->columnCount();
-//	int nrOfColumns = rootItem_->columnCount();
-//	return nrOfColumns;
-//}
-
-//int OpenInfraPlatform::UserInterface::IfcTreeModel::columnCount(const QModelIndex &parent) const
-//{
-//	if (!parent.isValid())
-//		return 1; //correct for testing?
-//	else
-//		return 3;
-//}
-
-//Qt::ItemFlags OpenInfraPlatform::UserInterface::IfcTreeModel::flags(const QModelIndex &index) const
-//{
-//	if (!index.isValid())
-//		return Qt::NoItemFlags;
-//
-//	return QAbstractItemModel::flags(index);
-//}
-//
-//QVariant OpenInfraPlatform::UserInterface::IfcTreeModel::data(const QModelIndex &index, int role) const
-//{
-//	if (!index.isValid())
-//		return QVariant();
-//
-//	IfcTreeItem *item = static_cast<IfcTreeItem*>(index.internalPointer());
-//
[…]
