+++ UserInterface/Dialogues/IfcTreeItem.h
@@ -43,6 +43,9 @@ namespace OpenInfraPlatform {
 			IfcTreeItem *parentItem();
 			QString getIfcClassName() const;
 
+			//void setItemData(QList<QVariant> itemData);
+			//void createChildren();
+
 		private:
 			QList<IfcTreeItem*> childItems_;
 			OpenInfraPlatform::EarlyBinding::EXPRESSEntity *data_;
@@ -50,10 +53,184 @@ namespace OpenInfraPlatform {
 			IfcTreeItem *parentItem_;
 
 			//struct getAttributeDescription;
+
+			//std::shared_ptr<OpenInfraPlatform::EarlyBinding::EXPRESSObject> m_managedData = nullptr;
+
+			////Helper struct which parses the attributes of a derived entity which are IfcAlignment1x1Types and hold flat values.
+			//struct parseType {
+
+			//	void operator()(const char* name, std::string string)
+			//	{
+			//		std::shared_ptr<OpenInfraPlatform::EarlyBinding::EXPRESSObject> ptr = nullptr;
+			//		TreeItem* child = new TreeItem(ptr, thisPtr);
+			//		QList<QVariant> itemData;
+			//		itemData << QVariant(name) << QVariant(string.data()) << QVariant("std::string");
+			//		child->setItemData(itemData);
+			//		thisPtr->appendChild(child);
+			//	}
+
+			//	//Function operator() which creates a statement from a boolean value.
+			//	void operator()(const char* name, bool value)
+			//	{
+			//		std::shared_ptr<OpenInfraPlatform::EarlyBinding::EXPRESSObject> ptr = nullptr;
+			//		TreeItem* child = new TreeItem(ptr, thisPtr);
+			//		QList<QVariant> itemData;
+			//		itemData << QVariant(name) << QVariant(value) << QVariant("bool");
+			//		child->setItemData(itemData);
+			//		thisPtr->appendChild(child);
+			//	}
+
+			//	void operator()(const char* name, std::shared_ptr<OpenInfraPlatform::EarlyBinding::EXPRESSEntity> value)
+			//	{
+			//		TreeItem* child = new TreeItem(std::static_pointer_cast<OpenInfraPlatform::EarlyBinding::EXPRESSObject>(value), thisPtr);
+			//		QList<QVariant> itemData;
+			//		itemData << QVariant(name) << QVariant(value ? value->getId() : -1); //<< QVariant(value ? value->classname() : "nullptr");
+			//		child->setItemData(itemData);
+			//		thisPtr->appendChild(child);
+			//	}
+
+			//	//TODO: Get value stored in type
+			//	void operator()(const char* name, std::shared_ptr<OpenInfraPlatform::EarlyBinding::EXPRESSType> value)
+			//	{
+			//		TreeItem* child = new TreeItem(std::static_pointer_cast<OpenInfraPlatform::EarlyBinding::EXPRESSObject>(value), thisPtr);
+			//		QList<QVariant> itemData;
+
+			//		//std::stringstream ss;
+			//		if (value)
+			//			//readStepData(std::string& arg)
+			//			//value->getStepData(ss);
+
+
+			//		//auto classname = value->classname(); //Gibt z.B. IfcLabel o.ä. zurück
+			//		//static_cast<classname>(value);
+			//		//auto castedValue = value.m_value; //gibt m_value zurück aus IfcLabel, also den eigentlichen Wert, der als String o.ö. gespeichert ist
+			//		//auto castedType = typeid(castedValue).name(); //gibt den Typ von m_value zurück, also z.B. String
+			//		//
+			//		//if (castedType == "string" || "char")
[…]
