### 2021-07-15 b6148c3
Removed valid-check of EXPRESSReference entity (weak pointer). This check is the first task in the function convertIfcFace which is called at this place.
