+++ Core/src/IfcGeometryConverter/GeomUtils.cpp
@@ -1275,45 +1275,41 @@ bool GeomUtils::checkMeshSet( const carve::mesh::MeshSet<3>* mesh_set,
 
 	/*!Calculate recursive multiplication for calculation Taylor series
 	* \param[in]  value                  value of Taylor series
-	* \param[in]  polynomialConstants    polynomial constants for a polyomial term - a,b,c,d
-	* \param[in]  polynomialConstantCnt  type of polinomial term - cubic, linear 
-	* \param[in]  myArray
-	* \param[in]  myArrayParentIndex
-	* \param[in]  i_n
-	* \param[in]  n
+	* \param[in]  polynomialConstants    polynomial constants for a polyomial term - a,b,c,d 
+	* \param[in]  myArray                matrix with 1xn number of iteration points
+	* \param[in]  i_n                    aktuôä itaration point
+	* \param[in]  n                      number of itaration points
 	*
 	* \return
 	*/
 	void GeomUtils::recursiveMultiplicationTaylor(
-			double	value,
-			std::vector<double> polynomialConstants,
-			size_t	polynomialConstantCnt,
-			std::vector<double> myArray,
-			size_t	myArrayParentIndex,
-			size_t	i_n,
-			size_t	n
+			const double	value,
+			const std::vector<double>& polynomialConstants,
+			std::vector<double>& myArray,
+			const int	i_n,
+			const int	n
 		)
 	{
-		assert(value);
+		if (!value) throw oip::InconsistentGeometryException("...");
 		if (i_n < n - 1) {
-			assert(polynomialConstantCnt && polynomialConstants[0] == 0.);
-			for (size_t i = 1; i < polynomialConstantCnt; i++) {
+			if(polynomialConstants.size() && polynomialConstants[0] != 0.) throw oip::InconsistentGeometryException("...");
+			for (int i = 1; i < polynomialConstants.size(); i++) {
 				if (polynomialConstants[i]) {
 					GeomUtils::recursiveMultiplicationTaylor(
 						value * polynomialConstants[i],
-						polynomialConstants, polynomialConstantCnt,
-						myArray, myArrayParentIndex + i,
+						polynomialConstants,
+						myArray,
 						i_n + 1, n
 					);
 				}
 			}
 		}
 		else {
-			assert(i_n == n - 1);
-			for (size_t i = 1; i < polynomialConstantCnt; i++) {
+			if(i_n == n - 1);
+			for (int i = 1; i < polynomialConstants.size(); i++) {
 				if (polynomialConstants[i]) {
-					assert(myArrayParentIndex + i >= n && myArrayParentIndex + i < polynomialConstantCnt* n + 1);
-					myArray[myArrayParentIndex + i] += value * polynomialConstants[i];
+					if(myArray.size() + i <= n && myArray.size() + i > polynomialConstants.size()* n + 1) throw oip::InconsistentGeometryException("...");
+					myArray[myArray.size() + i] += value * polynomialConstants[i];
 				}
 			}
 		}
@@ -1324,43 +1320,42 @@ bool GeomUtils::checkMeshSet( const carve::mesh::MeshSet<3>* mesh_set,
 	/*!Calculate the Taylor series 
 	* \param[in]  n                      number of iteration points
 	* \param[in]  polynomialConstants    polynomial constants for a polyomial term - a,b,c,d
-	* \param[in]  polynomialConstantCnt  type of polinomial term - cubic, linear
 	* \param[in]  s                      the length of the curve between two points 
 	*
 	* \return
 	*/
[…]
