### 2021-07-16 b4c24cb
Moved 'swap of outerBoundary' into a distinct function
