### 2021-07-16 d9303cc
Changed return type of convertIfcFaceBoundList
