### 2021-07-16 ff1b65e
Changed return type of convertIfcFaceBound
