### 2021-07-16 785881d
Added test of reference expired to convertIfcFaceBound. Adjusted expired test of IfcLoop.
