### 2021-07-16 72e884d
Moved integrity test about number of loop points upwards
