### 2021-07-16 3f4c077
Removed not needed return statements
