### 2021-07-19 7b0399b
Created an action for Zoom to an Object function
