### 2021-07-26 8f1f498
Add new software architecture markdown
