+++ Documentation/markdown/SoftwareArchitecture.md
@@ -1,0 +1,14 @@
+# Software Architecture
+
+*NOTE:* This document shows the overall software architecture of the Open Infra Platform and the Express Binding Generator as of the current state (July 2021). 
+
+## Content 
+
+1. [Overall Software Architecture](#Overall_architecture)
+    * [Fundaments](#Fundaments)
+    * [Modules](#Modules)
+    * [User Interface](#UI)
+2. [Express Binding Generator](#Express_binding) 
+	* [Executables](#Exec)
+	* [Commands](#Commands)
+	* [Early Binding Library](#Early_binding_lib)
