+++ Documentation/markdown/SoftwareArchitecture.md
@@ -12,3 +12,72 @@
 	* [Executables](#Exec)
 	* [Commands](#Commands)
 	* [Early Binding Library](#Early_binding_lib)
+
+## <a name="Overall_architecture"></a> Overall Software Architecture
+
+This section will give a general overview of the software architecture of the Open Infra Platform with its core software modules and how they are connected.
+Furthermore, dependies on third party libraries are stated. 
+
+![](../images/software_architecture.PNG)
+
+### <a name="Fundaments"></a> Fundaments
+Base lies the backbone for everything else that follows. Main part are the third party libraries that are linked here. Other than that we have only IModel.h Interface, GeorefMetaData.h, and BBox.h.
+
+Third Party Libraries:
+* Boost ()
+* BlueFramework ()
+* Carve ()
+* Tclap ()
+* FFTW3 ()
+* Doxygen ()
+
+### <a name="Modules"></a> Modules
+
+The modules are sperate in two big parts. First of all we have the Core section which links to the third party libraries Eigen () and Qt5 ().
+In here we have two differnet file formats that we support. First of all, we have the IFC Geometry Converter. It makes you of the Express Binding Library, which is generated by the Express Binding Generator, explained more in Detail later on. The IFC file format supports all sorts of geometric representations ranging from boundary representation and constructive solid geometry up to sweep volumes. So implicit as well as explicit geometric representations. In order to be able to render the geometries described in an IFC file the task of the IFC Geometry Converter is to convert the geometric descriptions specified into a format that can be understood by the renderer. In our case these are triangles. Since IFC is updating constantly (adding and changing IfcRepresenationItems) we also are always trying to keep up to date. Unfortunately not all IfcRepresentationItems are supported at the moment. However, the OIP already supports many of them. To keep track of what is supported you can take a look at [List of supported IFC Representations](./SupportedIFCrepresentations.md). The other part that is also part of Core is the OFF (Object File Format) Geometry Converter. This format is used to specify tessellations of objects with triangles, quadralaterals, hexagons, and so on. The conversion process is a lot simpler that for the IFC format since the format is limited to this type og geometric representation.
+
+On the other hand, there is the Point Cloud Processing module. This makes it possible to visualize point clouds of the file type x, y, and z.
+This module has references to the third party libraries:
+* Qt5 ()
+* Liblas ()
+* Eigen ()
+* Cloud Compare ()
+
+By default the point cloud processing part is deactivated. In order to make use of it the corresponding CMake option needs to be selected (see [CMake Options Documentation](Documentation/markdown/CMakeOptions.md)). The good thing is that the differnet modules can be run simultaneously so that, e.g., one is able to open an IFC file and point cloud at the same time to dirrectly compare them to each other.
+
+### <a name="UI"></a> User Interface
+
+The underlying part of the user interface is the Rendering module. This contains a self written rendering engine. With the help of this things can be displayed. Depending on the displayed object different types of shaders are implemented. On top of this we have the Graphical User Interface. It bases on the Qt5 library which has a big catalogue of ready to use user interface elemtents. Other linked third party libraries are:
+* Zlib ()
+* Proj ()
+On the other hand there is also the Unit Testing part. This bases on the google test. Since many unit test base on the visual comparison (visual test) the unit test base includes a geometry model renderer that allows rendering without the use of the graphical user interface. On top of there a various unit tests for IFC, OFF, point clouds and also some other more general unit tests? These provide automated checking whether the implemented funtionalities are all still in place and working. For the case of IFC the [List of supported IFC Representations](./SupportedIFCrepresentations.md) provides information which representation currently has a unit testing in place.
+
[…]
