### 2021-07-27 1dcf6c9
Reformulate overall software architecture part
