+++ Documentation/markdown/SoftwareArchitecture.md
@@ -15,58 +15,59 @@
 
 ## <a name="Overall_architecture"></a> Overall Software Architecture
 
-This section will give a general overview of the software architecture of the Open Infra Platform with its core software modules and how they are connected.
-Furthermore, dependies on third party libraries are stated. 
+This section will give a general overview of the software architecture of the Open Infra Platform with its core software modules together with their interrelations.
+Furthermore, dependencies on third-party libraries are stated. 
 
 ![](../images/software_architecture.PNG)
 
 ### <a name="Fundaments"></a> Fundaments
-Base lies the backbone for everything else that follows. Main part are the third party libraries that are linked here. Other than that we have only IModel.h Interface, GeorefMetaData.h, and BBox.h.
+The base module constitutes the backbone of the Open Infra Platform, with all of the other software parts building upon it. As its main part, several third-party libraries are linked here. Other than that, it also provides very fundamental elements like the IModel interface that is implemented by the upper layers for every supported file format (e.g., IFC and OFF) and will hold all of the information read from the input files. Furthermore, classes and structures for georeferencing metadata and simple bounding boxes are given.
 
-Third Party Libraries:
-* Boost (header-only, library for linear algebra, mutlithreading and in general storage management)
-* BlueFramework (open-source rendering engine developed by TUM)
-* Carve (library to work with constructive solid geometry)
-* Tclap (header-only, very small and flexible library providing an interface for accessing and defining command line arguments)
-* FFTW3 (header-only, used for computing fourier trasformations)
-* Doxygen (header-only, tool to generate documentation)
+Linked third-party libraries:
+* Boost (library for linear algebra, multithreading, and storage management; header-only)
+* BlueFramework (open-source C++ 3D graphics rendering engine developed by TUM)
+* Carve (constructive solid geometry library)
+* Tclap (small and flexible library providing an interface for accessing and defining command line arguments; header-only)
+* FFTW3 (computation of Fourier transformations; header-only)
+* Doxygen (tool to generate documentation; header-only)
 
 ### <a name="Modules"></a> Modules
 
-The modules are sperate in two big parts. First of all we have the Core section which links to the third party libraries Eigen () and Qt5 ().
-In here we have two differnet file formats that we support. First of all, we have the IFC Geometry Converter. It makes you of the Express Binding Library, which is generated by the Express Binding Generator, explained more in Detail later on. The IFC file format supports all sorts of geometric representations ranging from boundary representation and constructive solid geometry up to sweep volumes. So implicit as well as explicit geometric representations. In order to be able to render the geometries described in an IFC file the task of the IFC Geometry Converter is to convert the geometric descriptions specified into a format that can be understood by the renderer. In our case these are triangles. Since IFC is updating constantly (adding and changing IfcRepresenationItems) we also are always trying to keep up to date. Unfortunately not all IfcRepresentationItems are supported at the moment. However, the OIP already supports many of them. To keep track of what is supported you can take a look at [List of supported IFC Representations](./SupportedIFCrepresentations.md). The other part that is also part of Core is the OFF (Object File Format) Geometry Converter. This format is used to specify tessellations of objects with triangles, quadralaterals, hexagons, and so on. The conversion process is a lot simpler that for the IFC format since the format is limited to this type og geometric representation.
+The modules can be separated into two central parts. First of all, the Core module links to the third-party libraries Eigen and Qt.
[…]
