### 2021-07-29 0ee9281
Fixed mistakes, added new function "getNumberOfTessellationSegmentsByLength" for segmentation
