### 2021-07-29 655bc35
Fixed mistakes, added new function "getNumberOfTessellationSegmentsByLength" for segmentation
