### 2021-07-29 34bd9d8
Fixed mistakes, added new function "getNumberOfTessellationSegmentsByLength" for segmentation
