+++ Core/src/IfcGeometryConverter/GeomUtils.cpp
@@ -1230,10 +1230,10 @@ bool GeomUtils::checkMeshSet( const carve::mesh::MeshSet<3>* mesh_set,
 
 	/*!Calculate recursive multiplication for calculation Taylor series
 	* \param[in]  value                  value of Taylor series
-	* \param[in]  polynomialConstants    polynomial constants for a polyomial term - a,b,c,d 
+	* \param[in]  polynomialConstants    polynomial constants for a polynomial term - a,b,c,d 
 	* \param[in]  myArray                matrix with 1xn number of iteration points
-	* \param[in]  i_n                    aktuôä itaration point
-	* \param[in]  n                      number of itaration points
+	* \param[in]  i_n                    actual iteration point
+	* \param[in]  n                      number of iteration points
 	*
 	* \return
 	*/
@@ -1245,9 +1245,9 @@ bool GeomUtils::checkMeshSet( const carve::mesh::MeshSet<3>* mesh_set,
 			const int	n
 		)
 	{
-		if (!value) throw oip::InconsistentGeometryException("...");
+		if (!value) throw oip::InconsistentGeometryException("Invalid polynomial constants!");//to add
 		if (i_n < n - 1) {
-			if(polynomialConstants.size() && polynomialConstants[0] != 0.) throw oip::InconsistentGeometryException("...");
+			if(polynomialConstants.size() == 0 || polynomialConstants[0] != 0.) throw oip::InconsistentGeometryException("Invalid polynomial constants!");
 			for (int i = 1; i < polynomialConstants.size(); i++) {
 				if (polynomialConstants[i]) {
 					GeomUtils::recursiveMultiplicationTaylor(
@@ -1260,10 +1260,10 @@ bool GeomUtils::checkMeshSet( const carve::mesh::MeshSet<3>* mesh_set,
 			}
 		}
 		else {
-			if(i_n == n - 1);
+			if (i_n != n - 1) throw oip::InconsistentGeometryException("Invalid polynomial constants!");
 			for (int i = 1; i < polynomialConstants.size(); i++) {
 				if (polynomialConstants[i]) {
-					if(myArray.size() + i <= n && myArray.size() + i > polynomialConstants.size()* n + 1) throw oip::InconsistentGeometryException("...");
+					if(myArray.size() + i <= n && myArray.size() + i > polynomialConstants.size()* n + 1) throw oip::InconsistentGeometryException("Invalid polynomial constants!");
 					myArray[myArray.size() + i] += value * polynomialConstants[i];
 				}
 			}
@@ -1274,7 +1274,7 @@ bool GeomUtils::checkMeshSet( const carve::mesh::MeshSet<3>* mesh_set,
 
 	/*!Calculate the Taylor series 
 	* \param[in]  n                      number of iteration points
-	* \param[in]  polynomialConstants    polynomial constants for a polyomial term - a,b,c,d
+	* \param[in]  polynomialConstants    polynomial constants for a polynomial term - a,b,c,d
 	* \param[in]  s                      the length of the curve between two points 
 	*
 	* \return
@@ -1291,7 +1291,7 @@ bool GeomUtils::checkMeshSet( const carve::mesh::MeshSet<3>* mesh_set,
 		// 	x =  indeterminate
 		// 	   =>	pC[c-1] * x^(c-1) + pC[c-2] * x^(c-2) + ... + pC[1] * x^(1) + pC[0] * x^(0) where pC[0] == 0
 		//
-		if(polynomialConstants.size() && polynomialConstants[0] != 0.) throw oip::InconsistentGeometryException("...");
[…]
