### 2021-07-30 81ca3f1
current mistakes fixed, further comments needed
