+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -937,6 +937,7 @@ namespace OpenInfraPlatform
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcClothoid>())
 					{
 						return convertIfcClothoid(spiral.as<typename IfcEntityTypesT::IfcClothoid>(),
+<<<<<<< HEAD
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcClothoid
 
@@ -1009,27 +1010,29 @@ namespace OpenInfraPlatform
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcClothoid>())
 					{
 						return convertIfcClothoid(conic.as<typename IfcEntityTypesT::IfcClothoid>(),
+=======
+>>>>>>> 3a97f03f (current mistakes fixed, further comments needed)
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcClothoid
 
 					// IfcCosine SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcCosine>())
 					{
-						return convertIfcCosine(conic.as<typename IfcEntityTypesT::IfcCosine>(),
+						return convertIfcCosine(spiral.as<typename IfcEntityTypesT::IfcCosine>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcCosine
 
 					// IfcHelmertCurve SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcHelmertCurve>())
 					{
-						return convertIfcHelmertCurve(conic.as<typename IfcEntityTypesT::IfcHelmertCurve>(),
+						return convertIfcHelmertCurve(spiral.as<typename IfcEntityTypesT::IfcHelmertCurve>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcHelmertCurve
 
 					// IfcSine SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcSine>())
 					{
-						return convertIfcSine(conic.as<typename IfcEntityTypesT::IfcSine>(),
+						return convertIfcSine(spiral.as<typename IfcEntityTypesT::IfcSine>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcSine
 
@@ -1161,6 +1164,7 @@ namespace OpenInfraPlatform
 					
 					// Apply position
 <<<<<<< HEAD
+<<<<<<< HEAD
 					if (!clothoidPoints.empty())
 					{
 						std::vector<carve::geom::vector<3>> newClothoidPoints;
@@ -1231,6 +1235,9 @@ namespace OpenInfraPlatform
 						newClothoidPoints.push_back(point3d);
 =======
 					if (clothoidPoints.size() > 0)
+=======
+					if (!clothoidPoints.empty())
+>>>>>>> 3a97f03f (current mistakes fixed, further comments needed)
 					{
 						std::vector<carve::geom::vector<3>> newClothoidPoints;
 						for (unsigned int i = 0; i < clothoidPoints.size(); ++i)
@@ -1738,6 +1745,7 @@ namespace OpenInfraPlatform
 					std::vector<carve::geom::vector<3>>& segmentStartPoints
 				) const noexcept(false)
 				{
+<<<<<<< HEAD
[…]
