### 2021-07-30 3a97f03
current mistakes fixed, further comments needed
