### 2021-07-30 29b0b3f
current mistakes fixed, further comments needed
