+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -922,28 +922,28 @@ namespace OpenInfraPlatform
 					// IfcClothoid SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcClothoid>())
 					{
-						return convertIfcClothoid(conic.as<typename IfcEntityTypesT::IfcClothoid>(),
+						return convertIfcClothoid(spiral.as<typename IfcEntityTypesT::IfcClothoid>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcClothoid
 
 					// IfcCosine SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcCosine>())
 					{
-						return convertIfcCosine(conic.as<typename IfcEntityTypesT::IfcCosine>(),
+						return convertIfcCosine(spiral.as<typename IfcEntityTypesT::IfcCosine>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcCosine
 
 					// IfcHelmertCurve SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcHelmertCurve>())
 					{
-						return convertIfcHelmertCurve(conic.as<typename IfcEntityTypesT::IfcHelmertCurve>(),
+						return convertIfcHelmertCurve(spiral.as<typename IfcEntityTypesT::IfcHelmertCurve>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcHelmertCurve
 
 					// IfcSine SUBTYPE OF IfcSpiral
 					else if (spiral.isOfType<typename IfcEntityTypesT::IfcSine>())
 					{
-						return convertIfcSine(conic.as<typename IfcEntityTypesT::IfcSine>(),
+						return convertIfcSine(spiral.as<typename IfcEntityTypesT::IfcSine>(),
 							targetVec, segmentStartPoints, trim1Vec, trim2Vec, senseAgreement, trimmingPreference);
 					} // end if IfcSine
 
@@ -1035,7 +1035,7 @@ namespace OpenInfraPlatform
 					std::vector<carve::geom::vector<3>> clothoidPoints;
 					clothoidPoints[0] = carve::geom::VECTOR(0., 0., 0.);
 			
-					for (int i = 1; i < numSegments; ++i) 
+					for (int i = 0; i < numSegments; ++i) 
 					{
 						clothoidPoints.push_back(getPointOnCurve(clothoid, trim1Vec, trimmingPreference));
 						lengthNew += segmentLength;
@@ -1043,7 +1043,7 @@ namespace OpenInfraPlatform
 					}
 					
 					// Apply position
-					if (clothoidPoints.size() > 0)
+					if (!clothoidPoints.empty())
 					{
 						std::vector<carve::geom::vector<3>> newClothoidPoints;
 						for (unsigned int i = 0; i < clothoidPoints.size(); ++i)
@@ -1503,22 +1503,28 @@ namespace OpenInfraPlatform
 					std::vector<carve::geom::vector<3>>& segmentStartPoints
 				) const noexcept(false)
 				{
-					throw oip::UnhandledException(segment);
-					/*
+					//throw oip::UnhandledException(segment);
+					
+					// IfcCurveSegment SUPTYPE of IfcSegment
 					if (segment.isOfType<typename IfcEntityTypesT::IfcCurveSegment>())
 					{
[…]
