### 2021-08-17 30a64e9
Add correction notes and reduce characters per line
