### 2021-08-17 03fe3a8
Improvements to software architecture description
