+++ Documentation/markdown/SoftwareArchitecture.md
@@ -1,7 +1,7 @@
 # Software Architecture
 
 *NOTE:* This document shows the overall software architecture of the Open Infra Platform 
-and the Express Binding Generator as of the current state (July 2021). 
+and the Express Binding Generator as of the current state (August 2021). 
 
 ## Content 
 
@@ -26,80 +26,67 @@ Furthermore, dependencies on third-party libraries are stated.
 The base module constitutes the backbone of the Open Infra Platform, 
 with all of the other software parts building upon it. 
 As its main part, several third-party libraries are linked here. 
-Other than that, it also provides very fundamental elements like the IModel interface 
-that is implemented by the upper layers for every supported file format (e.g., IFC and OFF) 
-and will!!! hold all of the information read from the input files. 
+Other than that, it also provides very fundamental elements like the _IModel_ interface 
+that is implemented by the upper layers for every supported file format (e.g., `IFC` and `OFF`)
+and holds all of the information read from the input files. 
 Furthermore, classes and structures for georeferencing metadata and simple bounding boxes are given.
-
-Linked third-party libraries:
-* Boost (library for linear algebra, multithreading, and storage management; header-only)
-* BlueFramework (open-source C++ 3D graphics rendering engine developed by TUM)
-* Carve (constructive solid geometry library)
-* Tclap (small and flexible library providing an interface for accessing and defining command line arguments; header-only)
-* FFTW3 (computation of Fourier transformations; header-only)
-* Doxygen (tool to generate documentation; header-only)
-(main readme? -> oder seperates readme für alles thirdparty)
-lizenzen readme -> nur noch link
+The third-party libraries linked to this module are Boost, BlueFramework, Carve, Tclap, FFTW3, and Doxygen.
+More details on these libraries can be found [here](Licenses/readme.md).
 
 ### <a name="Modules"></a> Modules
 
 The modules can be separated into two central parts. 
-First of all, the Core module links to the third-party libraries Eigen and Qt.
-It also comprises one submodule for the IFC and OFF file format each. 
-In the case of IFC, the IFC Geometry Converter takes care of the geometric conversion processes. 
-It makes use of the Express Binding Library, which is generated by the Express Binding Generator (Link nach unten), 
-(and explained in further detail later on). 
+First of all, the _Core_ module links to the third-party libraries `Eigen` and `Qt`.
+It also comprises one submodule for the IFC (Industry Foundation Classes) and OFF (Object File Format) file format each. 
+In the case of IFC, the _IFC Geometry Converter_ takes care of the geometric conversion processes. 
+It makes use of the _Express Binding Library_, which is generated by the [Express Binding Generator](#Express_binding). 
 The IFC file format supports all sorts of geometric representations 
-ranging from boundary representation and constructive solid geometry to sweep volumes. 
+ranging from boundary representation and constructive solid geometry to sweep volumes.
 This includes implicit as well as explicit geometric representations. 
 In order to be able to render these complex geometries described in an IFC file, 
 the _IFC Geometry Converter_ has to convert the geometric descriptions 
 specified in IFC into a triangle mesh that can be understood by the rendering engine.
 This step is based on the geometry kernel `carve`. 
 Since IFC encompasses a multitude of representations and is constantly updating, 
-only a subset of the IfcRepresentationItems are supported at the moment, 
+only a subset of the _IfcRepresentationItems_ are supported at the moment, 
 focusing on the most frequently used ones. 
 However, OIP is continuously expanding to cover more and more types of representations. 
 To check which representation items are currently supported, 
-the user can take a look at the (weiter oben)[List of supported IFC Representations](./SupportedIFCrepresentations.md). 
[…]
