+++ Documentation/markdown/SoftwareArchitecture.md
@@ -23,12 +23,12 @@ Furthermore, dependencies on third-party libraries are stated.
 ![](../images/software_architecture.PNG)
 
 ### <a name="Fundaments"></a> Fundaments
-The base module constitutes the backbone of the Open Infra Platform, 
+The _Base_ module constitutes the backbone of the Open Infra Platform, 
 with all of the other software parts building upon it. 
 As its main part, several third-party libraries are linked here. 
 Other than that, it also provides very fundamental elements like the _IModel_ interface 
 that is implemented by the upper layers for every supported file format (e.g., `IFC` and `OFF`)
-and holds all of the information read from the input files. 
+and provides some basic information about the input files. 
 Furthermore, classes and structures for georeferencing metadata and simple bounding boxes are given.
 The third-party libraries linked to this module are Boost, BlueFramework, Carve, Tclap, FFTW3, and Doxygen.
 More details on these libraries can be found [here](Licenses/readme.md).
@@ -40,6 +40,7 @@ First of all, the _Core_ module links to the third-party libraries `Eigen` and `
 It also comprises one submodule for the IFC (Industry Foundation Classes) and OFF (Object File Format) file format each. 
 In the case of IFC, the _IFC Geometry Converter_ takes care of the geometric conversion processes. 
 It makes use of the _Express Binding Library_, which is generated by the [Express Binding Generator](#Express_binding). 
+
 The IFC file format supports all sorts of geometric representations 
 ranging from boundary representation and constructive solid geometry to sweep volumes.
 This includes implicit as well as explicit geometric representations. 
@@ -53,6 +54,7 @@ focusing on the most frequently used ones.
 However, OIP is continuously expanding to cover more and more types of representations. 
 To check which representation items are currently supported, 
 the user can take a look at the [List of supported IFC Representations](./SupportedIFCrepresentations.md). 
+
 The other part of the _Core_ module consists of the _OFF Geometry Converter_. 
 This format specifies tessellations of objects with the help of triangles, quadrilaterals, hexagons, and similar surfaces. 
 The conversion process is a lot simpler in comparison to the IFC format 
@@ -75,8 +77,9 @@ at the same time to compare them to each other directly.
 ### <a name="UI"></a> User Interface
 
 The underlying part of the user interface is the _Rendering module_. 
-This links to a self-written rendering engine called _BlueFramework_. 
+This links to a self-written rendering engine called _BlueFramework_ ([see repository](https://github.com/tumcms/Blue-Framework)). 
 Depending on the objects that are to be displayed, different types of shaders and effects are implemented. 
+
 The _Graphical User Interface_ is built on top of this, 
 which has a 3D viewport to display geometric objects and several menus and buttons that provide functions to interact with them. 
 The GUI bases on the `Qt` library, which has an extensive catalog of ready-to-use user interface elements. 
@@ -84,7 +87,7 @@ Other third-party libraries that the GUI uses are Zlib and Proj (for more detail
 
 ### <a name="UT"></a> Unit Testing
 
-On the other hand, there is the _Unit Testing_ part, which bases on `Google Test`. 
+The _Unit Testing_ part of OIP is based on `Google Test`. 
 Since many unit tests that are implemented in OIP use visual comparison, 
 the unit test base includes a geometry model renderer that allows rendering without the use of the _Graphical User Interface_. 
 On top of that, there are various individual unit tests for IFC, OFF, and point clouds (and also some other more general unit tests?). 
@@ -100,7 +103,7 @@ and automatically create early binding C++ classes mirroring the IFC entities.
 This allows to find coding errors already at compile time. 
 The _Express Binding Generator_ consists of three main parts: 
[…]
