### 2021-08-22 8320b1d
Added new UT for IfcSectionalSolidHorizontal
