### 2021-08-23 8560cf3
Check IfcSectionedSolidHorizontal, fixed some mistakes
