### 2021-08-24 2267315
Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment
