+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -491,12 +491,7 @@ namespace OpenInfraPlatform {
 					// **************************************************************************************************************************
 					for (auto &segment: compositeCurve->Segments) {
 						std::vector<carve::geom::vector<3>> segment_vec;
-
-#if defined(OIP_MODULE_EARLYBINDING_IFC4X1) || defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC1)
-						convertIfcCurve(segment->ParentCurve, segment_vec, segmentStartPoints);
-#else
 						convertIfcSegment(segment, segment_vec, segmentStartPoints);
-#endif
 						if (!segment_vec.empty()) {
 							GeomUtils::appendPointsToCurve(segment_vec, targetVec);
 						}
@@ -855,13 +850,8 @@ namespace OpenInfraPlatform {
 					GeomUtils::appendPointsToCurve(basisCurvePoints, targetVec);
 				}
 
-<<<<<<< HEAD
-				// IfcSpiral SUBTYPE of IfcCurve (exists starting IFC4X3_RC3)
-#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
-=======
-				// IfcClothoid SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
+				// IfcSpiral SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
->>>>>>> origin/development
 				/**********************************************************************************************/
 				/*! \brief Converts an \c IfcSpiral to a tesselated curve.
 				* \param[in] spiral					A pointer to data from \c Spiral.
@@ -932,6 +922,7 @@ namespace OpenInfraPlatform {
 					// the rest we do not support
 					throw oip::UnhandledException(spiral);
 				}
+
 				// IfcClothoid SUBTYPE of IfcSpiral (exists starting IFC4X3_RC3)
 				/**********************************************************************************************/
 				/*! \brief Converts an \c IfcClothoid to a tesselated curve.
@@ -1019,8 +1010,10 @@ namespace OpenInfraPlatform {
 			
 					for (int i = 0; i < numSegments; ++i) 
 					{
-						clothoidPoints.push_back(getPointOnCurve(clothoid, trim1Vec, trimmingPreference));
-						lengthNew += segmentLength;
+						const IfcEntityTypesT::IfcParameterValue val = lengthNew;
+
+						clothoidPoints.push_back(getPointOnCurve(clothoid, val));
+						//val += segmentLength;
 					
 					}
 					
@@ -1494,7 +1487,7 @@ namespace OpenInfraPlatform {
 					// IfcCompositeCurveSegment SUPTYPE of IfcSegment
 					else if  (segment.isOfType<typename IfcEntityTypesT::IfcCompositeCurveSegment>())
 					{
-						return IfcCompositeCurveSegment(segment.as<typename IfcEntityTypesT::IfcCompositeCurveSegment>(), targetVec, segmentStartPoints);
+						return convertIfcCurve(segment.as<typename IfcEntityTypesT::IfcCompositeCurveSegment>()->ParentCurve, targetVec, segmentStartPoints);
 					}
 					else {
 						throw oip::UnhandledException(segment);
@@ -1515,16 +1508,88 @@ namespace OpenInfraPlatform {
 					std::vector<carve::geom::vector<3>>& segmentStartPoints
 				) const throw(...)
 				{
[…]
