### 2021-08-24 ba3e9d1
Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment
