+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -492,14 +492,14 @@ namespace OpenInfraPlatform
 					// **************************************************************************************************************************
 					for (auto &segment : compositeCurve->Segments) {
 						std::vector<carve::geom::vector<3>> segment_vec;
-
-#if defined(OIP_MODULE_EARLYBINDING_IFC4X1) || defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC1)
-						convertIfcCurve(segment->ParentCurve, segment_vec, segmentStartPoints);
-#else
 						convertIfcSegment(segment, segment_vec, segmentStartPoints);
+<<<<<<< HEAD
 #endif
 						if (!segment_vec.empty()) 
 						{
+=======
+						if (!segment_vec.empty()) {
+>>>>>>> 22673156 (Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment)
 							GeomUtils::appendPointsToCurve(segment_vec, targetVec);
 						}
 					}
@@ -871,6 +871,7 @@ namespace OpenInfraPlatform
 
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 				// IfcClothoid SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
 =======
@@ -880,6 +881,10 @@ namespace OpenInfraPlatform
 >>>>>>> 875f326b (Fixed mistakes in Spiral function)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
 >>>>>>> 92a31da5 (Added IfsSpiral converter)
+=======
+				// IfcSpiral SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
+#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
+>>>>>>> 22673156 (Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment)
 				/**********************************************************************************************/
 				/*! \brief Converts an \c IfcSpiral to a tesselated curve.
 				* \param[in] spiral					A pointer to data from \c Spiral.
@@ -950,6 +955,7 @@ namespace OpenInfraPlatform
 					// the rest we do not support
 					throw oip::UnhandledException(spiral);
 				}
+
 				// IfcClothoid SUBTYPE of IfcSpiral (exists starting IFC4X3_RC3)
 				/**********************************************************************************************/
 				/*! \brief Converts an \c IfcClothoid to a tesselated curve.
@@ -1037,8 +1043,10 @@ namespace OpenInfraPlatform
 			
 					for (int i = 0; i < numSegments; ++i) 
 					{
-						clothoidPoints.push_back(getPointOnCurve(clothoid, trim1Vec, trimmingPreference));
-						lengthNew += segmentLength;
+						const IfcEntityTypesT::IfcParameterValue val = lengthNew;
+
+						clothoidPoints.push_back(getPointOnCurve(clothoid, val));
+						//val += segmentLength;
 					
 					}
 					
@@ -1513,7 +1521,7 @@ namespace OpenInfraPlatform
 					// IfcCompositeCurveSegment SUPTYPE of IfcSegment
 					else if  (segment.isOfType<typename IfcEntityTypesT::IfcCompositeCurveSegment>())
 					{
-						return IfcCompositeCurveSegment(segment.as<typename IfcEntityTypesT::IfcCompositeCurveSegment>(), targetVec, segmentStartPoints);
[…]
