### 2021-08-24 674525b
Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment
