+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -495,10 +495,13 @@ namespace OpenInfraPlatform
 						std::vector<carve::geom::vector<3>> segment_vec;
 						convertIfcSegment(segment, segment_vec, segmentStartPoints);
 <<<<<<< HEAD
+<<<<<<< HEAD
 #endif
 						if (!segment_vec.empty()) 
 						{
 =======
+=======
+>>>>>>> 22673156 (Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment)
 						if (!segment_vec.empty()) {
 >>>>>>> 22673156 (Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment)
 							GeomUtils::appendPointsToCurve(segment_vec, targetVec);
@@ -875,6 +878,7 @@ namespace OpenInfraPlatform
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 				// IfcClothoid SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
 =======
@@ -890,6 +894,10 @@ namespace OpenInfraPlatform
 >>>>>>> 875f326b (Fixed mistakes in Spiral function)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC3)
 >>>>>>> 92a31da5 (Added IfsSpiral converter)
+=======
+				// IfcSpiral SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
+#if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
+>>>>>>> 22673156 (Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment)
 =======
 				// IfcSpiral SUBTYPE of IfcCurve (exists starting IFC4X3_RC4)
 #if defined(OIP_MODULE_EARLYBINDING_IFC4X3_RC4)
@@ -1039,6 +1047,7 @@ namespace OpenInfraPlatform
 					// the rest we do not support
 					throw oip::UnhandledException(spiral);
 				}
+
 				// IfcClothoid SUBTYPE of IfcSpiral (exists starting IFC4X3_RC3)
 				/**********************************************************************************************/
 				/*! \brief Converts an \c IfcClothoid to a tesselated curve.
@@ -1151,14 +1160,20 @@ namespace OpenInfraPlatform
 					for (int i = 0; i < numSegments; ++i) 
 					{
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> 22673156 (Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment)
 						const IfcEntityTypesT::IfcParameterValue val = lengthNew;
 
 						clothoidPoints.push_back(getPointOnCurve(clothoid, val));
 						//val += segmentLength;
+<<<<<<< HEAD
 =======
 						clothoidPoints.push_back(getPointOnCurve(clothoid, trim1Vec, trimmingPreference));
 						lengthNew += segmentLength;
 >>>>>>> 34bd9d8c (Fixed mistakes, added new function "getNumberOfTessellationSegmentsByLength" for segmentation)
+=======
+>>>>>>> 22673156 (Fixed mistakes in Taylor Series, convertIfcClothoid; added implementation for points in convertIfcCurveSegment)
 					
 					}
 					
@@ -1746,6 +1761,7 @@ namespace OpenInfraPlatform
 				) const noexcept(false)
 				{
 <<<<<<< HEAD
+<<<<<<< HEAD
[…]
