### 2021-08-24 1dd9945
changes made in convertIfcCurveSegment, getPointOnCurve for IfcClothoid which includes IfcParameterValue and IfcNonNegativeLengthMeasure as parameter
