+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 ﻿/* -*-c++-*- IfcPlusPlus - www.ifcplusplus.com  - Copyright (C) 2011 Fabian Gerold
  * This library is open source and may be redistributed and/or modified under
  * the terms of the OpenSceneGraph Public License (OSGPL) version 0.0 or
@@ -2576,3 +2577,2560 @@ namespace OpenInfraPlatform
 
 
 #endif
+=======
+﻿/* -*-c++-*- IfcPlusPlus - www.ifcplusplus.com  - Copyright (C) 2011 Fabian Gerold
+ * This library is open source and may be redistributed and/or modified under
+ * the terms of the OpenSceneGraph Public License (OSGPL) version 0.0 or
+ * (at your option) any later version.  The full license is in LICENSE file
+ * included with this distribution, and on the openscenegraph.org website.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * OpenSceneGraph Public License for more details.
+ *
+ * - modfied by Michael Kern (September 2014)
+*/
+
+// visual studio
+#pragma once
+// unix
+#ifndef CURVECONVERTER_H
+#define CURVECONVERTER_H
+
+#define _USE_MATH_DEFINES 
+#include <math.h>
+#include <boost/numeric/ublas/matrix.hpp>
+#include <boost/numeric/ublas/io.hpp>
+
+#include "CarveHeaders.h"
+
+#include "ConverterBase.h"
+
+#include "GeomUtils.h"
+#include "PlacementConverter.h"
+#include "ProfileConverter.h"
+#include "SplineConverter.h"
+
+#include "BlueFramework/Core/Diagnostics/log.h"
+
+
+namespace OpenInfraPlatform {
+	namespace Core {
+		namespace IfcGeometryConverter {
+
+			/*! \brief Converter functionality for \c IfcCurve's subtypes
+			*
+			* \param IfcEntityTypesT The IFC version templates
+			*/
+			template <
+				class IfcEntityTypesT
+			>
+			class CurveConverterT : public ConverterBaseT<IfcEntityTypesT>
+			{
+			public:
+				//! Constructor
+				CurveConverterT(
+					std::shared_ptr<GeometrySettings> geomSettings,
+					std::shared_ptr<UnitConverter<IfcEntityTypesT>> unitConverter,
+					std::shared_ptr<PlacementConverterT<IfcEntityTypesT>> pc)
+					:
+					ConverterBaseT<IfcEntityTypesT>(geomSettings, unitConverter),
+					placementConverter(pc)
+				{
+				}
+
+				//! Virtual destructor
+				virtual ~CurveConverterT()
+				{
+
+				}
+
+				/*! \brief Converts an \c IfcCurve to an array of segments to be rendered on screen.
+				 *
+				 * \param[in] ifcCurve		The \c IfcCurve to be converted.
+				 * \param[in] pos			The relative location of the origin of the representation's coordinate system within the geometric context.
+				 * \param[out] itemData		A pointer to be filled with the relevant data.
+				 *
+				 * \note Calls the other overload CurveConverterT::convertIfcCurve with corresponding input parameters. Adds their result to itemData.
+				 */
+				void convertIfcCurve(const EXPRESSReference<typename IfcEntityTypesT::IfcCurve>& ifcCurve,
+					const carve::math::Matrix& pos,
+					std::shared_ptr<ItemData>& itemData
[…]
