### 2021-08-24 b0cc65b
changes made in convertIfcCurveSegment, getPointOnCurve for IfcClothoid which includes IfcParameterValue and IfcNonNegativeLengthMeasure as parameter
