+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -1508,10 +1508,15 @@ namespace OpenInfraPlatform {
 					std::vector<carve::geom::vector<3>>& segmentStartPoints
 				) const throw(...)
 				{
-					//	ENTITY IfcCurveSegment					//	  SUBTYPE OF(IfcSegment);					//       Placement: IfcPlacement;					//       SegmentStart: IfcCurveMeasureSelect;					//       SegmentLength: IfcCurveMeasureSelect;					//       ParentCurve: IfcCurve;					//	  DERIVE					//		Dim : IfcDimensionCount: = ParentCurve.Dim;					//	END_ENTITY;
-
-					// start point
-					auto startPoint = getPointOnCurve(curveSegment->ParentCurve, curveSegment->SegmentStart);
+					//	ENTITY IfcCurveSegment
+					//	  SUBTYPE OF(IfcSegment);
+					//       Placement: IfcPlacement;
+					//       SegmentStart: IfcCurveMeasureSelect;
+					//       SegmentLength: IfcCurveMeasureSelect;
+					//       ParentCurve: IfcCurve;
+					//	  DERIVE
+					//		Dim : IfcDimensionCount: = ParentCurve.Dim;
+					//	END_ENTITY;
 
 					// Get length of the segment curve
 					double length = 0.;
@@ -1521,11 +1526,13 @@ namespace OpenInfraPlatform {
 					{
 						// Calculate the length using \c IfcNonNegativeLengthMeasure. 
 						length = curveSegment->SegmentLength.get<0>();// *this->UnitConvert()->getLengthInMeterFactor();
+						break;
 					}
 					case 1:
 					{
 						// Calculate the length using \c IfcParameterValue.
 						length = curveSegment->SegmentLength.get<1>();
+						break;
 					}
 					default:
 						throw oip::InconsistentGeometryException(curveSegment, "Could not determine SegmentLength!");
@@ -1543,21 +1550,36 @@ namespace OpenInfraPlatform {
 					{
 						// Calculate the length using \c IfcNonNegativeLengthMeasure. 
 						lengthAlong = curveSegment->SegmentStart.get<0>();// *this->UnitConvert()->getLengthInMeterFactor();
+						break;
 					}
 					case 1:
 					{
 						// Calculate the length using \c IfcParameterValue.
 						lengthAlong = curveSegment->SegmentStart.get<1>();
+						break;
 					}
 					default:
 						throw oip::InconsistentGeometryException(curveSegment, "Could not determine SegmentStart!");
 					}
 
-					auto runningLength = curveSegment->SegmentStart;
+				    auto runningLength = curveSegment->SegmentStart;
 					for( int i = 0; i < numSegments; ++i )
 					{ 
 						// determine point
-						segmentPoints.push_back(getPointOnCurve(curveSegment->ParentCurve, runningLength));
+						carve::geom::vector<3> point;
+						if (curveSegment->ParentCurve.isOfType<typename IfcEntityTypesT::IfcClothoid>())
+						{
+							point = getPointOnCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcClothoid>(), runningLength));
+						}
+						else if (curveSegment->ParentCurve.isOfType<typename IfcEntityTypesT::IfcLine>())
+						{
+							point = getPointOnCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcLine>(), runningLength));
+						}
+						else (curveSegment->ParentCurve.isOfType<typename IfcEntityTypesT::IfcCircle>())
+						{
+							point = getPointOnCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcCircle>(), runningLength));
+						}
+						segmentPoints.push_back(point);
[…]
