### 2021-08-24 7d5a62b
changes made in convertIfcCurveSegment, getPointOnCurve for IfcClothoid which includes IfcParameterValue and IfcNonNegativeLengthMeasure as parameter
