+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -1,4 +1,5 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
 ﻿/* -*-c++-*- IfcPlusPlus - www.ifcplusplus.com  - Copyright (C) 2011 Fabian Gerold
  * This library is open source and may be redistributed and/or modified under
  * the terms of the OpenSceneGraph Public License (OSGPL) version 0.0 or
@@ -2912,6 +2913,8 @@ namespace OpenInfraPlatform
 
 #endif
 =======
+=======
+>>>>>>> b0cc65be (changes made in convertIfcCurveSegment, getPointOnCurve for IfcClothoid which includes IfcParameterValue and IfcNonNegativeLengthMeasure as parameter)
 ﻿/* -*-c++-*- IfcPlusPlus - www.ifcplusplus.com  - Copyright (C) 2011 Fabian Gerold
  * This library is open source and may be redistributed and/or modified under
  * the terms of the OpenSceneGraph Public License (OSGPL) version 0.0 or
@@ -2942,7 +2945,10 @@ namespace OpenInfraPlatform
 #include "ConverterBase.h"
 
 #include "GeomUtils.h"
+<<<<<<< HEAD
 #include "SpiralUtils.h"
+=======
+>>>>>>> b0cc65be (changes made in convertIfcCurveSegment, getPointOnCurve for IfcClothoid which includes IfcParameterValue and IfcNonNegativeLengthMeasure as parameter)
 #include "PlacementConverter.h"
 #include "ProfileConverter.h"
 #include "SplineConverter.h"
@@ -3901,12 +3907,34 @@ namespace OpenInfraPlatform {
 					double length = L / A * sqrt(M_PI);
 		
 					// Get number of tesselated segments
+<<<<<<< HEAD
 					//int numSegments = this->GeomSettings()->getNumberOfTessellationSegmentsByLength(length);
 					//Calculate segment length
 					int numSegments = 100;
 					double segmentLength = length / numSegments;
 					double lengthNew = segmentLength;
 
+=======
+					int numSegments = this->GeomSettings()->getNumberOfTessellationSegmentsByLength(length);
+					//Calculate segment length
+					double segmentLength = length / numSegments;
+					double lengthNew = segmentLength;
+
+					// Create a tessalated vector by clothoid length 
+					/*std::vector<carve::geom::vector<3>> ti;
+					ti[0] = 0;
+					// Calculate intervals for tessaleted vector
+					double pointInterval_ti = (l - 0) / (numSegments - 1);
+
+					//Add points to new tesselated vector
+					for (int i = 1; i < numSegments; ++i)
+					{
+						ti.push_back(ti[i - 1] + pointInterval_ti);
+					}
+					*/
+					// Calculate clothoid points. Parametrisation
+
+>>>>>>> b0cc65be (changes made in convertIfcCurveSegment, getPointOnCurve for IfcClothoid which includes IfcParameterValue and IfcNonNegativeLengthMeasure as parameter)
 					std::vector<carve::geom::vector<3>> clothoidPoints;
 					clothoidPoints[0] = carve::geom::VECTOR(0., 0., 0.);
 			
@@ -4441,10 +4469,17 @@ namespace OpenInfraPlatform {
 					}
 
 					// Determine segments
+<<<<<<< HEAD
 					double numSegments = 100;
 					double deltaLength = length / numSegments;
 
 					std::vector<carve::geom::vector<3>> segmentPoints, segmentDirections;
+=======
+					double numSegments = 500;
+					double deltaLength = length / numSegments;
+
+					std::vector<carve::geom::vector<3>> segmentPoints;
[…]
