### 2021-08-25 d235625
Added changes in convertIfcCurveSegment; Added new files in IfcGeometryConverter for spiral interpretation. Here the integrals for the spirals will be calculated. Source code is taken from taken Peter Bonsma
