+++ Core/src/IfcGeometryConverter/SpiralUtils.h
@@ -1,0 +1,1000 @@
+//
+//  Author:  Peter Bonsma
+//  $Date: 1999-12-31 23:59:59 +0000 (Wed, 31 Jan 1999) $
+//  $Revision: 3999 $
+//  Project: Geometry Kernel
+//
+//  In case of use of the DLL:
+//  be aware of license fee for use of this DLL when used commercially
+//  more info for commercial use:  peter.bonsma@rdf.bg
+//
+
+#ifndef __RDF_LTD__SPIRAL_H
+#define __RDF_LTD__SPIRAL_H
+
+
+#include	"mathematicsGeometryDoublePrecision.h"
+
+
+double			IntegralTaylorSeriesCos(
+	double		* polynomialConstants,
+	uint_t		polynomialConstantCnt,
+	double		s
+);
+
+double			IntegralTaylorSeriesSin(
+	double		* polynomialConstants,
+	uint_t		polynomialConstantCnt,
+	double		s
+);
+
+double			NumericalIntegrationCos(
+	double		* polynomialConstants,
+	uint_t		polynomialConstantCnt,
+	double		s
+);
+
+double			NumericalIntegrationSin(
+	double		* polynomialConstants,
+	uint_t		polynomialConstantCnt,
+	double		s
+);
+
+
+static	inline	double	XbyAngleDeviationPolynomial(
+	double	* polynomialConstants,
+	uint_t	polynomialConstantCnt,
+	double	s
+)
+{
+	double	* localPolynomialConstants = new double[polynomialConstantCnt + 1];
+	uint_t	localPolynomialConstantCnt = 1;
+
+	localPolynomialConstants[0] = 0.;
+
+	for (uint_t i = 0; i < polynomialConstantCnt; i++) {
+		if (polynomialConstants[i]) {
+			localPolynomialConstants[i + 1] = polynomialConstants[i] / (i + 1.);
+			localPolynomialConstantCnt = i + 2;
+		}
+		else {
+			localPolynomialConstants[i + 1] = 0;
+		}
+	}
+
+	double	rValue =
+		IntegralTaylorSeriesCos(
+			localPolynomialConstants,
+			localPolynomialConstantCnt,
+			s
+		);
+
+#ifdef _DEBUG
+	double	rValueNumericalIntegration =
+		NumericalIntegrationCos(
+			localPolynomialConstants,
+			localPolynomialConstantCnt,
+			s
+		);
+#endif // _DEBUG
+	assert(std::fabs(rValueNumericalIntegration - rValue) < 0.000001);
+
+	delete[] localPolynomialConstants;
+
+	return	rValue;
+}
+
+static	inline	double	XbyAngleDeviationPolynomial__NumericalIntegration(
+	double	* polynomialConstants,
+	uint_t	polynomialConstantCnt,
+	double	s
+)
+{
+	double	* localPolynomialConstants = new double[polynomialConstantCnt + 1];
+	uint_t	localPolynomialConstantCnt = 1;
+
+	localPolynomialConstants[0] = 0.;
+
+	for (uint_t i = 0; i < polynomialConstantCnt; i++) {
+		if (polynomialConstants[i]) {
+			localPolynomialConstants[i + 1] = polynomialConstants[i] / (i + 1.);
+			localPolynomialConstantCnt = i + 2;
+		}
+		else {
+			localPolynomialConstants[i + 1] = 0;
+		}
+	}
+
[…]
