+++ Core/src/IfcGeometryConverter/SpiralUtils.cpp
@@ -1,0 +1,543 @@
+//
+//  Author:  Peter Bonsma
+//  $Date: 1999-12-31 23:59:59 +0000 (Wed, 31 Jan 1999) $
+//  $Revision: 3999 $
+//  Project: Geometry Kernel
+//
+//  In case of use of the DLL:
+//  be aware of license fee for use of this DLL when used commercially
+//  more info for commercial use:  peter.bonsma@rdf.bg
+//
+
+#include	"../stdafx.h"
+
+#include	<stdio.h>
+#include	<stdlib.h>
+#include	<string.h>
+#include	<assert.h>
+#include	<cmath>
+
+
+#include	"spiral.h"
+
+
+void	RecursiveMultiplication(
+	double	value,
+	double	* polynomialConstants,
+	uint_t	polynomialConstantCnt,
+	double	* myArray,
+	uint_t	myArrayParentIndex,
+	uint_t	i_n,
+	uint_t	n
+)
+{
+	assert(value);
+	if (i_n < n - 1) {
+		assert(polynomialConstantCnt && polynomialConstants[0] == 0.);
+		for (uint_t i = 1; i < polynomialConstantCnt; i++) {
+			if (polynomialConstants[i]) {
+				RecursiveMultiplication(
+					value * polynomialConstants[i],
+					polynomialConstants, polynomialConstantCnt,
+					myArray, myArrayParentIndex + i,
+					i_n + 1, n
+				);
+			}
+		}
+	}
+	else {
+		assert(i_n == n - 1);
+		for (uint_t i = 1; i < polynomialConstantCnt; i++) {
+			if (polynomialConstants[i]) {
+				assert(myArrayParentIndex + i >= n && myArrayParentIndex + i < polynomialConstantCnt* n + 1);
+				myArray[myArrayParentIndex + i] += value * polynomialConstants[i];
+			}
+		}
+	}
+}
+
+double	IntegralTaylorSeriesExpansionElement(
+	uint_t	n,
+	double	* polynomialConstants,
+	uint_t	polynomialConstantCnt,
+	double	s
+)
+{
+	//
+	//	pC = polynomialConstants
+	//	c = polynomialConstantCnt
+	// 	x =  indeterminate
+	// 	   =>	pC[c-1] * x^(c-1) + pC[c-2] * x^(c-2) + ... + pC[1] * x^(1) + pC[0] * x^(0) where pC[0] == 0
+	//
+	assert(polynomialConstantCnt && polynomialConstants[0] == 0.);
+
+	if (n) {
+#ifdef _ALLOC
+		double	* myArray = (double*)calloc(polynomialConstantCnt * n + 1, sizeof(double));
+#else
+		double	* myArray = new double[polynomialConstantCnt * n + 1]();
+#endif
+
+		RecursiveMultiplication(
+			1.,
+			polynomialConstants, polynomialConstantCnt,
+			myArray, 0,
+			0, n
+		);
+
+		double	value = 0., factor = 1.;
+		for (uint_t k = 0; k < polynomialConstantCnt * n + 1; k++) {
+			factor *= s;
+			value += myArray[k] * factor / (k + 1);
+		}
+
+		uint_t	tB = n;
+		while (tB > 1) {
+			value /= (double)tB;
+			tB--;
+		}
+
+#ifdef _ALLOC
+		free(myArray);
+#else
+		delete[] myArray;
+#endif
+
+		return	value;
+	}
+	else {
+		return	s;
+	}
+}
+
+double	IntegralTaylorSeriesCosExpansion(
+	uint_t	i,
+	double	* polynomialConstants,
+	uint_t	polynomialConstantCnt,
+	double	s
+)
+{
[…]
