### 2021-09-02 39a6ed9
Added new representation for IfcSegment
