### 2021-09-02 a4af284
Added new representation for IfcSegment
