### 2021-09-02 7f5a6fa
Added new representation for IfcSegment
