### 2021-09-03 b722b42
SpiralUtils have been modified
