+++ Core/src/IfcGeometryConverter/SpiralUtils.h
@@ -1,3 +1,20 @@
+/*
+	Copyright (c) 2021 Technical University of Munich
+	Chair of Computational Modeling and Simulation.
+
+	TUM Open Infra Platform is free software; you can redistribute it and/or modify
+	it under the terms of the GNU General Public License Version 3
+	as published by the Free Software Foundation.
+
+	TUM Open Infra Platform is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
 //
 //  Author:  Peter Bonsma
 //  $Date: 1999-12-31 23:59:59 +0000 (Wed, 31 Jan 1999) $
@@ -9,992 +26,1041 @@
 //  more info for commercial use:  peter.bonsma@rdf.bg
 //
 
-#ifndef __RDF_LTD__SPIRAL_H
-#define __RDF_LTD__SPIRAL_H
+// visual studio
+#pragma once
+// unix
+#ifndef SPIRALURILS_H
+#define SPIRALUTILS_H
 
+#include <vector>
 
-#include	"mathematicsGeometryDoublePrecision.h"
+#include "namespace.h"
 
+#include "CarveHeaders.h"
 
-double			IntegralTaylorSeriesCos(
-	double		* polynomialConstants,
-	uint_t		polynomialConstantCnt,
-	double		s
-);
+//#include <buw.BlueEngine.h>
+#include <BlueFramework/Core/Math/Vector.h>
 
-double			IntegralTaylorSeriesSin(
-	double		* polynomialConstants,
-	uint_t		polynomialConstantCnt,
-	double		s
-);
-
-double			NumericalIntegrationCos(
-	double		* polynomialConstants,
-	uint_t		polynomialConstantCnt,
-	double		s
-);
-
-double			NumericalIntegrationSin(
-	double		* polynomialConstants,
-	uint_t		polynomialConstantCnt,
-	double		s
-);
-
-
-static	inline	double	XbyAngleDeviationPolynomial(
-	double	* polynomialConstants,
-	uint_t	polynomialConstantCnt,
-	double	s
-)
+namespace OpenInfraPlatform
 {
-	double	* localPolynomialConstants = new double[polynomialConstantCnt + 1];
-	uint_t	localPolynomialConstantCnt = 1;
+	namespace Core {
+		namespace IfcGeometryConverter {
 
-	localPolynomialConstants[0] = 0.;
+			/*!	\brief Helper functions for calculations of geometries
 
-	for (uint_t i = 0; i < polynomialConstantCnt; i++) {
-		if (polynomialConstants[i]) {
-			localPolynomialConstants[i + 1] = polynomialConstants[i] / (i + 1.);
-			localPolynomialConstantCnt = i + 2;
-		}
-		else {
-			localPolynomialConstants[i + 1] = 0;
-		}
-	}
+			*/
+			class SpiralUtils {
+			public:
 
-	double	rValue =
-		IntegralTaylorSeriesCos(
-			localPolynomialConstants,
-			localPolynomialConstantCnt,
-			s
-		);
+				static double	IntegralTaylorSeriesCos(
+					const std::vector<double>& polynomialConstants,
+					const int	polynomialConstantCnt,
+					double	s);
+
+				static double	IntegralTaylorSeriesSin(
+					const std::vector<double>& polynomialConstants,
+					const int polynomialConstantCnt,
+					const double s);
+
+				static double	NumericalIntegrationCos(
+					const double length,
[…]
