+++ Core/src/IfcGeometryConverter/SpiralUtils.cpp
@@ -1,3 +1,20 @@
+/*
+	Copyright (c) 2021 Technical University of Munich
+	Chair of Computational Modeling and Simulation.
+
+	TUM Open Infra Platform is free software; you can redistribute it and/or modify
+	it under the terms of the GNU General Public License Version 3
+	as published by the Free Software Foundation.
+
+	TUM Open Infra Platform is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
 //
 //  Author:  Peter Bonsma
 //  $Date: 1999-12-31 23:59:59 +0000 (Wed, 31 Jan 1999) $
@@ -9,32 +26,46 @@
 //  more info for commercial use:  peter.bonsma@rdf.bg
 //
 
-#include	"../stdafx.h"
+#define _USE_MATH_DEFINES
+#include <math.h>
+#include <iostream>
+#include <utility>
+#include <sstream>
 
-#include	<stdio.h>
-#include	<stdlib.h>
-#include	<string.h>
-#include	<assert.h>
-#include	<cmath>
+#include "CarveHeaders.h"
 
+#include "GeometrySettings.h"
+#include "ProfileConverter.h"
+#include "CurveConverter.h"
 
-#include	"spiral.h"
+#include "SpiralUtils.h"
 
+OIP_NAMESPACE_OPENINFRAPLATFORM_CORE_IFCGEOMETRYCONVERTER_BEGIN
 
+/**********************************************************************************************/
+/*!Calculate recursive multiplication for calculation Taylor series
+* \param[in]  value                  value of Taylor series
+* \param[in]  polynomialConstants    polynomial constants for a polynomial term - a,b,c,d
+* \param[in]  myArray                matrix with 1xn number of iteration points
+* \param[in]  i_n                    actual iteration point
+* \param[in]  n                      number of iteration points
+*
+* \return
+*/
 void	RecursiveMultiplication(
-	double	value,
-	double	* polynomialConstants,
-	uint_t	polynomialConstantCnt,
-	double	* myArray,
-	uint_t	myArrayParentIndex,
-	uint_t	i_n,
-	uint_t	n
+	const double	value,
+	const std::vector<double>& polynomialConstants,
+	const int polynomialConstantCnt,
+	std::vector<double>&  myArray,
+	const int myArrayParentIndex,
+	const int i_n,
+	const int n
 )
 {
-	assert(value);
+	if (!value) throw oip::InconsistentGeometryException("Invalid polynomial constants!");
 	if (i_n < n - 1) {
-		assert(polynomialConstantCnt && polynomialConstants[0] == 0.);
-		for (uint_t i = 1; i < polynomialConstantCnt; i++) {
+		if(polynomialConstantCnt && polynomialConstants[0] != 0.)throw oip::InconsistentGeometryException("Invalid polynomial constants!");
+		for (int i = 1; i < polynomialConstantCnt; i++) {
 			if (polynomialConstants[i]) {
 				RecursiveMultiplication(
 					value * polynomialConstants[i],
@@ -46,21 +77,29 @@ void	RecursiveMultiplication(
 		}
 	}
 	else {
-		assert(i_n == n - 1);
[…]
