### 2021-09-06 da0af79
Moved function to the model window and removed redundant image
