### 2021-09-06 007f9be
Added implementation for ZoomToObject function
