+++ UserInterface/ViewPanel/View.cpp
@@ -389,9 +389,9 @@ void OpenInfraPlatform::UserInterface::View::enableMap( const bool checked )
 	viewport_->enableMap(checked);
 }
 
-void OpenInfraPlatform::UserInterface::View::ZoomToOneObject()
+void OpenInfraPlatform::UserInterface::View::ZoomToOneObject(buw::Vector3f zoomMinExtend_ ,buw::Vector3f zoomMaxExtend_)
 {
-	viewport_->ZoomToOneObject();
+	viewport_->ZoomToOneObject(zoomMinExtend_, zoomMaxExtend_);
 }
 
 //void OpenInfraPlatform::UserInterface::View::onViewCubeSelectionChanged(buw::eViewCubeOrientation::Enum state)
