### 2021-09-06 2aba2ef
Added delete Object function
