+++ UserInterface/MainWindow.h
@@ -187,6 +187,8 @@ namespace OpenInfraPlatform
 			void on_actionViewport_as_screenshot_triggered();
 			//void on_actionAdd_Georeference_triggered();
 
+			void on_actionBlendObject_triggered();
+
 			//void on_checkBoxDifferentColorsForVerticalAlignmentElements_clicked(bool checked);
 			//void on_checkBoxHighlightSelectedAlignmentSegment_clicked(bool checked);
 			//void on_checkBoxShowDifferentAlignmentElements_clicked(bool checked);
