### 2021-09-06 12c4c46
Added new function getAngleOnCurve and also the function for reading polynomials has been redesigned. Note: functions do not display the correct angle and do not read points correctly at this time
