### 2021-09-06 92ca753
Added new function getAngleOnCurve and also the function for reading polynomials has been redesigned. Note: functions do not display the correct angle and do not read points correctly at this time
