### 2021-09-07 5d90a20
Moved ifcLoop.expired() from convertIfcFaceBound into convertIfcLoop. Adjusted variable name according to the lowerCamelCase convention.
