### 2021-09-08 2d2ef2f
Trying to implement delete object function
