### 2021-09-08 8cc1112
Changed return type of convertIfcLoop and its sub-functions
