### 2021-09-08 3aab188
replaced variable bool polyOrientation by its information source
