### 2021-09-09 5c04f9e
Corrected ZoomToOneObject function during call, thus it would work.
