### 2021-09-12 e6e6482
Updated function "Blend object", thus it would work on click
