### 2021-09-12 e1fcc3f
Updated function "Delete object", thus it would work on click
