+++ UserInterface/MainWindow.cpp
@@ -552,17 +552,15 @@ void OpenInfraPlatform::UserInterface::MainWindow::updateModelsUI()
 					// do nothing
 				}
 
-				//6. Function delete an object
-				if (std::dynamic_pointer_cast<oip::IfcModel>(model))
-				{
-					auto itemDeleteObject = new QTreeWidgetItem(itemModel);
-					itemDeleteObject->setText(0, "Delete object");
+				//7. Function delete an object
+				auto itemDeleteObject = new QTreeWidgetItem(itemModel);
+				itemDeleteObject->setText(0, "Delete object");
 
-					QPushButton *launchDeleteObjectButton = new QPushButton();
-					launchDeleteObjectButton->setText("Delete object");
-					QObject::connect(launchDeleteObjectButton, SIGNAL(clicked()), this, SLOT(on_actionDeleteObject_triggered(model, filename)));
-					modelsTreeWidget_->setItemWidget(itemDeleteObject, 1, launchDeleteObjectButton);
-				}
+				QPushButton *launchDeleteObjectButton = new QPushButton();
+				launchDeleteObjectButton->setText("Delete object");
+				QObject::connect(launchDeleteObjectButton, &QPushButton::clicked, [this, model, filename] {on_actionDeleteObject_triggered(model, filename); });
+				modelsTreeWidget_->setItemWidget(itemDeleteObject, 1, launchDeleteObjectButton);
+				
 				// expanded per default
 				itemModel->setExpanded(true);
 			}
@@ -921,7 +919,7 @@ void OpenInfraPlatform::UserInterface::MainWindow::actionGetCameraState() {
 	*/
 }
 
-void OpenInfraPlatform::UserInterface::MainWindow::on_actionDeleteObject_triggered(const std::shared_ptr<oip::IfcModel>& model, const QString& filename){
+void OpenInfraPlatform::UserInterface::MainWindow::on_actionDeleteObject_triggered(const std::shared_ptr<oip::IModel>& model, const QString& filename){
 	
 	QMessageBox::information(this, tr("Delete object"),
 		tr("The Button works!"), QMessageBox::Ok);
