### 2021-09-15 2fca2fe
Adapt cmake to new boost version
