+++ external/QtXlsxWriter/src/xlsx/xlsxzipreader.cpp
@@ -61,6 +61,10 @@
 	#include "QtGui/5.12.2/QtGui/private/qzipreader_p.h"
 	#elif (QT_VERSION == QT_VERSION_CHECK(5,14,2))
 		#include "QtGui/5.14.2/QtGui/private/qzipreader_p.h"
+    #elif (QT_VERSION == QT_VERSION_CHECK(5,15,0))
+        #include "QtGui/5.15.0/QtGui/private/qzipreader_p.h"
+	#elif (QT_VERSION == QT_VERSION_CHECK(5,15,1))
+		#include "QtGui/5.15.1/QtGui/private/qzipreader_p.h"
 	#endif
 #endif
 
