+++ external/QtXlsxWriter/src/xlsx/xlsxzipwriter.cpp
@@ -62,6 +62,10 @@
 		#include  "QtGui/5.12.2/QtGui/private/qzipwriter_p.h"
 	#elif (QT_VERSION == QT_VERSION_CHECK(5,14,2))
 		#include  "QtGui/5.14.2/QtGui/private/qzipwriter_p.h"
+    #elif (QT_VERSION == QT_VERSION_CHECK(5,15,0))
+        #include  "QtGui/5.15.0/QtGui/private/qzipwriter_p.h"
+    #elif (QT_VERSION == QT_VERSION_CHECK(5,15,1))
+        #include  "QtGui/5.15.1/QtGui/private/qzipwriter_p.h"
 	#endif
 
 #endif
