### 2021-09-17 cc2b38a
Changed swapOuterBoundaryToFront to work with reference. The input std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcFaceBound>> ifcFaceBounds of the function convertIfcFaceBoundList is changed to non-const non-reference, because at some point the original const source has to be converted to non-const (in order to perform the swap).
