+++ Core/src/IfcGeometryConverter/FaceConverter.h
@@ -791,24 +791,25 @@ namespace OpenInfraPlatform {
 				}
 				
 				std::vector<std::vector<carve::geom::vector<3>>> convertIfcFaceBoundList(
-					const std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcFaceBound>>& ifcFaceBounds,
+					std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcFaceBound>> ifcFaceBounds,
 					const carve::math::Matrix& pos) const noexcept(true)
 				{
-					const std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcFaceBound>> faceBounds = swapOuterBoundaryToFront(ifcFaceBounds);
+					swapOuterBoundaryToFront(ifcFaceBounds);
 
 					std::vector<std::vector<carve::geom::vector<3>>> faceBoundLoops;
-					faceBoundLoops.resize(faceBounds.size());
-					for (size_t i = 0; i < faceBounds.size(); i++)
+					faceBoundLoops.resize(ifcFaceBounds.size());
+					for (size_t i = 0; i < ifcFaceBounds.size(); i++)
 					{
 						// get vertices of one bound loop into faceBoundLoops[i]
-						faceBoundLoops[i] = convertIfcFaceBound(faceBounds[i], pos);
+						faceBoundLoops[i] = convertIfcFaceBound(ifcFaceBounds[i], pos);
 					}
 
 					return faceBoundLoops;
 				}
 
-				std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcFaceBound>> swapOuterBoundaryToFront(
-					std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcFaceBound>> ifcFaceBounds) const throw(...)
+				void swapOuterBoundaryToFront(
+					std::vector<EXPRESSReference<typename IfcEntityTypesT::IfcFaceBound>>& ifcFaceBounds
+				) const throw(...)
 				{
 					// As carve expects outer boundary of face to be at first index, outer boundary has to be moved to index 0 and inner boundary has index >= 1
 					for (auto it = ifcFaceBounds.begin(); it != ifcFaceBounds.end(); it++)
@@ -816,13 +817,13 @@ namespace OpenInfraPlatform {
 						if (it->isOfType<typename IfcEntityTypesT::IfcFaceOuterBound>())
 						{
 							// swap element 'it' with first element;
-							std::rotate(ifcFaceBounds.begin(), it, it + 1);
+							std::iter_swap(ifcFaceBounds.begin(), it);
 
 							// according to ifc-documentation, only one boundary (= loop) can be outer boundary:
 							// "One loop is optionally distinguished as the outer loop of the face."
 							// (https://standards.buildingsmart.org/IFC/DEV/IFC4_3/RC3/HTML/link/ifcface.htm)
 							// thus, break the for-loop after THE outer boundary was found
-							return ifcFaceBounds;
+							return;
 						}
 					}
 				}
