### 2021-09-17 7096ed8
Replaced for-loop with std::transform
