### 2021-09-18 9782715
Added doxygen comments to convertIfcFaceBound (and related functions)
