### 2021-09-18 8f5deca
Fixed doxygen comment of convertIfcTopologicalRepresentationItem
