### 2021-09-18 54e028b
Added IfcFace to convertIfcTopologicalRepresentationItem
