### 2021-09-18 5c2770c
Added IfcFaceBound to convertIfcTopologicalRepresentationItem
