### 2021-09-18 01a3edc
Added IfcLoop to convertIfcTopologicalRepresentationItem
