### 2021-09-18 3f00fd6
Fix: Compressed vector declaration
