### 2021-09-19 a9990e7
Added throw in swapOuterBoundaryToFront. The function throws, if there are multiple IfcFaceOuterBoundary entities.
