### 2021-09-20 c96ffdd
Defined some values, so it would be easier to find a problem in function (will delete them later)
