### 2021-09-21 1c374bb
Correcting small errors caused through the change to c++20
