### 2021-09-21 363c1cf
Specified auto-type declaration more specifically with const &.
