### 2021-09-22 5167ad5
Function getAngleOnCurve was renamed in getDirectionOfCurve. Some mistakes in code-interpretation were fixed. Function getPointOnCurve was updated for clothoid, namely overload was optimized. NOTE: The Angle is still displayed incorrectly. There may be a misinterpretation of the code from SpiralUtils
