+++ Core/src/IfcGeometryConverter/CurveConverter.h
@@ -1530,7 +1530,7 @@ namespace OpenInfraPlatform {
 					double numSegments = 100;
 					double deltaLength = length / numSegments;
 
-					std::vector<carve::geom::vector<3>> segmentPoints;
+					std::vector<carve::geom::vector<3>> segmentPoints, segmentDirections;
 					double lengthAlong = 0.;
 					switch (curveSegment->SegmentStart.which())
 					{
@@ -1554,20 +1554,24 @@ namespace OpenInfraPlatform {
 					for( int i = 0; i < numSegments; ++i )
 					{ 
 						// determine point
-						carve::geom::vector<3> point;
+						carve::geom::vector<3> point, direction;
 						if (curveSegment->ParentCurve.isOfType<typename IfcEntityTypesT::IfcClothoid>())
 						{
 							point = getPointOnCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcClothoid>(), runningLength);
+							direction = getDirectionOfCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcClothoid>(), runningLength);
 						}
 						else if (curveSegment->ParentCurve.isOfType<typename IfcEntityTypesT::IfcLine>())
 						{
 							point = getPointOnCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcLine>(), runningLength);
+							direction = getDirectionOfCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcLine>(), runningLength);
 						}
 						else if (curveSegment->ParentCurve.isOfType<typename IfcEntityTypesT::IfcCircle>())
 						{
 							point = getPointOnCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcCircle>(), runningLength);
+							direction = getDirectionOfCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcCircle>(), runningLength);
 						}
 						segmentPoints.push_back(point);
+						segmentDirections.push_back(direction);
 						// determine next length
 						lengthAlong += deltaLength;
 						switch (runningLength.which())
@@ -1602,22 +1606,33 @@ namespace OpenInfraPlatform {
 							0, 0, 1., 0.,
 							0., 0., 0., 1.);*/
 				
-						//get the angle
-						double theta = getAngleOnCurve(curveSegment->ParentCurve.as<typename IfcEntityTypesT::IfcClothoid>(), length);
+						//get the local coordinate system
+						carve::geom::vector<3> tangent = segmentDirections[0].normalize();
+						//tangent.y = tangent.y;
+						carve::geom::vector<3> localZ = carve::geom::VECTOR(0., 0., 1.),
+							orthogonal = carve::geom::cross(localZ, tangent).normalize(),
+							translate = -segmentPoints[0];
 						//calculate rotation
 						carve::math::Matrix rotationMatrix = carve::math::Matrix(
-							std::cos(theta), -std::sin(theta), 0., 0.,
-							std::sin(theta), std::cos(theta), 0., 0.,
-							0, 0, 1., 0.,
+							tangent.x, orthogonal.x, localZ.x, 0.,
+							tangent.y, orthogonal.y, localZ.y, 0.,
+							tangent.z, orthogonal.z, localZ.z, 0.,
 							0., 0., 0., 1.);
-					
+						/*carve::math::Matrix rotationMatrix = carve::math::Matrix(
+							0, -1, localZ.x, 0.,
+							1, 0, localZ.y, 0.,
+							0, 0, localZ.z, 0.,
+							0., 0., 0., 1.);*/
 						// apply placement
-						const carve::math::Matrix placement = placementConverter->convertIfcPlacement(curveSegment->Placement);
+						const carve::math::Matrix placement =
[…]
