### 2021-09-24 8792bc5
Added conditional construct for IfcFaceSurface into convertIfcFace
