### 2021-09-24 3a67468
Added body and call of convertIfcFaceSurface
