### 2021-09-24 865b5bb
Added body and call of convertIfcAdvancedFace
