### 2021-09-24 d76b7b8
Added check of IfcFaceBound-types in convertIfcAdvancedFace
