### 2021-09-24 305ec0c
Added check of IfcSurface-types in convertIfcAdvancedFace
