### 2021-09-24 cd721e1
Added function call to calculate IfcPlane in convertIfcAdvancedFace
