### 2021-09-24 e7cac46
Added call to IfcBSplineSurface in convertIfcAdvancedFace. Note: Surface is not append in return variable.
