### 2021-09-24 68b2bbe
Added mergePolyhedronsIntoOnePolyhedron to append surface to result
