### 2021-09-24 e78a63b
Fixing some minor errors
