+++ Core/src/IfcGeometryConverter/PlacementConverter.h
@@ -935,7 +935,7 @@ namespace OpenInfraPlatform {
 					std::vector<carve::geom::vector<2>> segmentStartPoints;
 					
 					std::shared_ptr<PlacementConverterT<IfcEntityTypesT>> placementConverter
-						= std::make_shared<PlacementConverterT<IfcEntityTypesT>>(this->GeomSettings(), this->UnitConvert());
+						= std::make_shared<PlacementConverterT<IfcEntityTypesT>>(GeomSettings(), this->UnitConvert());
 
 					CurveConverterT<IfcEntityTypesT> gridConv(this->GeomSettings(), this->UnitConvert(), placementConverter);
 					gridConv.convertIfcCurve2D(gridAxis->AxisCurve, axisVector, segmentStartPoints);
