### 2021-09-24 96c3f34
Started to implement IfcPlanarBox
