### 2021-09-29 a00fe11
Trying out different solutions for Zoom to object
