+++ UserInterface/MainWindow.cpp
@@ -932,13 +932,13 @@ void OpenInfraPlatform::UserInterface::MainWindow::on_actionZoomToOneObject_trig
 	buw::Vector3f fullBBoxCenter = fullBBox.center().cast<float>();
 
 	
-	buw::Vector3f delta = (fullBBox.center() - modelBBox.center()).cast<float>();
+	//buw::Vector3f delta = (fullBBox.center() - modelBBox.center()).cast<float>();
 
-	//buw::Vector3f zoomMinExtend = (modelBBox.min() - fullBBox.center()).cast<float>();
-	//buw::Vector3f zoomMaxExtend = (modelBBox.max() - fullBBox.center()).cast<float>();
+	buw::Vector3f zoomMinExtend = (modelBBox.min() - fullBBox.center()).cast<float>();
+	buw::Vector3f zoomMaxExtend = (modelBBox.max() - fullBBox.center()).cast<float>();
 
-	buw::Vector3f zoomMinExtend = modelBBox.min().cast<float>() - delta;
-	buw::Vector3f zoomMaxExtend = modelBBox.max().cast<float>() - delta;
+	//buw::Vector3f zoomMinExtend = modelBBox.min().cast<float>() - delta;
+	//buw::Vector3f zoomMaxExtend = modelBBox.max().cast<float>() - delta;
 
 	//buw::Vector3d offset = -zoomBox.center();
 	//buw::Vector3f zoomMinExtend = (zoomBox.min() - fullBBox.center()).cast<float>();
