### 2021-09-29 28e27f9
Implementing extent calculation
