### 2021-09-29 6825c5f
Split convertifcPlanarExtent in 2 function
