### 2021-09-29 8dee608
Implementing IfcPlanarBox
