### 2021-10-01 2034c1c
Fixed function parameter to be reference
