### 2021-10-01 b4bb324
Corrected IfcPlanarBox
