### 2021-10-01 bc916f2
Adjusted variable name of the FaceSurface-entity to be more meaningful
