### 2021-10-03 0195db3
Correcting small errors regarding the use of the "this" keyword
