### 2021-10-04 237f2b3
Moved mergePolyhedronsIntoOnePolyhedron into class ItemData
