+++ Core/src/IfcGeometryConverter/FaceConverter.h
@@ -865,7 +865,7 @@ namespace OpenInfraPlatform {
 						// ToDo: in general case, the B-Spline-Surface has to be trimmed by the loop boundary (or at least checked to be coincident)
 
 						// append surface-faces to target polygon
-						mergePolyhedronsIntoOnePolyhedron(inputDataFaceSurface, polygon);
+						inputDataFaceSurface->mergePolyhedronsIntoOnePolyhedron(polygon);
 					}
 					else if (faceSurface.isOfType<typename IfcEntityTypesT::IfcCylindricalSurface>())
 					{
@@ -1249,59 +1249,6 @@ namespace OpenInfraPlatform {
 							polygon->addFace(v0, v1, v2);
 					}
 
-				}
-
-				void mergePolyhedronsIntoOnePolyhedron(
-					const std::shared_ptr<ItemData> other,
-					std::shared_ptr<carve::input::PolyhedronData>& targetPolyhedron //Carve polygon of the converted face
-				)  const noexcept(false)
-				{
-					const std::vector<std::vector<std::shared_ptr<carve::input::PolyhedronData>>>& inputPolyhedronsCollection{
-						other->closed_polyhedrons,
-						other->open_polyhedrons,
-						other->open_or_closed_polyhedrons };
-
-					for (const auto& inputPolyhedrons : inputPolyhedronsCollection)
-					{
-						for (const std::shared_ptr<carve::input::PolyhedronData>& inputPolyhedron : inputPolyhedrons)
-						{
-							// number of existing vertices in target
-							const size_t indexOffset = targetPolyhedron->getVertexCount();
-
-							// --- add points ---
-							for (const auto& inputPoint : inputPolyhedron->points)
-							{
-								targetPolyhedron->addVertex(inputPoint);
-							}
-
-							// --- add faces ---
-							// iterator indicates position in vector inputPolyhedron
-							size_t iterator = 0;
-							// the indices of each face will be stored temporal in the vector indicesTemp
-							std::vector<size_t> indicesTemp;
-							// loop over faces: numberOfFaces = inputPolyhedron->getFaceCount()
-							for (size_t i = 0; i < inputPolyhedron->getFaceCount(); i++)
-							{
-								// get number of indices which describe one face
-								const size_t nIndicesOfFace = inputPolyhedron->faceIndices[iterator];
-								iterator++;
-
-								indicesTemp.reserve(nIndicesOfFace);
-								for (size_t j = iterator; j < iterator + nIndicesOfFace; j++)
-								{
-									// index shift by number of exiting vertices, store in temporal vector
-									indicesTemp.push_back(inputPolyhedron->faceIndices[j] + indexOffset);
-								}
-
-								// add face into target
-								targetPolyhedron->addFace(indicesTemp.begin(), indicesTemp.end());
-								indicesTemp.clear();
-
-								// set iterator to start of next face
-								iterator += nIndicesOfFace;
-							}
-						}
-					}
 				}
 
 				/*! \brief Typename definition (alias) of \c EXPRESSContainer for a easier use of \c OpenInfraPlatform::EarlyBinding::EXPRESSContainer.
