### 2021-10-05 1d90b6e
Added doxy-comments for functions
