+++ Core/src/IfcGeometryConverter/FaceConverter.h
@@ -1486,13 +1486,29 @@ namespace OpenInfraPlatform {
 					}
 				}
 
-
-
+				/**********************************************************************************************/
+				/*! \brief Converts \c IfcPlanarExtent to a triangulated vector.
+				*
+				* \param[in] planarExtent				A pointer to data from \c IfcPlanarExtent.
+				* \param[in] pos						A position matrix, which should be applied to the points.
+				*
+				* \param[out] itemData					A pointer to be filled with the relevant data.
+				*/
 				void convertIfcPlanarExtent(const EXPRESSReference<typename IfcEntityTypesT::IfcPlanarExtent> planarExtent,
 					const carve::math::Matrix& pos,
 					std::shared_ptr<ItemData>& itemData)
 					const noexcept(false)
 				{
+					// **************************************************************************************************************************
+					//	https://standards.buildingsmart.org/IFC/RELEASE/IFC4_1/FINAL/HTML/schema/ifcpresentationdefinitionresource/lexical/ifcplanarextent.htm
+					//	ENTITY IfcPlanarExtent
+					//		SUPERTYPE OF(IfcPlanarBox)
+					//		SUBTYPE OF(IfcGeometricRepresentationItem);
+					//			SizeInX: IfcLengthMeasure;
+					//			SizeInY: IfcLengthMeasure;
+					//	END_ENTITY;
+					// **************************************************************************************************************************
+
 					if (planarExtent.expired())
 						throw oip::ReferenceExpiredException(planarExtent);
 
@@ -1505,12 +1521,27 @@ namespace OpenInfraPlatform {
 					throw oip::UnhandledException(planarExtent);
 				}
 
-
+				/**********************************************************************************************/
+				/*! \brief Converts \c IfcPlanarBox to a triangulated vector.
+				*
+				* \param[in] planarBox					A pointer to data from \c IfcPlanarBox.
+				* \param[in] pos						A position matrix, which should be applied to the points.
+				*
+				* \param[out] itemData					A pointer to be filled with the relevant data.
+				*/
 				void convertIfcPlanarBox(EXPRESSReference<typename IfcEntityTypesT::IfcPlanarBox> planarBox,
 					const carve::math::Matrix& pos,
 					std::shared_ptr<ItemData>& itemData) 
 					const noexcept(false)
 				{
+					// **************************************************************************************************************************
+					//	https://standards.buildingsmart.org/IFC/RELEASE/IFC4_1/FINAL/HTML/schema/ifcpresentationdefinitionresource/lexical/ifcplanarbox.htm
+					//	ENTITY IfcPlanarBox
+					//		SUBTYPE OF(IfcPlanarExtent);
+					//		Placement: IfcAxis2Placement;
+					//	END_ENTITY;
+					// **************************************************************************************************************************
+
 					if (planarBox.expired())
 						throw oip::ReferenceExpiredException(planarBox);
 
@@ -1551,9 +1582,7 @@ namespace OpenInfraPlatform {
 
 					itemData->polylines.push_back(polylineData);
 					itemData->open_polyhedrons.push_back(polyhedronData);
-
 				}
-
 
 				protected:
 
