### 2021-10-06 e845efb
Trying to add hard-coded values to function
