### 2021-10-06 ba74506
Updating setupHelp and adding a markdown file for the latest release notes
