+++ Documentation/markdown/ReleaseNotes.md
@@ -1,0 +1,46 @@
+
+# News
+
+
+## 4.0 Release Notes
+
+*October 11th 2021*
+
+## Updates
+-------
+
++ **Main changes**
+
+  * Update to Visual Studio 2019 (`#319 <https://github.com/tumcms/Open-Infra-Platform/pull/319>`_)
+
+  * Major refactoring of SolidModelConverter, SplineConverter, GeometrySettings, GeomUtils and more
+
+  * Reading .off files enable (`#271 <https://github.com/tumcms/Open-Infra-Platform/pull/294>`_)
+
+  * First implementation of georeferencing IfcModels (`#396 <https://github.com/tumcms/Open-Infra-Platform/pull/396>`_)
+
+
++ **Various improvements**
+
+  * Supporting IfcIndexedPolygonalFaceWithVoids (`#448 <https://github.com/tumcms/Open-Infra-Platform/pull/448>`_)
+
+  * Supporting IfcPlanarBox (`#475 <https://github.com/tumcms/Open-Infra-Platform/pull/475>`_)
+
+  * Supporting IfcFixedReferenceSweptAreaSolid (`#406 <https://github.com/tumcms/Open-Infra-Platform/pull/406>`_)
+
+  * Added various unit tests
+
+  * Screen shot maker (`#446 <https://github.com/tumcms/Open-Infra-Platform/pull/446>`_)
+
+  * List with all supported Ifc entities (`#397 <https://github.com/tumcms/Open-Infra-Platform/pull/397>`_)
+
+  * Added detailed description of the software architecture (`#466 <https://github.com/tumcms/Open-Infra-Platform/pull/466>`_)
+
+
+Bug fixes
+---------
+
+* Fixing bug in alignment placement (`#443 <https://github.com/OSGeo/PROJ/issues/2692>`_)
+
+
+
