+++ Readme.md
@@ -10,6 +10,10 @@ It comes with its own `EXPRESS` parser and `C++` classes generator.
 
 ![Beta1_5_Windows8_1.png](./Documentation/images/Beta1_5_Windows8_1.png)
 
+## News
+
+See [Release Notes](Documentation/markdown/ReleaseNotes.md)
+
 ## Documentation
 
 - [Building OIP on your machine](Documentation/markdown/SetupHelp.md)
@@ -23,6 +27,9 @@ It comes with its own `EXPRESS` parser and `C++` classes generator.
 Out-of-sync:
 
 [Installation, Usage & Updating](Documentation/markdown/User.md)
+
+## Operating Sytems
+Currently, TUM Open Infra Platform only supports machines running on Windows. However, Linux support is considered for the future.
 
 ## License
 
