### 2021-10-07 100c1e1
Fixed initialization of temporal ItemData variable
