### 2021-10-07 377fe78
Small adjustments of the documentation
