### 2021-10-07 3ce4b4f
reformatted FAQ, added some more
