### 2021-10-07 6b11199
some minor corrections to setup help documentation
