### 2021-10-07 a706b96
small improvements to setup help
