### 2021-10-07 94049d6
further improvements to the setup help
