### 2021-10-13 65d6e4d
New unit tests for IFC4X3_RC4 was added. NOTE: Test "IFCHasAnEssentialEntity" not complete; Screen shots was not implemented
