### 2021-10-14 7d51cb7
First try to include libtiff library (not working yet)
