+++ PointCloudProcessing/CMakeLists.txt
@@ -30,6 +30,22 @@ add_definitions(-DBLUEINFRASTRUCTURE_EXPORT)
 
 # Find thirdparty libraries.
 FetchContent_Declare(
+	dep_libtiff
+	GIT_REPOSITORY https://gitlab.com/libtiff/libtiff.git
+	GIT_TAG v4.3.0
+)
+
+FetchContent_GetProperties(dep_libtiff)
+if(NOT dep_libtiff_POPULATED)
+	FetchContent_Populate(dep_libtiff)
+	add_subdirectory(${dep_libtiff_SOURCE_DIR}/libtiff ${dep_libtiff_BINARY_DIR} EXCLUDE_FROM_ALL)
+
+	set_target_properties(tiff_library PROPERTIES FOLDER "ThirdParty")
+endif()
+
+set(TIFF_INCLUDE_DIR ${dep_libtiff_SOURCE_DIR}/libtiff)
+
+FetchContent_Declare(
 	dep_libgeo
 	GIT_REPOSITORY https://github.com/OSGeo/libgeotiff.git
 	GIT_TAG 1.6.0
@@ -89,6 +105,7 @@ source_group(PointCloudProcessing										FILES ${PointCloudProcessing_Source})
 # Create a target for the library
 set(DEPENDENT_INCLUDE_DIRS
 	src
+	${dep_libtiff_SOURCE_DIR}/include
 	${dep_libgeo_SOURCE_DIR}/include
 	${dep_liblas_SOURCE_DIR}/include
 	${CMAKE_CURRENT_BINARY_DIR} 	
@@ -108,6 +125,7 @@ target_include_directories(OpenInfraPlatform.PointCloudProcessing PUBLIC ${DEPEN
 target_link_libraries(OpenInfraPlatform.PointCloudProcessing PUBLIC
 		las
 		las_c
+		tiff_library
 		geotiff_library
 		BlueFramework.ImageProcessing
 		BlueFramework.Rasterizer
