### 2021-10-21 1200e19
Added new function 'getNextUncommentedLine' to read only uncommented lines for off files
