+++ Core/src/OffConverter/OffReader.cpp
@@ -52,7 +52,11 @@ std::shared_ptr<OffModel> OffReader::readFile(const std::string& filename)
 		while (std::getline(offFile, line) && line != "OFF");
 
 		//read number of elements
-		getline(offFile, line);
+		//getline(offFile, line);
+		if (!OffReader::getNextUncommentedLine(offFile, line))
+		{
+			return false;
+		}
 		std::vector<int>	lineData;
 		std::stringstream	lineStream(line);
 
@@ -81,7 +85,11 @@ std::shared_ptr<OffModel> OffReader::readFile(const std::string& filename)
 		for (int i = 0; i < nrOfFaces; i++)
 		{
 			//which type of face (triangle, quad, ...)
-			std::getline(offFile, line);
+			if (!OffReader::getNextUncommentedLine(offFile, line))
+			{
+				return false;
+			}
+			//std::getline(offFile, line);
 			std::vector<uint32_t> faceVector;
 			std::stringstream lineStream(line);
 
@@ -130,7 +138,11 @@ std::vector<buw::Vector3f> OffReader::readVertices(const int nrOfVertices,
 	std::vector<buw::Vector3f> allVertices;
 	for (int i = 0; i < nrOfVertices; i++)
 	{
-		std::getline(offFile, line);
+		//std::getline(offFile, line);
+		if (OffReader::getNextUncommentedLine(offFile, line))
+		{
+			line;
+		}
 		buw::Vector3f position;
 		std::stringstream lineStream(line);
 
@@ -241,5 +253,16 @@ buw::Vector3f OffReader::calcNormal(const buw::Vector3f& vertex1,
 	return buw::Vector3f(normal.x, normal.y, normal.z);
 }
 
+bool OffReader::getNextUncommentedLine(std::ifstream& offFile, std::string& line)
+{
+	while (std::getline(offFile, line))
+	{
+		if (line[0] != '#')
+		{
+			return true;
+		}
+	}
+	return false;
+}
 
 OIP_NAMESPACE_OPENINFRAPLATFORM_CORE_OFFCONVERTER_END
