### 2021-10-21 d1a6359
Fixed mistake in 'getNextUncommentedLine'. Now ignored comments at the end of a line
