+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -2195,34 +2195,33 @@ namespace OpenInfraPlatform
 			) const noexcept(false)
 			{
 				// TODO: complete this function + refactoring 
-				/*
+				
 				//ENTITY IfcHalfSpaceSolid SUPERTYPE OF(ONEOF(IfcBoxedHalfSpace, IfcPolygonalBoundedHalfSpace))
-				EXPRESSReference<typename IfcEntityTypesT::IfcSurface> base_surface = halfSpaceSolid->BaseSurface;
+				EXPRESSReference<typename IfcEntityTypesT::IfcSurface> baseSurface = halfSpaceSolid->BaseSurface;
 
 				// base surface
-				if (!base_surface.template isOfType<typename IfcEntityTypesT::IfcElementarySurface>())
+				if (!baseSurface.template isOfType<typename IfcEntityTypesT::IfcElementarySurface>())
 				{
-					throw oip::InconsistentModellingException(base_surface,
+					throw oip::InconsistentModellingException(baseSurface,
 						"The base surface shall be an unbounded surface (subtype of IfcElementarySurface)");
 				}
-				EXPRESSReference<typename IfcEntityTypesT::IfcElementarySurface> elem_base_surface =
-					base_surface.template as <typename IfcEntityTypesT::IfcElementarySurface>();
+				
+				EXPRESSReference<typename IfcEntityTypesT::IfcElementarySurface> elemBaseSurface =
+					baseSurface.template as <typename IfcEntityTypesT::IfcElementarySurface>();
 
-				oip::EXPRESSReference<typename IfcEntityTypesT::IfcAxis2Placement3D>& base_surface_pos = elem_base_surface->Position;
-				carve::geom::plane<3> base_surface_plane;
-				carve::geom::vector<3> base_surface_position;
-				carve::math::Matrix base_position_matrix(carve::math::Matrix::IDENT());
-				if (base_surface_pos)
+				carve::geom::plane<3> baseSurfacePlane;
+				carve::geom::vector<3> baseSurfacePosition;
+				carve::math::Matrix basePositionMatrix(carve::math::Matrix::IDENT());
+				if (elemBaseSurface->Position)
 				{
-					placementConverter->getPlane(base_surface_pos.lock(), base_surface_plane, base_surface_position, UnitConvert()->getLengthInMeterFactor());
-					base_position_matrix = placementConverter->convertIfcAxis2Placement3D(base_surface_pos);
+					placementConverter->getPlane(elemBaseSurface->Position.lock(), baseSurfacePlane, baseSurfacePosition, UnitConvert()->getLengthInMeterFactor());
+					basePositionMatrix = placementConverter->convertIfcAxis2Placement3D(elemBaseSurface->Position);
 				}
 
 				// If the agreement flag is TRUE, then the subset is the one the normal points away from
-				bool agreement = halfSpaceSolid->AgreementFlag;
-				if (!agreement)
+				if (!halfSpaceSolid->AgreementFlag)
 				{
-					base_surface_plane.negate();
+					baseSurfacePlane.negate();
 				}
 
 				if (halfSpaceSolid.template isOfType<typename IfcEntityTypesT::IfcBoxedHalfSpace>())
@@ -2235,7 +2234,7 @@ namespace OpenInfraPlatform
 
 				// check dimenstions of other operand
 				double extrusion_depth = HALF_SPACE_BOX_SIZE;
-				carve::geom::vector<3> other_operand_pos = base_surface_position;
+				carve::geom::vector<3> other_operand_pos = baseSurfacePosition;
 				if (otherOperand)
 				{
 					carve::geom::aabb<3> aabb;
@@ -2288,28 +2287,40 @@ namespace OpenInfraPlatform
 
 					if (var == 0)
 					{
[…]
