+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -2233,8 +2233,8 @@ namespace OpenInfraPlatform
 				}
 
 				// check dimenstions of other operand
-				double extrusion_depth = HALF_SPACE_BOX_SIZE;
-				carve::geom::vector<3> other_operand_pos = baseSurfacePosition;
+				double extrusionDepth = HALF_SPACE_BOX_SIZE;
+				carve::geom::vector<3> otherOperandPos = baseSurfacePosition;
 				if (otherOperand)
 				{
 					carve::geom::aabb<3> aabb;
@@ -2258,52 +2258,40 @@ namespace OpenInfraPlatform
 						}
 					}
 					aabb.max();
-					carve::geom::vector<3>& aabb_extent = aabb.extent;
-					double max_extent = std::max(aabb_extent.x, std::max(aabb_extent.y, aabb_extent.z));
-					extrusion_depth = 2.0*max_extent;
-					other_operand_pos = aabb.pos;
+					carve::geom::vector<3>& aabbExtent = aabb.extent;
+					double maxExtent = std::max(aabbExtent.x, std::max(aabbExtent.y, aabbExtent.z));
+					extrusionDepth = 2.0 * maxExtent;
+					otherOperandPos = aabb.pos;
 				}
 
-
+				// ENTITY IfcPolygonalBoundedHalfSpace SUBTYPE OF IfcHalfSpaceSolid
 				if (halfSpaceSolid.template isOfType<typename IfcEntityTypesT::IfcPolygonalBoundedHalfSpace>())
 				{
-					EXPRESSReference<typename IfcEntityTypesT::IfcPolygonalBoundedHalfSpace> polygonal_half_space =
-						halfSpaceSolid.template as<typename IfcEntityTypesT::IfcPolygonalBoundedHalfSpace>();
-					// ENTITY IfcPolygonalBoundedHalfSpace
-					//	SUBTYPE OF IfcHalfSpaceSolid;
-					//	Position	 :	IfcAxis2Placement3D;
-					//	PolygonalBoundary	 :	IfcBoundedCurve;
-
 					convertIfcPolygonalBoundedHalfSpace(
 						halfSpaceSolid.template as<typename IfcEntityTypesT::IfcPolygonalBoundedHalfSpace>(), 
-						pos, itemData, otherOperand, extrusion_depth);
+						pos, itemData, otherOperand, extrusionDepth);
 					return;
 				}
+				// else, its an unbounded half space solid, create simple box
 				else
 				{
-					// else, its an unbounded half space solid, create simple box
-
+					//TO DO: adjust var
 					int var = 0;
-
 					if (var == 0)
 					{
-						//std::shared_ptr<carve::input::PolylineSetData> surface_data;
-						std::shared_ptr<ItemData> surface_data(new ItemData());
-						faceConverter->convertIfcSurface(baseSurface, carve::math::Matrix::IDENT(), surface_data);
+						std::shared_ptr<ItemData> surfaceData(new ItemData());
+						faceConverter->convertIfcSurface(baseSurface, carve::math::Matrix::IDENT(), surfaceData);
 
-						std::vector<std::shared_ptr<carve::input::PolylineSetData>> surfacePolylineData = surface_data->polylines;
-
+						std::vector<std::shared_ptr<carve::input::PolylineSetData>> surfacePolylineData = surfaceData->polylines;
 						std::vector<std::vector<carve::geom::vector<3>>> baseSurfacePointVector;
 
 						for (auto polyline : surfacePolylineData) {
 							baseSurfacePointVector.push_back(polyline->points);
 						}
 
-						std::vector<carve::geom::vector<3>> baseSurface_points = baseSurfacePointVector[0];
+						std::vector<carve::geom::vector<3>> baseSurfacePoints = baseSurfacePointVector[0];
 
[…]
