### 2021-10-22 e4ff92f
Some Refactoring to convertIfcHalfSpaceSolid function
