### 2021-10-22 850f998
Started new implementation for read the color of faces
