### 2021-10-22 c3d24ac
Reversed changes in 'getNextUncommentedLine'
