### 2021-10-25 d195de5
Adjustments to PointCloudProcessing due to update to VS2019
