+++ PointCloudProcessing/CMakeLists.txt
@@ -52,7 +52,7 @@ set(TIFF_INCLUDE_DIR "C:/OSGeo4W/include")
 set(TIFF_LIBRARY "C:/OSGeo4W/lib/tiff.lib")
 find_package(TIFF REQUIRED)
 
-FetchContent_Declare(
+#[[FetchContent_Declare(
 	dep_libgeo
 	GIT_REPOSITORY https://github.com/OSGeo/libgeotiff.git
 	GIT_TAG 1.6.0
@@ -64,7 +64,11 @@ if(NOT dep_libgeo_POPULATED)
 	add_subdirectory(${dep_libgeo_SOURCE_DIR}/libgeotiff ${dep_libgeo_BINARY_DIR} EXCLUDE_FROM_ALL)
 
 	set_target_properties(geotiff_library PROPERTIES FOLDER "ThirdParty")
-endif()
+endif()]]
+
+set(GeoTIFF_INCLUDE_DIR "C:/OSGeo4W/include")
+set(GeoTIFF_LIBRARY "C:/OSGeo4W/lib/geotiff_i.lib")
+#find_package(GeoTIFF REQUIRED)
 
 FetchContent_Declare(
 	dep_liblas
@@ -113,7 +117,8 @@ source_group(PointCloudProcessing										FILES ${PointCloudProcessing_Source})
 set(DEPENDENT_INCLUDE_DIRS
 	src
 	${TIFF_INCLUDE_DIR}
-	${dep_libgeo_SOURCE_DIR}/include
+	${GeoTIFF_INCLUDE_DIR}
+	#${dep_libgeo_SOURCE_DIR}/include
 	${dep_liblas_SOURCE_DIR}/include
 	${CMAKE_CURRENT_BINARY_DIR} 	
 	${blue_framework_SOURCE_DIR}/include
@@ -133,7 +138,8 @@ target_link_libraries(OpenInfraPlatform.PointCloudProcessing PUBLIC
 		las
 		las_c
 	    ${TIFF_LIBRARIES}
-		geotiff_library
+	    ${GeoTIFF_LIBRARIES}
+		#geotiff_library
 		BlueFramework.ImageProcessing
 		BlueFramework.Rasterizer
 		OpenInfraPlatform.Base
