### 2021-10-27 69e3c6b
Refactoring convertIfcBoxedHalfSpace
