### 2021-10-29 f083ca7
Fixes #384, added new function `readColorsFromFace`to read colors from the off.files
