+++ Core/src/OffConverter/OffReader.h
@@ -114,7 +114,7 @@ public:
 	*
 	* This function read color vector from the text line
 	*
-	* \param[in] vertex1 The first vertex of the triangle.
+	* \param[in] lineStream The line stream obtained from file reading (getline).
 	* \return The vector wich includes the numbers of RGB color
 	*/
 
