### 2021-10-30 1399296
Refactoring convertIfcPolygonalBoundedHalfSpace
