+++ Core/src/IfcGeometryConverter/SolidModelConverter.h
@@ -2427,88 +2427,87 @@ namespace OpenInfraPlatform
 				const std::shared_ptr<ItemData>& otherOperand,
 				double extrusionDepth) const noexcept(false)
 			{
-				throw oip::UnhandledException(polygonalHalfSpace);
-				/*
+				if (polygonalHalfSpace.expired())
+					throw oip::ReferenceExpiredException(polygonalHalfSpace);
+				
 				EXPRESSReference<typename IfcEntityTypesT::IfcSurface> baseSurface = polygonalHalfSpace->BaseSurface;
 
 				EXPRESSReference<typename IfcEntityTypesT::IfcElementarySurface> elemBaseSurface =
 					baseSurface.template as <typename IfcEntityTypesT::IfcElementarySurface>();
 
-				oip::EXPRESSReference<typename IfcEntityTypesT::IfcAxis2Placement3D>& baseSurfacePos = elemBaseSurface->Position;
 				carve::geom::plane<3> baseSurfacePlane;
 				carve::geom::vector<3> baseSurfacePosition;
 				carve::math::Matrix basePositionMatrix(carve::math::Matrix::IDENT());
-				if (baseSurfacePos)
+				if (elemBaseSurface->Position)
 				{
-					placementConverter->getPlane(baseSurfacePos.lock(), baseSurfacePlane, baseSurfacePosition, UnitConvert()->getLengthInMeterFactor());
-					basePositionMatrix = placementConverter->convertIfcAxis2Placement3D(baseSurfacePos);
+					placementConverter->getPlane(elemBaseSurface->Position.lock(), baseSurfacePlane, baseSurfacePosition, UnitConvert()->getLengthInMeterFactor());
+					basePositionMatrix = placementConverter->convertIfcAxis2Placement3D(elemBaseSurface->Position);
 				}
 
 				// If the agreement flag is TRUE, then the subset is the one the normal points away from
-				bool agreement = polygonalHalfSpace->AgreementFlag;
-				if (!agreement)
+				if (!polygonalHalfSpace->AgreementFlag)
 				{
 					baseSurfacePlane.negate();
 				}
 
-				carve::math::Matrix boundary_position_matrix(carve::math::Matrix::IDENT());
-				carve::geom::vector<3> boundary_plane_normal(carve::geom::VECTOR(0, 0, 1));
-				carve::geom::vector<3> boundary_position;
+				carve::math::Matrix boundaryPositionMatrix(carve::math::Matrix::IDENT());
+				carve::geom::vector<3> boundaryPlaneNormal(carve::geom::VECTOR(0, 0, 1));
+				carve::geom::vector<3> boundaryPosition;
 
 				if (polygonalHalfSpace->Position)
 				{
-					boundary_position_matrix = placementConverter->convertIfcAxis2Placement3D(polygonalHalfSpace->Position);
-					boundary_plane_normal = carve::geom::VECTOR(boundary_position_matrix._31, boundary_position_matrix._32, boundary_position_matrix._33);
-					boundary_position = carve::geom::VECTOR(boundary_position_matrix._41, boundary_position_matrix._42, boundary_position_matrix._43);
+					boundaryPositionMatrix = placementConverter->convertIfcAxis2Placement3D(polygonalHalfSpace->Position);
+					boundaryPlaneNormal = carve::geom::VECTOR(boundaryPositionMatrix._31, boundaryPositionMatrix._32, boundaryPositionMatrix._33);
+					boundaryPosition = carve::geom::VECTOR(boundaryPositionMatrix._41, boundaryPositionMatrix._42, boundaryPositionMatrix._43);
 				}
 
 				// PolygonalBoundary is given in 2D
[…]
