### 2021-10-30 a5dcbc8
Updating convertIfcHalfSpaseSolid
