### 2021-10-30 cb46ac8
Turn automoc on
