+++ Core/src/OffConverter/OffReader.cpp
@@ -99,7 +99,7 @@ std::shared_ptr<OffModel> OffReader::readFile(const std::string& filename)
 				readQuadFace(lineStream, indices, buwVertices, offVertices);
 			}
 			//read color
-			else if (faceType == 5)
+			else if (nrOfFaces-1)
 			{
 				readColorsFromFace(lineStream);
 			}
@@ -168,15 +168,15 @@ void OffReader::readTriangleFace(std::stringstream& lineStream,
 	buw::Vector3f normal = calcNormal(vector1, vector2, vector3);
 
 	//read color
-	buw::Vector3f color1 = OffReader::readColorsFromFace(lineStream);
-	buw::Vector3f color2 = OffReader::readColorsFromFace(lineStream);
-	buw::Vector3f color3 = OffReader::readColorsFromFace(lineStream);
+	buw::Vector3f color = OffReader::readColorsFromFace(lineStream);
+	//buw::Vector3f color2 = OffReader::readColorsFromFace(lineStream);
+	//buw::Vector3f color3 = OffReader::readColorsFromFace(lineStream);
 	//buw::Vector3f color(0.0f, 0.0f, 1.0f); //default color; to be changed later on 
 
 	//create vertices with position, color and normal and add to list of all vertices
-	vertices.push_back(buw::VertexPosition3Color3Normal3(vector1, color1, normal));
-	vertices.push_back(buw::VertexPosition3Color3Normal3(vector2, color2, normal));
-	vertices.push_back(buw::VertexPosition3Color3Normal3(vector3, color3, normal));
+	vertices.push_back(buw::VertexPosition3Color3Normal3(vector1, color, normal));
+	//vertices.push_back(buw::VertexPosition3Color3Normal3(vector2, color2, normal));
+	//vertices.push_back(buw::VertexPosition3Color3Normal3(vector3, color3, normal));
 }
 
 void OffReader::readQuadFace(std::stringstream& lineStream, 
@@ -215,20 +215,20 @@ void OffReader::readQuadFace(std::stringstream& lineStream,
 	//buw::Vector3f color2(0.0f, 0.0f, 1.0f); //default color; to be changed later on 
 	//buw::Vector3f color3(0.0f, 0.0f, 1.0f); //default color; to be changed later on
 
-	buw::Vector3f color1 = OffReader::readColorsFromFace(lineStream);
-	buw::Vector3f color2 = OffReader::readColorsFromFace(lineStream);
-	buw::Vector3f color3 = OffReader::readColorsFromFace(lineStream);
+	buw::Vector3f color = OffReader::readColorsFromFace(lineStream);
+	//buw::Vector3f color2 = OffReader::readColorsFromFace(lineStream);
+	//buw::Vector3f color3 = OffReader::readColorsFromFace(lineStream);
 
 	//create vertices with position, color and normal and add to list of all vertices
 	//first triangle
-	vertices.push_back(buw::VertexPosition3Color3Normal3(vector1, color1, normal1));
-	vertices.push_back(buw::VertexPosition3Color3Normal3(vector2, color2, normal1));
-	vertices.push_back(buw::VertexPosition3Color3Normal3(vector3, color3, normal1));
+	vertices.push_back(buw::VertexPosition3Color3Normal3(vector1, color, normal1));
+	vertices.push_back(buw::VertexPosition3Color3Normal3(vector2, color, normal1));
+	vertices.push_back(buw::VertexPosition3Color3Normal3(vector3, color, normal1));
 
 	//second triangle
[…]
