### 2021-11-03 4c2d15b
Remove unnecessary commented code
