### 2021-11-03 e06cbb1
Updated function `getNextUncommentedLine`. It return `string` instead of `bool`, so make code more readable and correct. Unnecessary comments were deleted
