+++ UnitTests/Off/sword/src/sword.cpp
@@ -1,0 +1,186 @@
+/*
+        Copyright (c) 2021 Technical University of Munich
+	Chair of Computational Modeling and Simulation.
+
+	TUM Open Infra Platform is free software; you can redistribute it and/or modify
+	it under the terms of the GNU General Public License Version 3
+	as published by the Free Software Foundation.
+
+	TUM Open Infra Platform is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#include <namespace.h>
+
+#include <OffVisualTest.h>
+
+#include <OffConverter/OffReader.h>
+
+
+using namespace testing;
+
+
+class Sword : public OffVisualTest 
+{
+protected:
+
+	// Test standard values
+	buw::Image4b _background = buw::Image4b(0, 0);
+
+	virtual void SetUp() override {
+		OffVisualTest::SetUp();
+
+		model = oip::OffReader::readFile(filename.string());
+
+		_background = rendererOff->captureImage();
+                rendererOff->setModel(model);
+	}
+
+	virtual void TearDown() override {
+		model.reset();
+		OffVisualTest::TearDown();
+	}
+
+	virtual std::string TestName() const { return "sword"; }
+
+	const boost::filesystem::path filename = dataPath("sword.off");
+
+	buw::ReferenceCounted<oip::OffModel> model = buw::makeReferenceCounted<oip::OffModel>();
+};
+
+TEST_F(Sword, AllEdgesAndFacesRed)
+{
+	// check for the number of vertices 
+	// does not coincide with the number of vertices mentioned in the off file since some vertices are
+	// stored once for every triangle they are part of
+	EXPECT_THAT(model->geometry().vertices.size(), Eq(3000));
+	// check for the number of indices (see comment for vertices)
+	EXPECT_THAT(model->geometry().indices.size(), Eq(3000));
+}
+
+TEST_F(Sword, ImageIsSaved)
+{
+	// Arrange
+	buw::Image4b image = rendererOff->captureImage();
+
+	// Act
+    buw::storeImage(testPath("sword.png").string(), image);
+
+	// Assert
+    EXPECT_NO_THROW(buw::loadImage4b(testPath("sword.png").string()));
+}
+
+TEST_F(Sword, PlaneSurfaceViews)
+{
+	// Arrange
+    const auto expected_front = buw::loadImage4b(dataPath("sword_front.png").string());
+    const auto expected_top = buw::loadImage4b(dataPath("sword_top.png").string());
+    const auto expected_bottom = buw::loadImage4b(dataPath("sword_bottom.png").string());
+    const auto expected_left = buw::loadImage4b(dataPath("sword_left.png").string());
+    const auto expected_right = buw::loadImage4b(dataPath("sword_right.png").string());
+    const auto expected_back = buw::loadImage4b(dataPath("sword_back.png").string());
+
+	// Act (Front)
+	rendererOff->setViewDirection(buw::eViewDirection::Front);
+	buw::Image4b image_front = CaptureImage();
+	// Act (Top)
+	rendererOff->setViewDirection(buw::eViewDirection::Top);
+	buw::Image4b image_top = CaptureImage();
+	// Act (Bottom)
+	rendererOff->setViewDirection(buw::eViewDirection::Bottom);
+	buw::Image4b image_bottom = CaptureImage();
[…]
