### 2021-11-07 ce21efd
Screenshots added. NOTE: Producing of screenshots for cube.off is note completely. There is problem with making them
