+++ UnitTests/Off/cube/src/cube.cpp
@@ -58,9 +58,9 @@ TEST_F(Cube, AllEdgesAndFacesRed)
 	// check for the number of vertices 
 	// does not coincide with the number of vertices mentioned in the off file since some vertices are
 	// stored once for every triangle they are part of
-	EXPECT_THAT(model->geometry().vertices.size(), Eq(8));
+	EXPECT_THAT(model->geometry().vertices.size(), Eq(30));
 	// check for the number of indices (see comment for vertices)
-	EXPECT_THAT(model->geometry().indices.size(), Eq(12));
+	EXPECT_THAT(model->geometry().indices.size(), Eq(30));
 }
 
 TEST_F(Cube, ImageIsSaved)
@@ -106,12 +106,12 @@ TEST_F(Cube, PlaneSurfaceViews)
 
 	// uncomment following lines to also save the screen shot
 
-    buw::storeImage(testPath("cube_front.png").string(), image_front);
-    buw::storeImage(testPath("cube_top.png").string(), image_top);
-    buw::storeImage(testPath("cube_bottom.png").string(), image_bottom);
-    buw::storeImage(testPath("cube_left.png").string(), image_left);
-    buw::storeImage(testPath("cube_right.png").string(), image_right);
-    buw::storeImage(testPath("cube_back.png").string(), image_back);
+   // buw::storeImage(testPath("cube_front.png").string(), image_front);
+   // buw::storeImage(testPath("cube_top.png").string(), image_top);
+   // buw::storeImage(testPath("cube_bottom.png").string(), image_bottom);
+   // buw::storeImage(testPath("cube_left.png").string(), image_left);
+   // buw::storeImage(testPath("cube_right.png").string(), image_right);
+   // buw::storeImage(testPath("cube_back.png").string(), image_back);
 
 
 	// Assert
@@ -162,14 +162,14 @@ TEST_F(Cube, VertexViews)
 
 	// uncomment following lines to also save the screen shot
 	
-    buw::storeImage(testPath("cube_front_left_bottom.png").string(), image_front_left_bottom);
-    buw::storeImage(testPath("cube_front_right_bottom.png").string(), image_front_right_bottom);
-    buw::storeImage(testPath("cube_top_left_front.png").string(), image_top_left_front);
-    buw::storeImage(testPath("cube_top_front_right.png").string(), image_top_front_right);
-    buw::storeImage(testPath("cube_top_left_back.png").string(), image_top_left_back);
-    buw::storeImage(testPath("cube_top_right_back.png").string(), image_top_right_back);
-    buw::storeImage(testPath("cube_back_left_bottom.png").string(), image_back_left_bottom);
-    buw::storeImage(testPath("cube_right_bottom_back.png").string(), image_right_bottom_back);
+    //buw::storeImage(testPath("cube_front_left_bottom.png").string(), image_front_left_bottom);
+    //buw::storeImage(testPath("cube_front_right_bottom.png").string(), image_front_right_bottom);
+    //buw::storeImage(testPath("cube_top_left_front.png").string(), image_top_left_front);
+    //buw::storeImage(testPath("cube_top_front_right.png").string(), image_top_front_right);
+    //buw::storeImage(testPath("cube_top_left_back.png").string(), image_top_left_back);
+    //buw::storeImage(testPath("cube_top_right_back.png").string(), image_top_right_back);
+    //buw::storeImage(testPath("cube_back_left_bottom.png").string(), image_back_left_bottom);
+    //buw::storeImage(testPath("cube_right_bottom_back.png").string(), image_right_bottom_back);
 
 
 	// Assert
