+++ UnitTests/Off/boxtorus/src/boxtorus.cpp
@@ -105,14 +105,14 @@ TEST_F(Boxtorus, PlaneSurfaceViews)
 	buw::Image4b image_back = CaptureImage();
 
 	// uncomment following lines to also save the screen shot
-
+	/*
     buw::storeImage(testPath("boxtorus_front.png").string(), image_front);
     buw::storeImage(testPath("boxtorus_top.png").string(), image_top);
     buw::storeImage(testPath("boxtorus_bottom.png").string(), image_bottom);
     buw::storeImage(testPath("boxtorus_left.png").string(), image_left);
     buw::storeImage(testPath("boxtorus_right.png").string(), image_right);
     buw::storeImage(testPath("boxtorus_back.png").string(), image_back);
-
+	*/
 
 	// Assert
 	EXPECT_EQ(image_front, expected_front);
@@ -161,7 +161,7 @@ TEST_F(Boxtorus, VertexViews)
 	buw::Image4b image_right_bottom_back = CaptureImage();
 
 	// uncomment following lines to also save the screen shot
-	
+	/*
     buw::storeImage(testPath("boxtorus_front_left_bottom.png").string(), image_front_left_bottom);
     buw::storeImage(testPath("boxtorus_front_right_bottom.png").string(), image_front_right_bottom);
     buw::storeImage(testPath("boxtorus_top_left_front.png").string(), image_top_left_front);
@@ -170,7 +170,7 @@ TEST_F(Boxtorus, VertexViews)
     buw::storeImage(testPath("boxtorus_top_right_back.png").string(), image_top_right_back);
     buw::storeImage(testPath("boxtorus_back_left_bottom.png").string(), image_back_left_bottom);
     buw::storeImage(testPath("boxtorus_right_bottom_back.png").string(), image_right_bottom_back);
-
+	*/
 
 	// Assert
 	EXPECT_EQ(image_front_left_bottom, expected_front_left_bottom);

[…]
