### 2021-11-10 bbb7fb7
Added statement if isn't provided in the line. Fixed mistakes in vertex and rgb color repr.
