### 2021-11-30 #3061 @jschlenger
Hi @CBenghi, @hugoledoux , @aothms, @abdoulayediak & @danielskatz

sorry for the late response. We've been struggling with the update to Visual Studio 2019 way longer than we expected, but now everything seems to be working fine.
I went through your comments, and down below, I put a list of your comments together with a response on how and where we tackled the raised issues.
Please take another look and let us know in case you have further comments.

> The platform is aimed at providing a development platform, but there is no information about how to use the platform as such

&#8594; Addition of a markdown file describing the overall software architecture to give developers a better understanding of how the OIP works and how they can adapt exiting modules to their needs (see [Software Architecture](https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/SoftwareArchitecture.md)).

> The paper text mentions a cross platform solution, but only Windows is currently supported. This needs to be better documented and perhaps explained: why only 64bit, why only msvc 2015/17

> Another point is the Windows-only support. Although I can see that the "cross-platform" claim was removed from the paper, it would be definitely valuable to state it early somewhere in the ReadMe of the repo.

&#8594; Update to Visual Studio 2019 in the most recently merged pull request (see github.com/tumcms/Open-Infra-Platform/pull/319); Adjustment was made to the main [Readme file](https://github.com/tumcms/Open-Infra-Platform/blob/development/Readme.md) and the paper that currently only Windows is supported.

> The documentation gives very little detail on geometry interpretation (which entities are supported and mapped to what geometric/topological entities).

&#8594; [List](https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/SupportedIFCrepresentations.md) with all supported entities was produced.

> Looking at the authors Contribution and authorship concerns I wonder if you have properly considered the contribution of Cara Coetzee

&#8594; Code contributions have been addressed properly in [Contribution and Authorship](https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/Contributors.md) file. Since this project is older than the repository itself (2013 vs 2017) there are also many authors' contributions not visible in the git history (we had some shenanigans with moving from one host to another, changing from hg to git, etc.). The decision was made, that only product managers would be the authors of the paper. The contributions of other (student) authors that worked on the project are acknowledged in the aforementioned file.

> Perhaps connected to the previous point, on line 57 you mention the specific focus on preliminary Express schemas. This uniqueness could be better highlighted in the project documentation, providing instructions for custom schema implementation.

&#8594; Explanation is given in the software architecture how one can introduce schema files of an IFC version (see [Software Architecture](https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/SoftwareArchitecture.md)).

> Mentioning limitations of OIP in comparison to other frameworks could better serve the research community (e.g. what you mentioned about the single IFC version at a time).

&#8594; Some limitations of OIP were added in the documentation (see [here](https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/SoftwareArchitecture.md)). Please advice if more is needed.

> Installation documentation could be slightly improved to clarify that plenty of warnings in cmake should be expected.

&#8594; Update of the installation documentation because of the update to VS2019 also mentioning the warnings that can be expected from cmake was introduced (see [here](https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/SoftwareArchitecture.md) and [here](https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/FAQ.md)).

> I've noticed a few apps in the solution that support your testing and QA. I would invite you to produce a quick readme markdown that might help developers understand their purpose and contribute cases.

&#8594; [A new markdown file](https://github.com/tumcms/Open-Infra-Platform/blob/development/Documentation/markdown/MiniApps.md) was added giving short explanations about the apps that are shipped together with OIP and what they can be used for.
