### 2021-12-07 #3061 @hugoledoux
👋 @CBenghi @aothms @abdoulayediak

the author fixed the issues you discussed, and modified the paper. Could you please have another look and let us know here if you're happy (also by clicking in the tickboxes above)?
