### 2021-12-22 #3061 @aothms
Dear authors, thank you for your efforts here, I think the paper + repo improved substantially. I just have the following two points that are a bit hard to check off in the current state:

I still find the following really minimal (if any):

*  State of the field: Do the authors describe how this software compares to other commonly-used packages?

For:

*  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I did see unit tests, but I didn't find a line in the docs with some background info and how to run them.
