### 2021-12-23 #3061 @jschlenger
Hi @aothms,

thank you for your comments!

We kept the comparison with other software very minimal since the papers by Hecht & Jaud (2019) and Valero et al. (2020) did a thorough comparison with various software tools. We directly mention them as source of information about OIP compared to other software.

I just added a new section to the documentation about unit tests and how to run them (see github.com/tumcms/Open-Infra-Platform/pull/519)
