### 2021-12-23 #3061 @CBenghi
Hello all,
Just a quick note from me about the timing of the review.
I'm currently on leave and will be able to review the update just after
Xmas.
Happy holidays,
Claudio

Il Gio 23 Dic 2021, 11:18 jschlenger ***@***.***> ha scritto:

> Hi @aothms <https://github.com/aothms>,
>
> thank you for your comments!
>
> We kept the comparison with other software very minimal since the papers
> by Hecht & Jaud (2019) and Valero et al. (2020) did a thorough comparison
> with various software tools. We directly mention them as source of
> information about OIP compared to other software.
>
> I just added a new section to the documentation about unit tests and how
> to run them (see github.com/tumcms/Open-Infra-Platform/pull/519
> <https://github.com/tumcms/Open-Infra-Platform/pull/519>)
>
