### 2022-04-19 #3061 @hugoledoux
a *HUGE* sorry is in order here @pjanck : I don't know how I let that go, but I seriously dropped the ball on that one. I thought it was done but no. Sorry.

I'm wrapping this up today, I'm going over the paper and proposing some fixes.

One thing I just noticed: the link above the first letter of your name is missing because it's a special char, but in previous it was there right? I just tried with a new paper and it should be there. Do you know what could be going on?
