### 2022-04-19 #3061 @hugoledoux
I went through the paper, please check my recommendations: https://github.com/tumcms/Open-Infra-Platform/pull/540

The \v{S} is something on our side, I'll see how this can be fixed... not action on your side needed.

At this point could you:
- [ ] Accept the PR above
