### 2022-04-26 #3061 @pjanck
> 10.1016/s0740-0020(08)00208-6 may be a valid DOI for title: IFC++

I can't seem to find this DOI anywhere. Also, IFC++ webpage does not list any publications that we could cite, or where a DOI could be identified. Deciding to ignore the comment.
