### 2022-04-26 #3061 @jschlenger
We have archived the release with Zenodo.
The DOI is: `10.5281/zenodo.6488493`
