### 2022-04-26 #3061 @danielskatz
@pjanck - I've suggested some minor changes to the paper in https://github.com/tumcms/Open-Infra-Platform/pull/546 while doing a final proofread, as the AEiC this week.
