### 2022-04-26 #3061 @pjanck
@danielskatz I've read through the paper once more and improved two typos.

I believe we're ready for publishing 🍰
